Contents

Contents

Index

Volume-Temperature Based Properties

Pressure-Temperature Based Properties

Automatic Differenciation functions

Index

Clapeyron.f_hess

Clapeyron.pip

Clapeyron.p∂p∂V

Clapeyron.volume

Clapeyron.∂2f

Clapeyron.∂2p

Clapeyron.∂f

Clapeyron.∂f∂T

Clapeyron.∂f∂V

Clapeyron.∂²³f

Volume-Temperature Based Properties

Clapeyron.pip — Function

pip(model::EoSModel,V,T,z=[1.0])

Phase identification parameter Π. as described in 1. If Π > 1, then the phase is clasified as a liquid or a liquid-like vapor, being a vapor or vapor-like liquid otherwise.

This identification parameter fails at temperatures and pressures well aboVe the critical point.

Calculated as:

Π = V*((∂²p/∂V∂T)/(∂p/∂T) - (∂²p/∂V²)/(∂p/∂V))

G. Venkatarathnama, L.R. Oellrich, Identification of the phase of a fluid using partial derivatives of pressure, volume,and temperature without reference to saturation properties: Applications in phase equilibria calculations, Fluid Phase Equilibria 301 (2011) 225–233

Pressure-Temperature Based Properties

Clapeyron.volume — Function

volume(model::EoSModel,p,T,z=SA[1.0];phase=:unknown,threaded=true)

calculates the volume (m³) of the compound modelled by model at a certain pressure,temperature and moles.

phase is a Symbol that determines the initial volume root to look for:

If phase =:liquid it will return the volume of the phase using a liquid initial point.

If phase =:vapor it will return the volume of the phase using a gas initial point.

The default is phase =:unknown. with this, both liquid and volume roots will be calculated, and the phase with the least amount of energy is returned.

The calculation of both volume roots can be calculated in serial (threaded=false) or in parallel (threaded=true)

Automatic Differenciation functions

Clapeyron.∂f∂T — Function

∂f∂T(model,V,T,z=SA[1.0])

returns f and ∂f/∂T at constant total volume and composition, where f is the total helmholtz energy, given by eos(model,V,T,z)

Clapeyron.∂f∂V — Function

∂f∂V(model,V,T,z=SA[1.0])

returns f and ∂f/∂V at constant temperature and composition, where f is the total helmholtz energy, given by eos(model,V,T,z), and V is the total volume

Clapeyron.∂f — Function

∂f(model,V,T,z)

returns zeroth order (value) and first order derivative information of the total helmholtz energy (given by eos(model,V,T,z)). the result is given in two values:

grad_f,fval = ∂2f(model,V,T,z)

where:

fval   = f(V,T) = eos(model,V,T,z)

grad_f = [ ∂f/∂V; ∂f/∂T]

Where V is the total volume, T is the temperature and f is the total helmholtz energy.

Clapeyron.p∂p∂V — Function

p∂p∂V(model,V,T,z=SA[1.0])

returns p and ∂p/∂V at constant temperature, where p is the pressure = pressure(model,V,T,z) and V is the total Volume.

Clapeyron.∂2f — Function

∂2f(model,V,T,z)

returns zeroth order (value), first order and second order derivative information of the total helmholtz energy (given by eos(model,V,T,z)). the result is given in three values:

hess_f,grad_f,fval = ∂2f(model,V,T,z)

where: ``` fval = f(V,T) = eos(model,V,T,z)

grad_f = [ ∂f/∂V; ∂f/∂T]

hess_f = [ ∂²f/∂V²; ∂²f/∂V∂T ∂²f/∂V∂T; ∂²f/∂V²] ```

Where V is the total volume, T is the temperature and f is the total helmholtz energy.

Clapeyron.∂2p — Function

∂2p(model,V,T,z)

returns zeroth order (value), first order and second order derivative information of the pressure. the result is given in three values:

hess_p,grad_p,pval = ∂2p(model,V,T,z)

where: ``` pval = p(V,T) = pressure(model,V,T,z)

grad_p = [ ∂p/∂V; ∂p/∂T]

hess_p = [ ∂²p/∂V²; ∂²p/∂V∂T ∂²p/∂V∂T; ∂²p/∂V²] ```

Where V is the total volume, T is the temperature and p is the pressure.

Clapeyron.f_hess — Function

f_hess(model,V,T,z)

returns the second order volume (V) and temperature (T) derivatives of the total helmholtz energy (given by eos(model,V,T,z)). the result is given in a 2x2 SMatrix, in the form:

[ ∂²f/∂V² ∂²f/∂V∂T ∂²f/∂V∂T ∂²f/∂V²]

use this instead of the ∂2f if you only need second order information. ∂2f also gives zeroth and first order derivative information, but due to a bug in the used AD, it allocates more than necessary.

Clapeyron.∂²³f — Function

∂²³f(model,V,T,z=SA[1.0])

returns ∂²A/∂V² and ∂³A/∂V³, in a single ForwardDiff pass. used mainly in crit_pure objective function