Contents

Contents

Index

Model Macros

Utility Macros

Index

Clapeyron.@comps

Clapeyron.@f

Clapeyron.@groups

Clapeyron.@nan

Clapeyron.@newmodel

Clapeyron.@newmodelgc

Clapeyron.@newmodelsimple

Clapeyron.@registermodel

Clapeyron.@sites

Model Macros

Clapeyron.@newmodelgc — Macro

@newmodelgc modelname parent paramstype

This is a data type that contains all the information needed to use an EoS model. It also functions as an identifier to ensure that the right functions are called.

The user is expected to create an outter constructor that takes this signature

function modelname(components::Array{String,1})

It should then return name(params::paramtype, groups::GroupParam, sites::SiteParam, idealmodel::IdealModel)

= Fields = The Struct consists of the following fields:

components: a string lists of components

icomponents: an iterator that goes through the indices corresponding to each component

groups: a GroupParam

sites: a SiteParam

params: the Struct paramstype that contains all parameters in the model

idealmodel: the IdealModel struct that determines which ideal model to use

assoc_options: struct containing options for the association solver. see AssocOptions

references: reference for this EoS

See the tutorial or browse the implementations to see how this is used.

Clapeyron.@newmodel — Macro

@newmodel name parent paramstype

This is exactly the same as the above but for non-GC models. All group parameters are absent in this struct. The sites are associated to the main component rather than the groups, and the respective fieldnames are named correspondingly.

Clapeyron.@newmodelsimple — Macro

@newmodelsimple

Even simpler model, primarily for the ideal models. Contains neither sites nor ideal models.

Clapeyron.@registermodel — Macro

@registermodel(model)

given an existing model, composed of Clapeyron EoS models, ClapeyronParams or EoSParams, it will generate the necessary traits to make the model compatible with Clapeyron routines.

Utility Macros

Clapeyron.@comps — Macro

@comps

This macro is an alias to

model.icomponents

The caveat is that model has to exist in the local namespace. model is expected to be an EoSModel type that contains the icomponents field. icomponents is an iterator that goes through all component indices.

Clapeyron.@groups — Macro

@groups

This macro is an alias to

model.groups.i_flattenedgroups

iflattenedgroups is an iterator that goes through all groups in flattenedgroups.

@groups(component)

This macro is an alias to

model.groups.i_groups[component]

i_groups[component] is an iterator that goes through all groups in relevent to a given component.

Clapeyron.@sites — Macro

@sites(component)

This macro is an alias to

model.sites.i_sites[component]

i_sites[component] is an iterator that goes through all sites relevant to each group in a GC model, and to each main component in a non-GC model.

Clapeyron.@f — Macro

@f(func,a,b,c,...)

This macro is an alias to

func(model, V, T, z, a, b, c, ...)

where func is the name of the function, model is the model struct, V is the volume, T is the absolute temperature, z is an array of number of moles of each component, and a, b, c, ... are arbitrary parameters that get passed to func.

It is very common for functions that are involved in the models to contain the model, V, T and z parameters, so this macro helps reduce code repetition as long as the first four parameters in the function are written exactly as above.

Clapeyron.@nan — Macro

@nan(function_call,default=NaN)

Wraps the function in a try-catch block, and if a DomainError or DivideError is raised, then returns default. for better results, its best to generate the default result beforehand