Contents

Contents

Index

Parsing Parameters from files

Parameter types

Index

Clapeyron.ParamOptions

Clapeyron.getparams

Parsing Parameters from files

Clapeyron.ParamOptions — Type

ParamOptions(;kwargs...)

Struct containing all the options related to parameter parsing:

userlocations::Vector{String} = String[]: List of used-defined locations to search.

usergrouplocations::Vector{String} = String[]: List of used-defined group locations to search.

asymmetricparams::Vector{String} = String[]: List of pair or association parameters that follow that param[i,j] ≠ param[j,i]

ignore_headers::Vector{String} = ["dipprnumber", "smiles"]: List of ignored headers.

ignore_missing_singleparams::Bool=false: Check if any single parameter has missing values. if set to false,getparams will fail when encountering a single parameter with missing values.

verbose::Bool = false: If true, show all operations done by getparams displayed in the terminal. this includes the warnings emmited by CSV.jl

species_columnreference::String ="species": column name to check for components. in pair and association params, it will check for #species#1 and #species#2, where #species# is the value of this option.

site_columnreference::String ="site": column name to check for sites in association params, it will check for #site#1 and #site#2, where #site# is the value of this option.

group_columnreference::String = "groups": column name to check for groups in group data.

normalisecomponents::Bool = true: If true, performs normalization of strings, on the CSV and input components

n_sites_columns::Dict{String,String} = Dict( "e" => "n_e","e1" => "n_e1","e2" => "n_e2","H" => "n_H"): dictionary to look number of sites. the number of sites is stored as columns in a single parameter csv file. for example, the number of sites of name e will be looked on the column n_e

return_sites::Bool = true: If set to false, association params will be ignored and sites will not be created, even if they exist in the list of locations.

Clapeyron.getparams — Function

params, sites = getparams(components,locations;kwargs...)

returns a Dict{String,ClapeyronParam} containing all the parameters found for the list of components in the available CSVs. locations are the locations relative to Clapeyron database. the available keywords are the ones used in ParamOptions

if return_sites is set to false, getparams will only return the found params.

Parameter types