Contents

Contents

Index

Primitive functions

Automatic Differenciation functions

Volume-Temperature Based Properties

Pressure-Temperature Based Bulk Properties

Single component properties

Multi component properties

Consistency and Stability

TP Flash

Initial guess functions

Index

Clapeyron.AntoineSaturation

Clapeyron.ChemPotDensitySaturation

Clapeyron.ChemPotVSaturation

Clapeyron.ClapeyronSaturation

Clapeyron.DETPFlash

Clapeyron.IsoFugacitySaturation

Clapeyron.MichelsenTPFlash

Clapeyron.RRTPFlash

Clapeyron.TPFlashMethod

Clapeyron.LLE_pressure

Clapeyron.LLE_temperature

Clapeyron.T_scale

Clapeyron.UCEP_mix

Clapeyron.UCST_mix

Clapeyron.VLLE_pressure

Clapeyron.VLLE_temperature

Clapeyron.a_res

Clapeyron.acentric_factor

Clapeyron.antoine_coef

Clapeyron.azeotrope_pressure

Clapeyron.azeotrope_temperature

Clapeyron.bubble_pressure

Clapeyron.bubble_temperature

Clapeyron.chemical_potential

Clapeyron.chemical_potential_res

Clapeyron.compressibility_factor

Clapeyron.crit_mix

Clapeyron.crit_pure

Clapeyron.dew_pressure

Clapeyron.dew_temperature

Clapeyron.diffusive_stability

Clapeyron.enthalpy

Clapeyron.enthalpy_vap

Clapeyron.entropy

Clapeyron.entropy_res

Clapeyron.eos

Clapeyron.eos_res

Clapeyron.f_hess

Clapeyron.fugacity_coefficient

Clapeyron.gibbs_duhem

Clapeyron.gibbs_free_energy

Clapeyron.gibbs_solvation

Clapeyron.helmholtz_free_energy

Clapeyron.idealmodel

Clapeyron.internal_energy

Clapeyron.isentropic_compressibility

Clapeyron.isobaric_expansivity

Clapeyron.isobaric_heat_capacity

Clapeyron.isochoric_heat_capacity

Clapeyron.isothermal_compressibility

Clapeyron.isstable

Clapeyron.joule_thomson_coefficient

Clapeyron.lb_volume

Clapeyron.mass_density

Clapeyron.mechanical_stability

Clapeyron.mixing

Clapeyron.molar_density

Clapeyron.numphases

Clapeyron.p_scale

Clapeyron.pip

Clapeyron.pressure

Clapeyron.p∂p∂V

Clapeyron.saturation_pressure

Clapeyron.saturation_temperature

Clapeyron.second_virial_coefficient

Clapeyron.speed_of_sound

Clapeyron.tp_flash

Clapeyron.tpd

Clapeyron.volume

Clapeyron.volume_virial

Clapeyron.x0_crit_pure

Clapeyron.x0_psat

Clapeyron.x0_sat_pure

Clapeyron.x0_saturation_temperature

Clapeyron.x0_volume

Clapeyron.x0_volume_gas

Clapeyron.x0_volume_liquid

Clapeyron.∂2f

Clapeyron.∂2p

Clapeyron.∂f

Clapeyron.∂f∂T

Clapeyron.∂f∂V

Clapeyron.∂²³f

Primitive functions

Almost all models in Clapeyron have at least one of the following functions defined.

Clapeyron.eos — Function

eos(model::EoSModel, V, T, z=SA[1.0])

Returns the total Helmholtz free energy.

Inputs:

model::EoSModel Thermodynamic model to evaluate

V Total volume, in [m³]

T Temperature, in [K]

z mole amounts, in [mol], by default is @SVector [1.0]

Outputs:

Total Helmholtz free energy, in [J]

by default, it calls R̄*T*∑(z)*(a_ideal(ideal_model,V,T,z) + a_res(model,V,T,z)) where ideal_model == idealmodel(model), where a_res is the reduced residual Helmholtz energy and a_ideal is the reduced ideal Helmholtz energy. You can mix and match ideal models if you provide:

[idealmodel](@ref)(model): extracts the ideal model from your Thermodynamic model

[a_res](@ref)(model,V,T,z): residual reduced Helmholtz free energy

Clapeyron.eos_res — Function

eos_res(model::EoSModel, V, T, z=SA[1.0])

Returns the residual Helmholtz free energy.

Inputs:

model::EoSModel Thermodynamic model to evaluate

V Total volume, in [m³]

T Temperature, in [K]

z mole amounts, in [mol], by default is @SVector [1.0]

Outputs:

Residual Helmholtz free energy, in [J]

by default, it calls R̄*T*∑(z)*(a_res(model,V,T,z)) where a_res is the reduced residual Helmholtz energy.

Clapeyron.idealmodel — Function

idealmodel(model::EoSModel)

retrieves the ideal model from the input's model.

Examples:

julia> pr = PR(["water"],idealmodel=MonomerIdeal)
PR{MonomerIdeal, PRAlpha, NoTranslation, vdW1fRule} with 1 component:
 "water"
Contains parameters: a, b, Tc, Pc, Mw 
julia> Clapeyron.idealmodel(pr)
MonomerIdeal with 1 component:
 "water"
Contains parameters: Mw

Clapeyron.a_res — Function

a_res(model::EoSModel, V, T, z,args...)

Reduced residual Helmholtz free energy.

Inputs:

model::EoSModel Thermodynamic model to evaluate

V Total volume, in [m³]

T Temperature, in [K]

z mole amounts, in [mol], by default is @SVector [1.0]

Outputs:

Residual Helmholtz free energy, no units

You can define your own EoS by adding a method to a_res that accepts your custom model.

Automatic Differenciation functions

All bulk properties in Clapeyron are calculated via a combination of these Automatic Differenciation Primitives over eos or eos_res

Clapeyron.∂f∂T — Function

∂f∂T(model,V,T,z=SA[1.0])

returns f and ∂f/∂T at constant total volume and composition, where f is the total helmholtz energy, given by eos(model,V,T,z)

Clapeyron.∂f∂V — Function

∂f∂V(model,V,T,z=SA[1.0])

returns f and ∂f/∂V at constant temperature and composition, where f is the total helmholtz energy, given by eos(model,V,T,z), and V is the total volume

Clapeyron.∂f — Function

∂f(model,V,T,z)

returns zeroth order (value) and first order derivative information of the total helmholtz energy (given by eos(model,V,T,z)). the result is given in two values:

grad_f,fval = ∂2f(model,V,T,z)

where:

fval   = f(V,T) = eos(model,V,T,z)

grad_f = [ ∂f/∂V; ∂f/∂T]

Where V is the total volume, T is the temperature and f is the total helmholtz energy.

Clapeyron.p∂p∂V — Function

p∂p∂V(model,V,T,z=SA[1.0])

returns p and ∂p/∂V at constant temperature, where p is the pressure = pressure(model,V,T,z) and V is the total Volume.

Clapeyron.∂2f — Function

∂2f(model,V,T,z)

returns zeroth order (value), first order and second order derivative information of the total helmholtz energy (given by eos(model,V,T,z)). the result is given in three values:

hess_f,grad_f,fval = ∂2f(model,V,T,z)

where: ``` fval = f(V,T) = eos(model,V,T,z)

grad_f = [ ∂f/∂V; ∂f/∂T]

hess_f = [ ∂²f/∂V²; ∂²f/∂V∂T ∂²f/∂V∂T; ∂²f/∂V²] ```

Where V is the total volume, T is the temperature and f is the total helmholtz energy.

Clapeyron.∂2p — Function

∂2p(model,V,T,z)

returns zeroth order (value), first order and second order derivative information of the pressure. the result is given in three values:

hess_p,grad_p,pval = ∂2p(model,V,T,z)

where: ``` pval = p(V,T) = pressure(model,V,T,z)

grad_p = [ ∂p/∂V; ∂p/∂T]

hess_p = [ ∂²p/∂V²; ∂²p/∂V∂T ∂²p/∂V∂T; ∂²p/∂V²] ```

Where V is the total volume, T is the temperature and p is the pressure.

Clapeyron.f_hess — Function

f_hess(model,V,T,z)

returns the second order volume (V) and temperature (T) derivatives of the total helmholtz energy (given by eos(model,V,T,z)). the result is given in a 2x2 SMatrix, in the form:

[ ∂²f/∂V² ∂²f/∂V∂T ∂²f/∂V∂T ∂²f/∂T²]

use this instead of the ∂2f if you only need second order information. ∂2f also gives zeroth and first order derivative information, but due to a bug in the used AD, it allocates more than necessary.

Clapeyron.∂²³f — Function

∂²³f(model,V,T,z=SA[1.0])

returns ∂²A/∂V² and ∂³A/∂V³, in a single ForwardDiff pass. used mainly in crit_pure objective function

Volume-Temperature Based Properties

Clapeyron.pressure — Function

pressure(model::EoSModel, V, T, z=SA[1.])

default units: [Pa]

Returns the pressure of the model at a given volume, temperature and composition, defined as:

p =  -∂A/∂V

Clapeyron.second_virial_coefficient — Function

second_virial_coefficient(model::EoSModel, T, z=SA[1.])

Calculates the second virial coefficient B, defined as:

B = lim(V->∞)[ V^2/RT *  (∂Aᵣ∂V + V*∂²Aᵣ∂V²) ]

where Aᵣ is the residual helmholtz energy.

Clapeyron.pip — Function

pip(model::EoSModel,V,T,z=[1.0])

Phase identification parameter Π. as described in 1. If Π > 1, then the phase is clasified as a liquid or a liquid-like vapor, being a vapor or vapor-like liquid otherwise.

This identification parameter fails at temperatures and pressures well aboVe the critical point.

Calculated as:

Π = V*((∂²p/∂V∂T)/(∂p/∂T) - (∂²p/∂V²)/(∂p/∂V))

G. Venkatarathnama, L.R. Oellrich, Identification of the phase of a fluid using partial derivatives of pressure, volume,and temperature without reference to saturation properties: Applications in phase equilibria calculations, Fluid Phase Equilibria 301 (2011) 225–233

Pressure-Temperature Based Bulk Properties

In general almost all bulk properties follow the pattern:

function property(model::EoSModel, p, T, z=SA[1.]; phase = :unknown,threaded=true)
    V = volume(model, p, T, z; phase=phase, threaded=threaded)
    return VT_property(model,V,T,z)
end

So, you can calculate the property with Volume-Temperature variables by calling VT_property(model,V,T,z). Another way to do this is by using units,provided by Unitful.jl:

using Unitful
r = 18u"kg/m^3"
T = 373.15"K"
prop = helholtz_free_energy(model,r,T,z,output = u"kJ")

Where r could be any molar or mass density, molar or mass volume, total volume or pressure. it also supports mass and mol amounts defined as units for the composition (z) If no units are provided for the composition, they will be considered moles.

Methods that require first order VT derivatives

Clapeyron.volume — Function

volume(model::EoSModel,p,T,z=SA[1.0];phase=:unknown,threaded=true)

calculates the volume (m³) of the compound modelled by model at a certain pressure,temperature and moles. phase is a Symbol that determines the initial volume root to look for:

If phase =:unknown (Default), it will return the physically correct volume root with the least gibbs energy.

If phase =:liquid, it will return the volume of the phase using a liquid initial point.

If phase =:vapor, it will return the volume of the phase using a gas initial point.

If phase =:stable, it will return the physically correct volume root with the least gibbs energy, and perform a stability test on the result.

All volume calculations are checked for mechanical stability, that is: dP/dV <= 0. The calculation of both volume roots can be calculated in serial (threaded=false) or in parallel (threaded=true)

Stability checks

The stability check is disabled by default. that means that the volume obtained just follows the the relation P = pressure(model,V,T,z). For single component models, this is alright, but phase splits (with different compositions that the input) can and will occur, meaning that the volume solution does not correspond to an existing phase. For unknown multicomponent mixtures, it is recommended to use a phase equilibrium procedure (like tp_flash) to obtain a list of valid compositions, and then perform a volume calculation over those compositions. You can also pass phase=:stable to perform the stability test inside the volume solver. Finally, you can perform the stability test after the volume solver:

v = volume(model,p,T,z)
isstable(model,v,T,z)

Clapeyron.helmholtz_free_energy — Function

helmholtz_free_energy(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Default units: [J]

Calculates the helmholtz free energy, defined as:

A = eos(model,V(p),T,z)

Internally, it calls Clapeyron.volume to obtain V and calculates the property via eos(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Clapeyron.molar_density — Function

molar_density(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

default units: [mol/m^3]

Calculates the molar density, defined as:

ρₙ =  ∑nᵢ/V

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_molar_density(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Clapeyron.mass_density — Function

mass_density(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

default units: [kg/m^3]

Calculates the molar density, defined as:

ρₙ =  Mr/V

Where Mr is the molecular weight of the model at the input composition.

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_mass_density(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Clapeyron.compressibility_factor — Function

compressibility_factor(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Calculates the compressibility factor Z, defined as:

Z = p*V(p)/R*T

The keywords phase and threaded are passed to Clapeyron.volume.

Clapeyron.gibbs_free_energy — Function

gibbs_free_energy(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Default units: [J]

Calculates the gibbs free energy, defined as:

G = A - V * ∂A/∂V

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_gibbs_free_energy(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Clapeyron.entropy — Function

entropy(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Default units: [J/K]

Calculates entropy, defined as:

S = -∂A/∂T

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_entropy(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Clapeyron.entropy_res — Function

entropy_res(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Default units: [J/K]

Calculates residual entropy, defined as:

S = -∂Ares/∂T

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_entropy_res(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Clapeyron.enthalpy — Function

enthalpy(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Default units: [J]

Calculates the enthalpy, defined as:

H = A - T * ∂A/∂T - V * ∂A/∂V

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_enthalpy(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Clapeyron.internal_energy — Function

internal_energy(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Default units: [J]

Calculates the internal energy, defined as:

U = A - T * ∂A/∂T

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_internal_energy(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Methods that require second order VT derivatives

Clapeyron.isochoric_heat_capacity — Function

isochoric_heat_capacity(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Default units: [J/T]

Calculates the isochoric heat capacity, defined as:

Cv = -T * ∂²A/∂T²

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_isochoric_heat_capacity(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Accurate ideal model required

This property requires at least second order ideal model temperature derivatives. If you are computing these properties, consider using a different ideal model than the BasicIdeal default (e.g. EoS(["species"];idealmodel=ReidIdeal)).

Clapeyron.isobaric_heat_capacity — Function

isobaric_heat_capacity(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Default units: [J/T]

Calculates the isobaric heat capacity, defined as:

Cp =  -T*(∂²A/∂T² - (∂²A/∂V∂T)^2 / ∂²A/∂V²)

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_isobaric_heat_capacity(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Accurate ideal model required

This property requires at least second order ideal model temperature derivatives. If you are computing these properties, consider using a different ideal model than the BasicIdeal default (e.g. EoS(["species"];idealmodel=ReidIdeal)).

Clapeyron.isothermal_compressibility — Function

isothermal_compressibility(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

default units: [Pa^-1]

Calculates the isothermal compressibility, defined as:

κT =  (V*∂p/∂V)^-1

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_isothermal_compressibility(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Clapeyron.isentropic_compressibility — Function

isentropic_compressibility(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

default units: [Pa^-1]

Calculates the isentropic compressibility, defined as:

κS =  (V*( ∂²A/∂V² - ∂²A/∂V∂T^2 / ∂²A/∂T² ))^-1

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_isentropic_compressibility(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Accurate ideal model required

This property requires at least second order ideal model temperature derivatives. If you are computing these properties, consider using a different ideal model than the BasicIdeal default (e.g. EoS(["species"];idealmodel=ReidIdeal)).

Clapeyron.speed_of_sound — Function

speed_of_sound(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

default units: [m/s]

Calculates the speed of sound, defined as:

c =  V * √(∂²A/∂V² - ∂²A/∂V∂T^2 / ∂²A/∂T²)/Mr)

Where Mr is the molecular weight of the model at the input composition.

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_speed_of_sound(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Accurate ideal model required

This property requires at least second order ideal model temperature derivatives. If you are computing these properties, consider using a different ideal model than the BasicIdeal default (e.g. EoS(["species"];idealmodel=ReidIdeal)).

Clapeyron.isobaric_expansivity — Function

isobaric_expansivity(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

default units: [K^-1]

Calculates the isobaric expansivity, defined as:

α =  -∂²A/∂V∂T / (V*∂²A/∂V²)

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_isobaric_expansivity(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Clapeyron.joule_thomson_coefficient — Function

joule_thomson_coefficient(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

default units: [K/Pa]

Calculates the joule thomson coefficient, defined as:

μⱼₜ =  -(∂²A/∂V∂T - ∂²A/∂V² * ((T*∂²A/∂T² + V*∂²A/∂V∂T) / (T*∂²A/∂V∂T + V*∂²A/∂V²)))^-1

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_joule_thomson_coefficient(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Accurate ideal model required

This property requires at least second order ideal model temperature derivatives. If you are computing these properties, consider using a different ideal model than the BasicIdeal default (e.g. EoS(["species"];idealmodel=ReidIdeal)).

Methods that first order composition derivatives

Clapeyron.chemical_potential — Function

chemical_potential(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Default units: [J/mol]

Calculates the chemical potential, defined as:

μᵢ = ∂A/∂nᵢ

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_chemical_potential(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Clapeyron.chemical_potential_res — Function

chemical_potential_res(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Default units: [J/mol]

Calculates the residual chemical potential, defined as:

μresᵢ = ∂Ares/∂nᵢ

Internally, it calls Clapeyron.volume to obtain V and calculates the property via VT_chemical_potential_res(model,V,T,z).

The keywords phase and threaded are passed to the volume solver.

Clapeyron.fugacity_coefficient — Function

fugacity_coefficient(model::EoSModel, p, T, z=SA[1.]; phase = :unknown, threaded=true)

Calculates the fugacity coefficient φᵢ, defined as:

log(φᵢ) =  μresᵢ/RT - log(Z)

Where μresᵢ is the vector of residual chemical potentials and Z is the compressibility factor.

The keywords phase and threaded are passed to Clapeyron.volume.

Mixing

Clapeyron.mixing — Function

mixing(model::EoSModel, p, T, z=SA[1.], property; phase = :unknown,threaded=true)

Calculates the mixing function for a specified property as:

f_mix = f(p,T,z) - ∑zᵢ*f_pureᵢ(p,T)

The keywords phase and threaded are passed to the volume solver.

Single component properties

Clapeyron.saturation_pressure — Function

saturation_pressure(model::EoSModel, T)
saturation_pressure(model::EoSModel,T,method::SaturationMethod)
saturation_pressure(model,T,x0::Union{Tuple,Vector})

Performs a single component saturation equilibrium calculation, at the specified temperature T, of one mol of pure sustance specified by model

Returns (p₀, Vₗ, Vᵥ) where p₀ is the saturation pressure (in Pa), Vₗ is the liquid saturation volume (in m³) and Vᵥ is the vapour saturation volume (in m³).

If the calculation fails, returns (NaN, NaN, NaN)

By default, it uses ChemPotVSaturation

Examples:

julia-repl

julia> pr = PR(["water"])
PR{BasicIdeal, PRAlpha, NoTranslation, vdW1fRule} with 1 component:
 "water"
Contains parameters: a, b, Tc, Pc, Mw

julia> saturation_pressure(pr,373.15) #default, uses Clapeyron.ChemPotVSaturation
(96099.38979351855, 2.2674781912892906e-5, 0.03201681565699426)

julia> saturation_pressure(pr,373.15,IsoFugacitySaturation()) #iso fugacity
(96099.38979351871, 2.2674781912892933e-5, 0.03201681565699359)

julia> saturation_pressure(pr,373.15,IsoFugacitySaturation(p0 = 1.0e5)) #iso fugacity, with starting point
(96099.38979351871, 2.2674781912892933e-5, 0.03201681565699547)

Clapeyron.ChemPotVSaturation — Type

ChemPotVSaturation <: SaturationMethod
ChemPotVSaturation(V0)
ChemPotVSaturation(;log10vl = nothing,
                    log10vv = nothing,
                    crit = nothing,
                    f_limit = 0.0,
                    atol = 1e-8,
                    rtol = 1e-12,
                    max_iters = 10^4)

Default saturation_pressure Saturation method used by Clapeyron.jl. It uses equality of Chemical Potentials with a volume basis. If no volumes are provided, it will use x0_sat_pure.

If those initial guesses fail and the specification is near critical point, it will try one more time, using Corresponding States instead.

V0 is [log10(Vₗ₀),log10(Vᵥ₀)] , where Vₗ₀ and Vᵥ₀ are initial guesses for the liquid and vapour volumes.

f_limit, atol, rtol, max_iters are passed to the non linear system solver.

Clapeyron.ChemPotDensitySaturation — Type

ChemPotDensitySaturation <: SaturationMethod
ChemPotDensitySaturation(;vl = nothing,
                        vv = nothing,
                        crit = nothing,
                        f_limit = 0.0,
                        atol = 1e-8,
                        rtol = 1e-12,
                        max_iters = 10^4)

Saturation method for saturation_pressure. It uses equality of Chemical Potentials with a density basis. If no volumes are provided, it will use x0_sat_pure.

vl and vl are initial guesses for the liquid and vapour volumes.

f_limit, atol, rtol, max_iters are passed to the non linear system solver.

Clapeyron.IsoFugacitySaturation — Type

IsoFugacitySaturation <: SaturationMethod
IsoFugacitySaturation(;p0 = nothing,
    vl = nothing,
    vv = nothing,
    crit = nothing,
    max_iters = 20,
    p_tol = sqrt(eps(Float64)))

Saturation method for saturation_pressure. Uses the isofugacity criteria. Ideal for Cubics or other EoS where the volume calculations are cheap. If p0 is not provided, it will be calculated via x0_psat.

Clapeyron.saturation_temperature — Function

saturation_pressure(model::EoSModel, p)
saturation_pressure(model::EoSModel, p, method::SaturationMethod)
saturation_pressure(model, p, T0::Number)

Performs a single component saturation temperature equilibrium calculation, at the specified pressure T, of one mol of pure sustance specified by model

Returns (T₀, Vₗ, Vᵥ) where p₀ is the saturation Temperature (in K), Vₗ is the liquid saturation volume (in m³) and Vᵥ is the vapour saturation volume (in m³).

If the calculation fails, returns (NaN, NaN, NaN)

By default, it uses AntoineSaturation

Examples:

julia-repl

julia> pr = PR(["water"])
PR{BasicIdeal, PRAlpha, NoTranslation, vdW1fRule} with 1 component:
 "water"
Contains parameters: a, b, Tc, Pc, Mw

julia> Ts,vl,vv = saturation_temperature(pr,1e5) # AntoineSaturation by default
(374.24014010712983, 2.269760164801948e-5, 0.030849387955737825)

julia> saturation_pressure(pr,Ts)
(100000.00004314569, 2.269760164804427e-5, 0.03084938795785433)

Clapeyron.ClapeyronSaturation — Type

ClapeyronSaturation <: SaturationMethod
ClapeyronSaturation(T0 = nothing, crit = nothing, satmethod = ChemPotVSaturation())

Saturation method for saturation_temperature. It solves iteratively saturation_temperature(model,Ti,satmethod) until convergence, by using the Clapeyron equation:

dp/dT = ΔS/ΔV

It descends from the critical point (or T0, if provided). Reliable, but slow.

It is recommended that T0 > Tsat, as the temperature decrease iteration series is more stable. Default method for saturation_temperature until Clapeyron 0.3.7

Clapeyron.AntoineSaturation — Type

AntoineSaturation <: SaturationMethod
AntoineSaturation(;T0 = nothing,
                    vl = nothing,
                    vv = nothing,
                    f_limit = 0.0,
                    atol = 1e-8,
                    rtol = 1e-12,
                    max_iters = 10^4)

Saturation method for saturation_temperature .Default method for saturation temperature from Clapeyron 0.3.7. It solves the Volume-Temperature system of equations for the saturation condition.

If only T0 is provided, vl and vv are obtained via x0_sat_pure. If T0 is not provided, it will be obtained via x0_saturation_temperature. It is recommended to overload x0_saturation_temperature, as the default starting point calls crit_pure, resulting in slower than ideal times.

f_limit, atol, rtol, max_iters are passed to the non linear system solver.

Clapeyron.enthalpy_vap — Function

enthalpy_vap(model::EoSModel, T,method = ChemPotVSaturation(x0_sat_pure(model,T)))

Calculates ΔH, the difference between saturated vapour and liquid enthalpies at temperature T, in J

Clapeyron.crit_pure — Function

crit_pure(model::EoSModel,x0=nothing)

Calculates the critical point of a single component modelled by model.

Returns (Tc, pc, Vc) where Tc is the critical temperature (in K), pc is the critical pressure (in Pa) and Vc is the critical volume (in m³)

Clapeyron.acentric_factor — Function

acentric_factor(model::EoSModel;crit = crit_pure(model), satmethod = ChemPotVSaturation())

calculates the acentric factor using its definition:

ω = -log10(psatᵣ) -1, at Tᵣ = 0.7

To do so, it calculates the critical temperature (using crit_pure) and performs a saturation calculation (with saturation_pressure(model,0.7Tc,satmethod))

Multi component properties

Clapeyron.bubble_pressure — Function

bubble_pressure(model::EoSModel, T, x; v0 = x0_bubble_pressure(model,T,x))

calculates the bubble pressure and properties at a given temperature. Returns a tuple, containing:

Bubble Pressure [Pa]

liquid volume at Bubble Point [m³]

vapour volume at Bubble Point [m³]

Gas composition at Bubble Point

Clapeyron.bubble_temperature — Function

bubble_temperature(model::EoSModel, p, x; T0 = x0_bubble_pressure(model,p,x))

calculates the bubble temperature and properties at a given pressure. Returns a tuple, containing:

Bubble Temperature [K]

liquid volume at Bubble Point [m³]

vapour volume at Bubble Point [m³]

Gas composition at Bubble Point

Clapeyron.dew_pressure — Function

dew_pressure(model::EoSModel, T, y; v0 = x0_dew_pressure(model,T,y))

calculates the dew pressure and properties at a given temperature. Returns a tuple, containing:

Dew Pressure [Pa]

liquid volume at Dew Point [m³]

vapour volume at Dew Point [m³]

Liquid composition at Dew Point

Clapeyron.dew_temperature — Function

dew_temperature(model::EoSModel, p, y; T0 = x0_dew_temperature(model,p,y))

calculates the dew temperature and properties at a given pressure. Returns a tuple, containing:

Dew Temperature [K]

liquid volume at Dew Point [m³]

vapour volume at Dew Point [m³]

Liquid composition at Dew Point

Clapeyron.azeotrope_pressure — Function

azeotrope_pressure(model::EoSModel, T; v0 = x0_azeotrope_pressure(model,T))

calculates the azeotrope pressure and properties at a given temperature. Returns a tuple, containing:

Azeotrope Pressure [Pa]

liquid volume at Azeotrope Point [m³]

vapour volume at Azeotrope Point [m³]

Azeotrope composition

Clapeyron.azeotrope_temperature — Function

azeotrope_temperature(model::EoSModel, T; v0 = x0_bubble_pressure(model,T,[0.5,0.5]))

Calculates the azeotrope temperature and properties at a given pressure. Returns a tuple, containing:

Azeotrope Temperature [K]

liquid volume at Azeotrope Point [m³]

vapour volume at Azeotrope Point [m³]

Azeotrope composition

Clapeyron.LLE_pressure — Function

LLE_pressure(model::EoSModel, T, x; v0 = x0_LLE_pressure(model,T,x))

calculates the Liquid-Liquid equilibrium pressure and properties at a given temperature.

Returns a tuple, containing:

LLE Pressure [Pa]

liquid volume of composition x₁ = x at LLE Point [m³]

liquid volume of composition x₂ at LLE Point [m³]

Liquid composition x₂

Clapeyron.LLE_temperature — Function

LLE_temperature(model::EoSModel, p, x; T0 = x0_LLE_temperature(model,p,x))

calculates the Liquid-Liquid equilibrium temperature and properties at a given pressure.

Returns a tuple, containing:

LLE Pressure [Pa]

liquid volume of composition x₁ = x at LLE Point [m³]

liquid volume of composition x₂ at LLE Point [m³]

Liquid composition x₂

Clapeyron.VLLE_pressure — Function

VLLE_pressure(model::EoSModel, T; v0 = x0_LLE_pressure(model,T))

calculates the Vapor-Liquid-Liquid equilibrium pressure and properties of a binary mixture at a given temperature.

Returns a tuple, containing:

VLLE Pressure [Pa]

Liquid volume of composition x₁ at VLLE Point [m³]

Liquid volume of composition x₂ at VLLE Point [m³]

Vapour volume of composition y at VLLE Point [m³]

Liquid composition x₁

Liquid composition x₂

Liquid composition y

Clapeyron.VLLE_temperature — Function

VLLE_temperature(model::EoSModel, p; T0 = x0_LLE_temperature(model,p))

calculates the Vapor-Liquid-Liquid equilibrium temperature and properties of a binary mixture at a given temperature.

Returns a tuple, containing:

VLLE temperature [K]

Liquid volume of composition x₁ at VLLE Point [m³]

Liquid volume of composition x₂ at VLLE Point [m³]

Vapour volume of composition y at VLLE Point [m³]

Liquid composition x₁

Liquid composition x₂

Liquid composition y

Clapeyron.crit_mix — Function

crit_mix(model::EoSModel,z;v0=x=x0_crit_mix(model,z))

Returns the critical mixture point at a ginven composition.

Returns a tuple, containing:

Critical Mixture Temperature [K]

Critical Mixture Pressure [Pa]

Critical Mixture Volume [m³]

Clapeyron.UCEP_mix — Function

UCEP_mix(model::EoSModel;v0=x0_UCEP_mix(model))

Calculates the Upper Critical End Point of a binary mixture.

returns:

UCEP Temperature [K]

UCEP Pressure [Pa]

liquid volume at UCEP Point [m³]

vapour volume at UCEP Point [m³]

liquid molar composition at UCEP Point

vapour molar composition at UCEP Point

Clapeyron.UCST_mix — Function

UCST_mix(model::EoSModel,T;v0=x0_UCST_mix(model,T))

Calculates the Upper critical solution point of a mixture at a given Temperature.

returns:

UCST Pressure [Pa]

volume at UCST Point [m³]

molar composition at UCST Point

Clapeyron.gibbs_solvation — Function

gibbs_solvation(model::EoSModel, T)

Calculates the solvation free energy as:

g_solv = -R̄*T*log(K)

where the first component is the solvent and second is the solute.

Consistency and Stability

Clapeyron.gibbs_duhem — Function

gibbs_duhem(model,V,T,z=[1.0])

performs a Gibbs-Duhem check on the input conditions:

∑zᵢμᵢ - G ≈ 0

Where G is the total gibbs energy. it can help diagnose if a user-defined eos is consistent.

return |∑zᵢμᵢ - G|, ∑zᵢμᵢ and G at the specified conditions.

Clapeyron.isstable — Function

isstable(model,V,T,z)::Bool

Performs stability tests for a (V,T,z) pair, and warn if any tests fail. returns true/false.

Checks:

mechanical stability: isothermal compressibility is not negative.

diffusive stability: all eigenvalues of ∂²A/∂n² are positive.

Clapeyron.mechanical_stability — Function

mechanical_stability(model,V,T,z)::Bool

Performs a mechanical stability for a (V,T,z) pair, returns true/false. Checks if isothermal compressibility is not negative.

Clapeyron.diffusive_stability — Function

diffusive_stability(model,V,T,z)::Bool

Performs a diffusive stability for a (V,T,z) pair, returns true/false. Checks if all eigenvalues of ∂²A/∂n² are positive.

Clapeyron.tpd — Function

tpd(model,p,T,z;verbose=false)

Calculates the Tangent plane distance function (tpd). It returns:

a vector with trial phase compositions where tpd < 0

a vector with the tpd values

a vector with symbols indicating the phase of the input composition

a vector with symbols indicating the phase of the trial composition

It iterates over each two-phase combination, starting from pure trial compositions, it does succesive substitution, then Gibbs optimization.

If the vectors are empty, then the procedure couldn't find a negative tpd. That is an indication that the phase is (almost) surely stable.

TP Flash

Clapeyron.tp_flash — Function

tp_flash(model, p, T, n, method::TPFlashMethod = DETPFlash())

Routine to solve non-reactive multicomponent flash problem. The default method uses Global Optimization. see DETPFlash

Inputs:

T, Temperature

p, Pressure

n, vector of number of moles of each species

Outputs - Tuple containing:

xᵢⱼ, Array of mole fractions of species j in phase i

nᵢⱼ, Array of mole numbers of species j in phase i, [mol]

G, Gibbs Free Energy of Equilibrium Mixture [J]

Clapeyron.TPFlashMethod — Type

TPFlashMethod

Abstract type for tp_flash routines. it requires defining numphases(method) and tp_flash_impl(model,p,T,n,method). If the method accept component-dependent inputs, it also should define index_reduction(method,nonzero_indices)

Clapeyron.DETPFlash — Type

DETPFlash(;numphases = 2;max_steps = 1e4*(numphases-1),population_size =20,time_limit = Inf,verbose = false, logspace = false)

Method to solve non-reactive multicomponent flash problem by finding global minimum of Gibbs Free Energy via Differential Evolution.

User must assume a number of phases, numphases. If true number of phases is smaller than numphases, model should predict either (a) identical composition in two or more phases, or (b) one phase with negligible total number of moles. If true number of phases is larger than numphases, a thermodynamically unstable solution will be predicted.

The optimizer will stop at max_steps evaluations or at time_limit seconds

Clapeyron.RRTPFlash — Type

RRTPFlash{T}(;K0 = nothing,rtol= 1e-10,atol = 1e-10,max_iters = 100)

Method to solve non-reactive multicomponent flash problem by using successive substitution in the Rachford-Rice equation.

Only two phases are supported. if K0 is nothing, it will be calculated via the Wilson correlation.

The optimizer will stop at max_iters evaluations, when the absolute tolerance is less than atol or when the relative tolerance is less than rtol

Clapeyron.MichelsenTPFlash — Type

MichelsenTPFlash{T}(;kwargs...)

Method to solve non-reactive multicomponent flash problem by Michelsen's method.

Only two phases are supported. if K0 is nothing, it will be calculated via the Wilson correlation.

Keyword Arguments:

equilibrium = equilibrium type ":vle" for liquid vapor equilibria, ":lle" for liquid liquid equilibria

K0 (optional), initial guess for the constants K

x0 (optional), initial guess for the composition of phase x

y0 = optional, initial guess for the composition of phase y

vol0 = optional, initial guesses for phase x and phase y volumes

K_tol = tolerance to stop the calculation

ss_iters = number of Successive Substitution iterations to perform

second_order = wheter to solve the gibbs energy minimization using the analytical hessian or not

Clapeyron.numphases — Function

numphases(method::TPFlashMethod)

return the number of phases supported by the TP flash method. by default its set to 2. it the method allows it, you can set the number of phases by doing method(;numphases = n).

Initial guess functions

Clapeyron.lb_volume — Function

lb_volume(model::EoSModel,z=SA[1.0])

Returns the lower bound volume.

It has different meanings depending on the Equation of State, but symbolizes the minimum allowable volume at a certain composition:

SAFT EoS: the packing volume

Cubic EoS, covolume (b) parameter

On empiric equations of state, the value is chosen to match the volume of the conditions at maximum pressure and minimum temperature , but the equation itself normally can be evaluated at lower volumes.

On SAFT and Cubic EoS, volumes lower than lb_volume will likely error.

The lower bound volume is used for guesses of liquid volumes at a certain pressure, saturated liquid volumes and critical volumes.

Clapeyron.T_scale — Function

T_scale(model::EoS,z=SA[1.0])

Represents a temperature scaling factor.

On any EoS based on Critical parameters (Cubic or Empiric EoS), the temperature scaling factor is chosen to be the critical temperature.

On SAFT or other molecular EoS, the temperature scaling factor is chosen to be a function of the potential depth ϵ.

Used as scaling factors in saturation_pressure and as input for solving crit_pure

Clapeyron.p_scale — Function

p_scale(model::SAFTModel,z=SA[1.0])

Represents a pressure scaling factor

On any EoS based on Critical parameters (Cubic or Empiric EoS), the pressure scaling factor is chosen to be a function of the critical pressure.

On SAFT or other molecular EoS, the temperature scaling factor is chosen to a function of ∑(zᵢϵᵢ(σᵢᵢ)³)

Used as scaling factors in saturation_pressure and as input for solving crit_pure

Clapeyron.x0_volume — Function

x0_volume(model,p,T,z; phase = :unknown)

Returns an initial guess of the volume at a pressure, temperature, composition and suggested phase.

If the suggested phase is :unkwown or :liquid, calls x0_volume_liquid.

If the suggested phase is :gas, calls x0_volume_gas.

Clapeyron.x0_volume_liquid — Function

x0_volume_liquid(model,T,z)

Returns an initial guess to the liquid volume, dependent on temperature and composition. by default is 1.25 times lb_volume.

Clapeyron.x0_volume_gas — Function

x0_volume_gas(model,p,T,z)

Returns an initial guess to the gas volume, depending of pressure, temperature and composition. by default uses volume_virial

Clapeyron.volume_virial — Function

volume_virial(model::EoSModel,p,T,z=SA[1.0])
volume_virial(B::Real,p,T,z=SA[1.0])

Calculates an aproximation to the gas volume at specified pressure, volume and composition, by aproximating:

Z(v) ≈ 1 + B(T)/v

where Z is the compressibility factor and B is the second virial coefficient. If B>0, (over the inversion temperature) returns NaN. If the solution to the problem is complex (Z = 1 + B/v implies solving a quadratic polynomial), returns -2*B. If you pass an EoSModel as the first argument, B will be calculated from the EoS at the input T. You can provide your own second virial coefficient instead of a model.

Clapeyron.x0_sat_pure — Function

x0_sat_pure(model::EoSModel,T,z=SA[1.0])

Returns a 2-tuple corresponding to (log10(Vₗ),log10(Vᵥ)), where Vₗ and Vᵥ are the liquid and vapor initial guesses.

Used in saturation_pressure methods that require initial volume guesses.

It can be overloaded to provide more accurate estimates if necessary.

Clapeyron.x0_psat — Function

x0_psat(model::EoSModel, T,crit = nothing)

Initial point for saturation pressure, given the temperature and V,T critical coordinates. On moderate pressures it will use a Zero Pressure initialization. On pressures near the critical point it will switch to spinodal finding.

Used in saturation_pressure methods that require initial pressure guesses.

if the initial temperature is over the critical point, it returns NaN.

It can be overloaded to provide more accurate estimates if necessary.

Clapeyron.x0_saturation_temperature — Function

x0_saturation_temperature(model::EoSModel,p)

Returns a 3-tuple corresponding to (T,Vₗ,Vᵥ), T is the initial guess for temperature and Vₗ and Vᵥ are the liquid and vapor initial guesses. Used in saturation_temperature with AntoineSaturation.

Clapeyron.antoine_coef — Function

antoine_coef(model)

should return a 3-Tuple containing reduced Antoine Coefficients. The Coefficients follow the correlation:

lnp̄ = log(p / p_scale(model))
T̃ = T/T_scale(model)
lnp̄ = A - B/(T̄ + C))

By default returns nothing. This is to use alternative methods in case Antoine coefficients aren't available. Used mainly in single and multicomponent temperature calculations.

Clapeyron.x0_crit_pure — Function

x0_crit_pure(model::SAFTModel)

Returns a 2-tuple corresponding to (k,log10(Vc0)), where k is Tc0/T_scale(model,z)