Index

Clapeyron.BasicIdeal

Clapeyron.IAPWS95Ideal

Clapeyron.JobackIdeal

Clapeyron.LJRefIdeal

Clapeyron.MonomerIdeal

Clapeyron.ReidIdeal

Clapeyron.WalkerIdeal

Correlation Models

Correlation models are, as their name says, fitted equations that express one property of a compound. they meant to be used in conjunction with other models (like Activity models that require a saturated liquid volume), or via a CompositeModel. because they only overload one property, the way to define a correlation is different than normal EoSModels

Saturation Correlations

Saturation Correlations are any EoSModel that are subtypes of SaturationModel. return psat(T) and the upper limit (Tc,Pc) pair. To define saturation correlations, you need to overload:

function crit_pure(model::MySaturationModel <: SaturationModel)
    ...
    return (Tc,Pc,NaN)
end

function Clapeyron.saturation_pressure_impl(model::MySaturationModel <: SaturationModel,T,::SaturationCorrelation)
    ...
    return (psat,NaN,NaN)
end

Saturation Models and Types

Clapeyron.SaturationModel — Type

SaturationModel <: EoSModel

Abstract type for Saturation correlation models.

Clapeyron.SaturationCorrelation — Type

SaturationCorrelation <: SaturationMethod

saturation method used for dispatch on saturation correlations.

Missing docstring.

Missing docstring for Clapeyron.LeeKeslerSat. Check Documenter's build log for details.

Clapeyron.DIPPR101Sat — Type

DIPPR101Sat <: SaturationModel

DIPPR101Sat(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

pc: Single Parameter (Float64) - Critical Pressure [Pa]

A: Single Parameter (Float64)

B: Single Parameter (Float64)

C: Single Parameter (Float64)

D: Single Parameter (Float64)

E: Single Parameter (Float64)

Tmin: Single Parameter (Float64) - mininum Temperature range [K]

Tmax: Single Parameter (Float64) - maximum Temperature range [K]

Model Parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

A: Single Parameter (Float64)

B: Single Parameter (Float64)

C: Single Parameter (Float64)

D: Single Parameter (Float64)

E: Single Parameter (Float64)

Tmin: Single Parameter (Float64) - mininum Temperature range [K]

Tmax: Single Parameter (Float64) - maximum Temperature range [K]

Description

DIPPR 101 Equation for saturation pressure:

psat(T) = exp(A + B/T + C•log(T) + D•T^E)

References

Design Institute for Physical Properties, 1996. DIPPR Project 801 DIPPR/AIChE

Liquid Volume Correlations

Liquid Volume Correlations are any EoSModel that are subtypes of LiquidVolumeModel. They return volume(model,p,T,z, phase = :liquid).

Virial Models

Virial models are defined in terms of the second virial coefficient, B(T,z). this allows for a fast calculation of volume(model,p,T,z, phase = :gas). they cannot represent the liquid phase.