Index

Clapeyron.BMAlpha

Clapeyron.CPAAlpha

Clapeyron.HVRule

Clapeyron.KayRule

Clapeyron.LCVMRule

Clapeyron.MHV1Rule

Clapeyron.MHV2Rule

Clapeyron.MTAlpha

Clapeyron.MTTranslation

Clapeyron.NoAlpha

Clapeyron.NoTranslation

Clapeyron.PR

Clapeyron.PR78Alpha

Clapeyron.PRAlpha

Clapeyron.PSRKRule

Clapeyron.PenelouxTranslation

Clapeyron.RK

Clapeyron.RKAlpha

Clapeyron.RackettTranslation

Clapeyron.SoaveAlpha

Clapeyron.TwuAlpha

Clapeyron.VTPRRule

Clapeyron.WSRule

Clapeyron.sCPAAlpha

Clapeyron.vdW

Clapeyron.vdW1fRule

Clapeyron.PR78

Clapeyron.PSRK

Clapeyron.SRK

Clapeyron.UMRPR

Clapeyron.VTPR

Clapeyron.ab_premixing

Clapeyron.mixing_rule

Clapeyron.translation

Clapeyron.α_function

Cubic Models

All cubic models in Clapeyron.jl follow a common evaluation order:

function CubicModel(args...)
    #get params for database, initialize other models, etc
    #we need Tc,pc and kij for the generation of the matrices aᵢⱼ and bᵢⱼ
    a,b = a,b = ab_premixing(CubicModel,mixing_model,Tc,pc,k)
    #rest of the initialization, return model
end

function cubic_ab(model::CubicModel,V,T,z=SA[1.0])
    invn2 = (one(n)/n)^2
    a = model.params.a.values
    b = model.params.b.values
    α = α_function(model,V,T,z,model.alpha)
    c = translation(model,V,T,z,model.translation)
    ā,b̄,c̄ = mixing_rule(model,V,T,z,model.mixing,α,a,b,c)
    return ā ,b̄, c̄
end

function a_res(model::CubicModel,V,T,z)
    ā ,b̄, c̄ = cubic_ab(model,V,T,z)
    #depends on the specific EoS
    return result
end

A Mixing Rule Model creates aᵢⱼ and bᵢⱼ from the critical temperature, critical pressure and a matrix of pair coefficients.

An Alpha Model creates a vector of αᵢ(T) values

A Translation Model creates a vector of cᵢ values

The same Mixing rule, given aᵢⱼ, bᵢⱼ, αᵢ(T) and cᵢ returns the the mixture values of ā, b̄ and c̄ that are then used by the corresponding cubic model. a Mixing Rule can contain activity models to participate in the mixing (for example, Huron-Vidal rules)

Main Models

Clapeyron.vdW — Type

vdW(components::Vector{String};
idealmodel=BasicIdeal,
alpha = NoAlpha,
mixing = vdW1fRule,
activity=nothing,
translation=NoTranslation,
userlocations=String[], 
ideal_userlocations=String[],
alpha_userlocations = String[],
mixing_userlocations = String[],
activity_userlocations = String[],
translation_userlocations = String[],
verbose=false)

Input parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

k: Pair Parameter (Float64)

Model Parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

a: Pair Parameter (Float64)

b: Pair Parameter (Float64)

Input models

idealmodel: Ideal Model

alpha: Alpha model

mixing: Mixing model

activity: Activity Model, used in the creation of the mixing model.

translation: Translation Model

Description

van der Wals Equation of state.

P = RT/(V-Nb) + a•α(T)/V²

References

van der Waals JD. Over de Continuiteit van den Gasen Vloeistoftoestand. PhD thesis, University of Leiden; 1873

Clapeyron.RK — Type

RK(components::Vector{String}; idealmodel=BasicIdeal,
alpha = PRAlpha,
mixing = vdW1fRule,
activity=nothing,
translation=NoTranslation,
userlocations=String[],
ideal_userlocations=String[],
alpha_userlocations = String[],
mixing_userlocations = String[],
activity_userlocations = String[],
translation_userlocations = String[],
verbose=false)

Input parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

k: Pair Parameter (Float64)

Model Parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

a: Pair Parameter (Float64)

b: Pair Parameter (Float64)

Input models

idealmodel: Ideal Model

alpha: Alpha model

mixing: Mixing model

activity: Activity Model, used in the creation of the mixing model.

translation: Translation Model

Description

Redlich-Kwong Equation of state.

P = RT/(V-Nb) + a•α(T)/(V(V+Nb))

References

Redlich, O., & Kwong, J. N. S. (1949). On the thermodynamics of solutions; an equation of state; fugacities of gaseous solutions. Chemical Reviews, 44(1), 233–244. doi:10.1021/cr60137a013

Clapeyron.PR — Type

PR(components::Vector{String}; idealmodel=BasicIdeal,
alpha = PRAlpha,
mixing = vdW1fRule,
activity=nothing,
translation=NoTranslation,
userlocations=String[],
ideal_userlocations=String[],
alpha_userlocations = String[],
mixing_userlocations = String[],
activity_userlocations = String[],
translation_userlocations = String[],
verbose=false)

Input parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

k: Pair Parameter (Float64)

Model Parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

a: Pair Parameter (Float64)

b: Pair Parameter (Float64)

Input models

idealmodel: Ideal Model

alpha: Alpha model

mixing: Mixing model

activity: Activity Model, used in the creation of the mixing model.

translation: Translation Model

Description

Peng-Robinson Equation of state.

P = RT/(V-Nb) + a•α(T)/(V-Nb₁)(V-Nb₂)
b₁ = (1 + √2)b
b₂ = (1 - √2)b

References

Peng, D.Y., & Robinson, D.B. (1976). A New Two-Constant Equation of State. Industrial & Engineering Chemistry Fundamentals, 15, 59-64. doi:10.1021/I160057A011

Clapeyron.SRK — Function

SRK(components::Vector{String}; idealmodel=BasicIdeal,
alpha = SoaveAlpha,
mixing = vdW1fRule,
activity = nothing,
translation=NoTranslation,
userlocations=String[], 
ideal_userlocations=String[],
alpha_userlocations = String[],
mixing_userlocations = String[],
activity_userlocations = String[],
translation_userlocations = String[],
verbose=false)

Description

Soave-Redlich-Kwong equation of state. it uses the following models:

Translation Model: NoTranslation

Alpha Model: SoaveAlpha

Mixing Rule Model: vdW1fRule

References

Soave, G. (1972). Equilibrium constants from a modified Redlich-Kwong equation of state. Chemical Engineering Science, 27(6), 1197–1203. doi:10.1016/0009-2509(72)80096-4

Clapeyron.PSRK — Function

function PSRK(components::Vector{String}; idealmodel=BasicIdeal,
alpha = SoaveAlpha,
mixing = PSRKRule,
activity = PSRKUNIFAC,
translation=PenelouxTranslation,
userlocations=String[], 
ideal_userlocations=String[],
alpha_userlocations = String[],
mixing_userlocations = String[],
activity_userlocations = String[],
translation_userlocations = String[],
verbose=false)

Description

Predictive Soave-Redlich-Kwong equation of state. it uses the following models:

Translation Model: NoTranslation

Alpha Model: SoaveAlpha

Mixing Rule Model: PSRKRule with PSRKUNIFAC activity model

References

Horstmann, S., Jabłoniec, A., Krafczyk, J., Fischer, K., & Gmehling, J. (2005). PSRK group contribution equation of state: comprehensive revision and extension IV, including critical constants and α-function parameters for 1000 components. Fluid Phase Equilibria, 227(2), 157–164. doi:10.1016/j.fluid.2004.11.002

Clapeyron.PR78 — Function

PR78(components::Vector{String}; idealmodel=BasicIdeal,
alpha = PR78Alpha,
mixing = vdW1fRule,
activity = nothing,
translation=NoTranslation,
userlocations=String[], 
ideal_userlocations=String[],
alpha_userlocations = String[],
mixing_userlocations = String[],
verbose=false)

Peng Robinson (1978) equation of state. it uses the following models:

Translation Model: PR78Alpha

Alpha Model: PR78Alpha

Mixing Rule Model: vdW1fRule

References

Robinson DB, Peng DY. The characterization of the heptanes and heavier fractions for the GPA Peng-Robinson programs. Tulsa: Gas Processors Association; 1978

Clapeyron.UMRPR — Function

UMRPR(components::Vector{String}; idealmodel=BasicIdeal,
alpha = MTAlpha,
mixing = UMRRule,
activity = UNIFAC,
translation=MTTranslation,
userlocations=String[], 
ideal_userlocations=String[],
alpha_userlocations = String[],
mixing_userlocations = String[],
verbose=false)

Universal Mixing Rule Peng Robinson equation of state. it uses the following models:

Translation Model: MTTranslation

Alpha Model: MTAlpha

Mixing Rule Model: UMRRule with UNIFAC activity

References

Voutsas, E., Magoulas, K., & Tassios, D. (2004). Universal mixing rule for cubic equations of state applicable to symmetric and asymmetric systems: Results with the Peng−Robinson equation of state. Industrial & Engineering Chemistry Research, 43(19), 6238–6246. doi:10.1021/ie049580p

Clapeyron.VTPR — Function

VTPR(components::Vector{String}; idealmodel=BasicIdeal,
mixing = VTPRRule,
alpha = TwuAlpha,
translation = RackettTranslation,
activity = VTPRUNIFAC,
userlocations=String[], 
ideal_userlocations=String[],
alpha_userlocations = String[],
mixing_userlocations = String[],
verbose=false)

Volume-translated Peng Robinson equation of state. it uses the following models:

Translation Model: RackettTranslation

Alpha Model: TwuAlpha

Mixing Rule Model: VTPRRule with VTPRUNIFAC activity

References

Ahlers, J., & Gmehling, J. (2001). Development of an universal group contribution equation of state. Fluid Phase Equilibria, 191(1–2), 177–188. doi:10.1016/s0378-3812(01)00626-4

Alpha (α(T)) Models

Clapeyron.α_function — Function

α_function(model::CubicModel,V,T,z,αmodel::AlphaModel)

Interface function used in cubic models. it should return a vector of αᵢ(T).

Example:

function α_function(model::CubicModel,V,T,z,alpha_model::RKAlphaModel)
    return 1 ./ sqrt.(T ./ Tc)
end

Clapeyron.NoAlpha — Type

NoAlpha(args...) <: NoAlphaModel

Input Parameters

None

Description

Cubic alpha (α(T)) model. Default for vdW EoS

αᵢ = 1 ∀ i

Clapeyron.RKAlpha — Type

RKAlpha <: RKAlphaModel

RKAlpha(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

w: Single Parameter (Float64)

Model Parameters

acentricfactor: Single Parameter (Float64)

Description

Cubic alpha (α(T)) model. Default for RK EoS.

αᵢ = 1/√(Trᵢ)
Trᵢ = T/Tcᵢ

Clapeyron.SoaveAlpha — Type

SoaveAlpha <: SoaveAlphaModel

SoaveAlpha(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

w: Single Parameter (Float64)

Model Parameters

acentricfactor: Single Parameter (Float64)

Description

Cubic alpha (α(T)) model. Default for SRK EoS.

αᵢ = (1+mᵢ(1-√(Trᵢ)))^2
Trᵢ = T/Tcᵢ
mᵢ = 0.480 + 1.547ωᵢ - 0.176ωᵢ^2

to use different polynomial coefficients for mᵢ, overload Clapeyron.α_m(::CubicModel,::SoaveAlphaModel) = (c₁,c₂,...cₙ)

Clapeyron.PRAlpha — Type

PRAlpha <: SoaveAlphaModel

PRAlpha(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

w: Single Parameter (Float64)

Model Parameters

acentricfactor: Single Parameter (Float64)

Description

Cubic alpha (α(T)) model. Default for PR EoS.

αᵢ = (1+mᵢ(1-√(Trᵢ)))^2
Trᵢ = T/Tcᵢ
mᵢ = 0.37464 + 1.54226ωᵢ - 0.26992ωᵢ^2

Clapeyron.PR78Alpha — Type

PR78Alpha <: PR78AlphaModel

PR78Alpha(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

w: Single Parameter (Float64)

Model Parameters

acentricfactor: Single Parameter (Float64)

Description

Cubic alpha (α(T)) model. Default for PR78 EoS.

αᵢ = (1+mᵢ(1-√(Trᵢ)))^2
Trᵢ = T/Tcᵢ
if ωᵢ ≤ 0.491
    mᵢ = 0.37464 + 1.54226ωᵢ - 0.26992ωᵢ^2
else
    mᵢ = 0.379642 + 1.487503ωᵢ - 0.164423ωᵢ^2 - 0.016666ωᵢ^3

Clapeyron.CPAAlpha — Type

CPAAlpha <: CPAAlphaModel

CPAAlpha(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

c1: Single Parameter (Float64)

Description

Cubic alpha (α(T)) model. Default for CPA EoS.

αᵢ = (1+c¹ᵢ(1-√(Trᵢ)))^2

Clapeyron.sCPAAlpha — Type

sCPAAlpha <: sCPAAlphaModel

CPAAlpha(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

c1: Single Parameter

Description

Cubic alpha (α(T)) model. Default for sCPA EoS.

αᵢ = (1+c¹ᵢ(1-√(Trᵢ)))^2

Clapeyron.MTAlpha — Type

MTAlpha <: MTAlphaModel

MTAlpha(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

w: Single Parameter (Float64)

Model Parameters

acentricfactor: Single Parameter (Float64)

Description

Cubic alpha (α(T)) model. Default for UMRPR EoS.

αᵢ = (1+mᵢ(1-√(Trᵢ)))^2
Trᵢ = T/Tcᵢ
mᵢ = 0.384401 + 1.52276ωᵢ - 0.213808ωᵢ^2 + 0.034616ωᵢ^3 - 0.001976ωᵢ^4 

References

Magoulas, K., & Tassios, D. (1990). Thermophysical properties of n-Alkanes from C1 to C20 and their prediction for higher ones. Fluid Phase Equilibria, 56, 119–140. doi:10.1016/0378-3812(90)85098-u

Clapeyron.BMAlpha — Type

BMAlpha <: BMAlphaModel

MTAlpha(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

w: Single Parameter (Float64)

Model Parameters

acentricfactor: Single Parameter (Float64)

Description

Cubic alpha (α(T)) model. Default for UMRPR EoS.

if Trᵢ > 1
    αᵢ = (exp((1-2/(2+mᵢ))*(1-Trᵢ^(1+mᵢ/2))))^2
else
    αᵢ = (1+mᵢ*(1-√(Trᵢ)))^2

Trᵢ = T/Tcᵢ

for PR models:
    mᵢ = 0.37464 + 1.54226ωᵢ - 0.26992ωᵢ^2
for RK models:
    mᵢ = 0.480 + 1.547ωᵢ - 0.176ωᵢ^2

References

.M. Boston, P.M. Mathias, Proceedings of the 2nd International Conference on Phase Equilibria and Fluid Properties in the Chemical Process Industries, West Berlin, March, 1980, pp. 823–849

Clapeyron.TwuAlpha — Type

TwuAlpha <: TwuAlphaModel

TwuAlpha(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

M: Single Parameter

N: Single Parameter

L: Single Parameter

Model Parameters

M: Single Parameter

N: Single Parameter

L: Single Parameter

Description

Cubic alpha (α(T)) model. Default for VTPR EoS.

αᵢ = Trᵢ^(N*(M-1))*exp(L*(1-Trᵢ^(N*M))
Trᵢ = T/Tcᵢ

References

Twu, C. H., Lee, L. L., & Starling, K. E. (1980). Improved analytical representation of argon thermodynamic behavior. Fluid Phase Equilibria, 4(1–2), 35–44. doi:10.1016/0378-3812(80)80003-3

Volume Translation Models

Clapeyron.translation — Function

translation(model::CubicModel,V,T,z,translation_model::TranslationModel)

Interface function used in cubic models. it should return a vector of cᵢ. such as Ṽ = V + mixing(c,z)

Example:

function α_function(model::CubicModel,V,T,z,translation_model::RackettTranslation)
    Tc = model.params.Tc.values
    Pc = model.params.Pc.values
    Vc = translation_model.params.Vc.values
    R = Clapeyron.R̄
    Zc = Pc .* Vc ./ (R .* Tc)
    c = 0.40768 .* (0.29441 .- Zc) .* R .* Tc ./ Pc
    return c
end

Clapeyron.NoTranslation — Type

NoTranslation(args...) <: TranslationModel

Input Parameters

None

Description

Default volume translation model for cubic models. it performs no translation:

V = V₀ + mixing_rule(cᵢ)
cᵢ = 0 ∀ i

Clapeyron.RackettTranslation — Type

RackettTranslation <: RackettTranslationModel

RackettTranslation(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

vc: Single Parameter (Float64) - Critical Volume [m³/mol]

Description

Rackett Translation model for cubics:

V = V₀ + mixing_rule(cᵢ)
cᵢ = 0.40768*RTcᵢ/Pcᵢ*(0.29441-Zcᵢ)
Zcᵢ = Pcᵢ*Vcᵢ/(RTcᵢ)

References

Rackett, H. G. (1970). Equation of state for saturated liquids. Journal of Chemical and Engineering Data, 15(4), 514–517. doi:10.1021/je60047a012

Clapeyron.PenelouxTranslation — Type

PenelouxTranslation <: PenelouxTranslationModel

PenelouxTranslation(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

vc: Single Parameter (Float64) - Critical Volume [m³/mol]

Description

Peneloux Translation model for cubics:

V = V₀ + mixing_rule(cᵢ)
cᵢ = -0.252*RTcᵢ/Pcᵢ*(1.5448Zcᵢ - 0.4024)
Zcᵢ = Pcᵢ*Vcᵢ/(RTcᵢ)

References

Péneloux A, Rauzy E, Fréze R. (1982) A consistent correction for Redlich‐Kwong‐Soave volumes. Fluid Phase Equilibria 1, 8(1), 7–23. doi:10.1016/0378-3812(82)80002-2

Clapeyron.MTTranslation — Type

MTTranslation <: MTTranslationModel

MTTranslation(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

w: Single Parameter (Float64)

Model Parameters

acentricfactor: Single Parameter (Float64)

Description

Magoulas Tassios Translation model for cubics:

V = V₀ + mixing_rule(cᵢ)
cᵢ = T₀ᵢ+(T̄cᵢ-T̄₀ᵢ)*exp(β*abs(1-Trᵢ))
Trᵢ = T/T̄cᵢ
T̄cᵢ = (RTcᵢ/Pcᵢ)*(0.3074-Zcᵢ)
T̄₀ᵢ = (RTcᵢ/Pcᵢ)*(-0.014471 + 0.067498ωᵢ - 0.084852ωᵢ^2 + 0.067298ωᵢ^3 - 0.017366ωᵢ^4)
Zcᵢ = 0.289 - 0.0701ωᵢ - 0.0207ωᵢ^2
βᵢ  = -10.2447 - 28.6312ωᵢ

References

Magoulas, K., & Tassios, D. (1990). Thermophysical properties of n-Alkanes from C1 to C20 and their prediction for higher ones. Fluid Phase Equilibria, 56, 119–140. doi:10.1016/0378-3812(90)85098-u

Mixing Rule Models

Clapeyron.ab_premixing — Function

ab_premixing(::Type{T},mixing,Tc,pc,kij) where T <: ABCubicModel

given Tc::SingleParam, pc::SingleParam, kij::PairParam and mixing <: MixingRule, it will return PairParams a and b, containing values aᵢⱼ and bᵢⱼ. by default, it performs the van der Wals One-Fluid mixing rule. that is:

aᵢⱼ = sqrt(aᵢ*aⱼ)*(1-kᵢⱼ)
bᵢⱼ = (bᵢ + bⱼ)/2

Clapeyron.mixing_rule — Function

mixing_rule(model::CubicModel,V,T,z,mixing_model::MixingRule,α,a,b,c)

Interface function used by cubic models. with matrices a and b, vectors α and c, a model::CubicModel and mixing_model::MixingRule, returns the scalars ā,b̄ and c̄, corresponding to the values mixed by the amount of components and the specifics of the mixing rule.

Example

function mixing_rule(model::CubicModel,V,T,z,mixing_model::vdW1fRule,α,a,b,c)
    ∑z = sum(z)
    ā = dot(z .* sqrt(α),a,z .* sqrt(α))/(∑z*∑z) #∑∑aᵢⱼxᵢxⱼ√(αᵢαⱼ)
    b̄ = dot(z,b,z)/(∑z*∑z)  #∑∑bᵢⱼxᵢxⱼ
    c̄ = dot(z,c)/∑z ∑cᵢxᵢ
    return ā,b̄,c̄
end

Clapeyron.vdW1fRule — Type

vdW1fRule <: vdW1fRuleModel

vdW1fRule(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

None

Description

van der Wals One-Fluid mixing rule for cubic parameters:

aᵢⱼ = √(aᵢaⱼ)(1-kᵢⱼ)
bᵢⱼ = (bᵢ + bⱼ)/2
ā = ∑aᵢⱼxᵢxⱼ√(αᵢ(T)αⱼ(T))
b̄ = ∑bᵢⱼxᵢxⱼ
c̄ = ∑cᵢxᵢ

Clapeyron.KayRule — Type

KayRule <: KayRuleModel

KayRule(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

None

Description

Kay mixing rule for cubic parameters:

aᵢⱼ = √(aᵢaⱼ)(1-kᵢⱼ)
bᵢⱼ = (bᵢ + bⱼ)/2
ā = b̄*(∑[aᵢⱼxᵢxⱼ√(αᵢ(T)αⱼ(T))/bᵢⱼ])^2
b̄ = (∑∛(bᵢⱼ)xᵢxⱼ)^3
c̄ = ∑cᵢxᵢ

Clapeyron.HVRule — Type

HVRule{γ} <: HVRuleModel

HVRule(components::Vector{String};
activity = Wilson,
userlocations::Vector{String}=String[],
activity_userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

None

Input models

activity: Activity Model

Description

Huron-Vidal Mixing Rule

aᵢⱼ = √(aᵢaⱼ)(1-kᵢⱼ)
bᵢⱼ = (bᵢ +bⱼ)/2
b̄ = ∑bᵢⱼxᵢxⱼ
c̄ = ∑cᵢxᵢ
ā = b̄(∑[xᵢaᵢᵢαᵢ/(bᵢᵢ)] - gᴱ/λ)
for Redlich-Kwong:
    λ = log(2) (0.6931471805599453)
for Peng-Robinson:
    λ = 1/(2√(2))log((2+√(2))/(2-√(2))) (0.6232252401402305)

λ is a coefficient indicating the relation between gᴱ and gᴱ(cubic) at infinite pressure. see [1] for more information. it can be customized by defining HV_λ(::HVRuleModel,::CubicModel)

References

Huron, M.-J., & Vidal, J. (1979). New mixing rules in simple equations of state for representing vapour-liquid equilibria of strongly non-ideal mixtures. Fluid Phase Equilibria, 3(4), 255–271. doi:10.1016/0378-3812(79)80001-1

Clapeyron.MHV1Rule — Type

MHV1Rule{γ} <: MHV1RuleModel

MHV1Rule(components::Vector{String};
activity = Wilson,
userlocations::Vector{String}=String[],
activity_userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

None

Input models

activity: Activity Model

Description

Modified Huron-Vidal Mixing Rule, First Order

aᵢⱼ = √(aᵢaⱼ)(1-kᵢⱼ)
bᵢⱼ = (bᵢ + bⱼ)/2
b̄ = ∑bᵢⱼxᵢxⱼ
c̄ = ∑cᵢxᵢ
ā = b̄RT(∑[xᵢaᵢᵢαᵢ/(RTbᵢᵢ)] - [gᴱ/RT + ∑log(bᵢᵢ/b̄)]/q)

if the model is Peng-Robinson:
    q = 0.53
if the model is Redlich-Kwong:
    q = 0.593

to use different values for q, overload Clapeyron.MHV1q(::CubicModel,::MHV1Model) = q

References

Michelsen, M. L. (1990). A modified Huron-Vidal mixing rule for cubic equations of state. Fluid Phase Equilibria, 60(1–2), 213–219. doi:10.1016/0378-3812(90)85053-d

Clapeyron.MHV2Rule — Type

MHV2Rule{γ} <: MHV2RuleModel

MHV2Rule(components::Vector{String};
activity = Wilson,
userlocations::Vector{String}=String[],
activity_userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

None

Input models

activity: Activity Model

Description

Modified Huron-Vidal Mixing Rule, Second Order.

aᵢⱼ = √(aᵢaⱼ)(1 - kᵢⱼ)
bᵢⱼ = (bᵢ + bⱼ)/2
b̄ = ∑bᵢⱼxᵢxⱼ
c̄ = ∑cᵢxᵢ
ᾱᵢ  = aᵢαᵢ/bᵢRT
ċ = -q₁*Σᾱᵢxᵢ - q₂*Σᾱᵢxᵢ^2 - gᴱ/RT - ∑log(bᵢᵢ/b̄)
ā = (-q₁ - √(q₁^2 - 4q₂ċ))/(2q₂)

if the model is Peng-Robinson:
    q₁ = -0.4347, q₂ = -0.003654
if the model is Redlich-Kwong:
    q₁ = -0.4783, q₂ = -0.0047
    (-0.4783,-0.0047)

to use different values for q₁ and q₂, overload Clapeyron.MHV1q(::CubicModel,::MHV2Model) = (q₁,q₂)

References

Michelsen, M. L. (1990). A modified Huron-Vidal mixing rule for cubic equations of state. Fluid Phase Equilibria, 60(1–2), 213–219. doi:10.1016/0378-3812(90)85053-d

Clapeyron.LCVMRule — Type

LCVMRule{γ} <: LCVMRuleModel

LCVMRule(components::Vector{String};
activity = Wilson,
userlocations::Vector{String}=String[],
activity_userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

None

Input models

activity: Activity Model

Description

Linear Combination of Vidal and Michaelsen (LCVM) Mixing Rule

aᵢⱼ = √(aᵢaⱼ)(1 - kᵢⱼ)
bᵢⱼ = (bᵢ + bⱼ)/2
b̄ = ∑bᵢⱼxᵢxⱼ
c̄ = ∑cᵢxᵢ
ᾱᵢ  = aᵢαᵢ/bᵢRT
ċ = -q₁*Σᾱᵢxᵢ - q₂*Σᾱᵢxᵢ^2 - gᴱ/RT - ∑log(bᵢᵢ/b̄)
ā = b̄RT(-1.827[gᴱ/RT - 0.3∑log(bᵢᵢ/b̄)] + Σᾱᵢxᵢ)

References

Boukouvalas, C., Spiliotis, N., Coutsikos, P., Tzouvaras, N., & Tassios, D. (1994). Prediction of vapor-liquid equilibrium with the LCVM model: a linear combination of the Vidal and Michelsen mixing rules coupled with the original UNIFAC. Fluid Phase Equilibria, 92, 75–106. doi:10.1016/0378-3812(94)80043-x

Clapeyron.WSRule — Type

WSRule{γ} <: WSRuleModel

WSRule(components::Vector{String};
activity = Wilson,
userlocations::Vector{String}=String[],
activity_userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

None

Input models

activity: Activity Model

Description

Wong-Sandler Mixing Rule.

aᵢⱼ = √(aᵢaⱼ)(1 - kᵢⱼ)
bᵢⱼ = (bᵢ + bⱼ)/2
c̄ = ∑cᵢxᵢ
B̄ = Σxᵢxⱼ(bᵢⱼ - aᵢⱼ√(αᵢαⱼ)/RT)
b̄  = B̄/(1 - gᴱ/λRT - Σxᵢaᵢαᵢ/bᵢRT)
ā = RT(b̄ - B̄)
for Redlich-Kwong:
    λ = log(2) (0.6931471805599453)
for Peng-Robinson:
    λ = 1/(2√(2))log((2+√(2))/(2-√(2))) (0.6232252401402305)

λ is a coefficient indicating the relation between gᴱ and gᴱ(cubic) at infinite pressure. see [1] for more information. it can be customized by defining WS_λ(::WSRuleModel,::CubicModel)

References

Wong, D. S. H., & Sandler, S. I. (1992). A theoretically correct mixing rule for cubic equations of state. AIChE journal. American Institute of Chemical Engineers, 38(5), 671–680. doi:10.1002/aic.690380505

Orbey, H., & Sandler, S. I. (1995). Reformulation of Wong-Sandler mixing rule for cubic equations of state. AIChE journal. American Institute of Chemical Engineers, 41(3), 683–690. doi:10.1002/aic.690410325

Clapeyron.VTPRRule — Type

VTPRRule{γ} <: VTPRRuleModel

VTPRRule(components::Vector{String};
activity = UNIFAC,
userlocations::Vector{String}=String[],
activity_userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

None

Input models

activity: Activity Model

Description

Mixing Rule used by the Volume-translated Peng-Robinson (VTPR) equation of state. only works with activity models that define an lnγ_res function (UNIFAC models)

aᵢⱼ = √(aᵢaⱼ)(1-kᵢⱼ)
bᵢⱼ = ((bᵢ^(3/4) + bⱼ^(3/4))/2)^(4/3)
log(γʳ)ᵢ = lnγ_res(model.activity,V,T,z) 
gᴱᵣₑₛ = ∑RTlog(γʳ)ᵢxᵢ
b̄ = ∑bᵢⱼxᵢxⱼ
c̄ = ∑cᵢxᵢ
ā = b̄RT(∑[xᵢaᵢᵢαᵢ/(RTbᵢᵢ)] - gᴱᵣₑₛ/(0.53087RT))

References

Ahlers, J., & Gmehling, J. (2001). Development of an universal group contribution equation of state. Fluid Phase Equilibria, 191(1–2), 177–188. doi:10.1016/s0378-3812(01)00626-4

Clapeyron.PSRKRule — Type

PSRKRule{γ} <: MHV1RuleModel

PSRKRule(components::Vector{String};
activity = Wilson,
userlocations::Vector{String}=String[],
activity_userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

None

Input models

activity: Activity Model

Description

Mixing Rule used by the Predictive Soave-Redlich-Kwong (PSRK) equation of state, derived from the First Order modified Huron-Vidal Mixing Rule.