Index

Clapeyron.BasicIdeal

Clapeyron.IAPWS95Ideal

Clapeyron.JobackIdeal

Clapeyron.LJRefIdeal

Clapeyron.MonomerIdeal

Clapeyron.ReidIdeal

Clapeyron.WalkerIdeal

Clapeyron.idealmodel

Ideal Models

All Clapeyron.jl models can be separated between an ideal and a residual contribution. The ideal contribution can be obtained via integration of the ideal isobaric heat capacity:

$\frac{A_{\mathrm{ideal}}}{Nk_\mathrm{B}T} = \sum_{i=1}^{N_{\mathrm{Component}}} x_i\left[\ln{\frac{\rho_i}{\rho_0}} + \frac{1}{Nk_\mathrm{B}T} \int_{T_0}^T \!\!C_{p,i}^0 dT + \frac{H_{0,i}}{Nk_\mathrm{B}T}- \frac{1}{Nk_{B}}\!\!\int_{T_0}^T \frac{C_{p,i}^0}{T} dT -\ln{\frac{T}{T_0}}-\frac{S_{0,i}}{Nk_\mathrm{B}} - 1\right]$

Clapeyron.idealmodel — Function

idealmodel(model::EoSModel)

retrieves the ideal model from the input's model.

Examples:

julia> pr = PR(["water"],idealmodel=IAPWS95Ideal)   
PR{IAPWS95Ideal} with 1 component:
 "water"
Contains parameters: a, b, acentricfactor, Tc, Mw   
julia> Clapeyron.idealmodel(pr)
IAPWS95Ideal()

Clapeyron.BasicIdeal — Type

BasicIdeal <: IdealModel
BasicIdeal(components::Array{String,1}; 
userlocations::Array{String,1}=String[], 
verbose=false)

Input parameters

None

Description

Default Ideal Model. Constant specific heat capacity equal to 5R/2. it's Helmholtz energy is equal to:

    a₀ = A₀/nRT =  ∑(xᵢlog(nxᵢ/V)) - 1 - 1.5log(T)

Clapeyron.ReidIdeal — Type

ReidIdeal <: IdealModel
ReidIdeal(components; 
userlocations::Array{String,1}=String[], 
verbose=false)

Input parameters

a: Single Parameter (Float64)

b: Single Parameter (Float64)

c: Single Parameter (Float64)

d: Single Parameter (Float64)

Model parameters

coeffs: Single Parameter (NTuple{4,Float64})

Description

Reid Ideal Model. Helmholtz energy obtained via integration of specific heat capacity:

Cpᵢ(T) = aᵢ  + bᵢT + cᵢT^2 + dᵢT^3
Cp(T) = ∑Cpᵢxᵢ

Clapeyron.JobackIdeal — Type

JobackIdeal <: JobackIdealModel
JobackIdeal(components; 
userlocations::Array{String,1}=String[], 
verbose=false)

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

N_a: Single Parameter (Float64)

T_c: Single Parameter (Float64)

P_c: Single Parameter (Float64)

V_c: Single Parameter (Float64)

T_b: Single Parameter (Float64)

T_m: Single Parameter (Float64)

H_form: Single Parameter (Float64)

G_form: Single Parameter (Float64)

a: Single Parameter (Float64)

b: Single Parameter (Float64)

c: Single Parameter (Float64)

d: Single Parameter (Float64)

H_fusion: Single Parameter (Float64)

H_vap: Single Parameter (Float64)

eta_a: Single Parameter (Float64)

eta_b: Single Parameter (Float64)

Description

Joback Group Contribution Ideal Model. GC version of ReidIdeal. Helmholtz energy obtained via integration of specific heat capacity:

aᵢ = ∑(νᵢₖbₖ) - 37.93
bᵢ = ∑(νᵢₖbₖ) + 0.210
cᵢ = ∑(νᵢₖcₖ) - 3.91e-4
dᵢ = ∑(νᵢₖbₖ) + 2.06e-7
Cpᵢ(T) = aᵢ  + bᵢT + cᵢT^2 + dᵢT^3

The GC-averaged Reid Model is available by using ReidIdeal(model::JobackIdeal).

The estimated critical point of a single component can be obtained via crit_pure(model::JobackIdeal)

References

Joback, K. G., & Reid, R. C. (1987). Estimation of pure-component properties from group-contributions. Chemical Engineering Communications, 57(1–6), 233–243. doi:10.1080/00986448708960487

Clapeyron.MonomerIdeal — Type

MonomerIdeal <: MonomerIdealModel
MonomerIdeal(components::Array{String,1}; 
userlocations::Array{String,1}=String[], 
verbose=false)

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

Model Parameters

None

Description

Monomer Ideal Model, result obtained from statistical mechanics Λ

    Λᵢ = h/√(kᵦTMwᵢ/Nₐ)    
    a₀ = A₀/nRT = ∑xᵢlog(ρᵢΛᵢ^3)

Clapeyron.WalkerIdeal — Type

WalkerIdeal <: WalkerIdealModel
WalkerIdeal(components::Array{String,1}; 
userlocations::Array{String,1}=String[], 
verbose=false)

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

Nrot: Single Parameter (Int)

theta1: Single Parameter (Float64)

theta2: Single Parameter (Float64)

theta3: Single Parameter (Float64)

theta4: Single Parameter (Float64)

deg1: Single Parameter (Int)

deg2: Single Parameter (Int)

deg3: Single Parameter (Int)

deg4: Single Parameter (Int)

Description

Walker [1] Group Contribution Ideal Model.

Cpᵢ(T)/R = (5+NRot)/2 ∑νᵢₖ∑gₖᵥ(θₖᵥ/T)^2*exp(θₖᵥ/T)/(1-exp(θₖᵥ/T)) , v ∈ 1:4 

References

Walker, P. J., & Haslam, A. J. (2020). A new predictive group-contribution ideal-heat-capacity model and its influence on second-derivative properties calculated using a free-energy equation of state. Journal of Chemical and Engineering Data, 65(12), 5809–5829. doi:10.1021/acs.jced.0c00723

Clapeyron.IAPWS95Ideal — Type

IAPWS95Ideal <: IdealModel
IAPWS95Ideal(components; 
userlocations::Array{String,1}=String[], 
verbose=false)

IAPWS95Ideal()

Input parameters

None

Description

IAPWS95 ideal helmholtz model for use in other models. Only valid for water. Check IAPWS95 for more information.

References

Wagner, W., & Pruß, A. (2002). The IAPWS formulation 1995 for the thermodynamic properties of ordinary water substance for general and scientific use. Journal of physical and chemical reference data, 31(2), 387–535. doi:10.1063/1.1461829

IAPWS R6-95 (2018). Revised Release on the IAPWS Formulation 1995 for the Thermodynamic Properties of Ordinary Water Substance for General and Scientific Use

Clapeyron.LJRefIdeal — Type

LJRefIdeal <: IdealModel
LJRef(components;
userlocations=String[],
verbose=false)

Input parameters

sigma: Single Parameter (Float64) - particle size [Å]

epsilon: Single Parameter (Float64) - dispersion energy [K]

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

Description

Leonard-Jones Reference equation of state. Ideal Part. valid from 0.5 < T/Tc < 7 and pressures up to p/pc = 500.

``` τᵢ = 1.32ϵᵢ/T δᵢ = n(Nₐσᵢ^3)/0.31V a⁰ᵢ(δ,τ) = log(δᵢ) + 1.5log(τᵢ) + 1.515151515τᵢ + 6.262265814 a⁰(δ,τ,z) = ∑xᵢ(a⁰ᵢ + log(xᵢ))

````

LJRefIdeal acts as a wrapper of LJRef model, you can access it with LJRef(model::LJRefIdeal).

Mutiple component warning

The original model was done with only one component in mind. to support multiple components, a VDW 1-fluid mixing rule (shown above) is implemented, but it is not tested.

References

Thol, M., Rutkai, G., Köster, A., Lustig, R., Span, R., & Vrabec, J. (2016). Equation of state for the Lennard-Jones fluid. Journal of physical and chemical reference data, 45(2), 023101. doi:10.1063/1.4945000