Index

Clapeyron.SLKRule

Clapeyron.SLk0k1lMixingRule

Clapeyron.SanchezLacombe

Clapeyron.mix_vε

Sanchez-Lacombe Model

Clapeyron.SanchezLacombe — Type

SanchezLacombe(components::Vector{String}; 
idealmodel=BasicIdeal, 
mixing = SLk0k1lMixingRule, 
userlocations=String[], 
ideal_userlocations=String[], 
mixing_userlocations = String[],
verbose=false)

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

epsilon: Single Parameter (Float64) - Nonbonded interaction energy per monomer [J/mol]

vol: Single Parameter (Float64) - Closed Packed Specific volume [m^3/mol]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

epsilon: Pair Parameter (Float64) - Nonbonded interaction energy per monomer [J/mol]

vol: Pair Parameter (Float64) - Closed Packed Specific volume [m^3/mol]

Input models

idealmodel: Ideal Model

mixing: Mixing model

Description

Sanchez-Lacombe Lattice Fluid Equation of State.

xᵢ = zᵢ/∑zᵢ
r̄ = ∑xᵢrᵢ
vᵣ,εᵣ = mix_vε(model,V,T,z,model.mixing,r̄,∑zᵢ)
ρ̃ = r̄*vᵣ/v
T̃ = R̄*T/εᵣ
aᵣ = r̄*(- ρ̃ /T̃ + (1/ρ̃  - 1)*log(1 - ρ̃ ) + 1)

References

Neau, E. (2002). A consistent method for phase equilibrium calculation using the Sanchez–Lacombe lattice–fluid equation-of-state. Fluid Phase Equilibria, 203(1–2), 133–140. doi:10.1016/s0378-3812(02)00176-0

Clapeyron.mix_vε — Function

mix_vε(model::SanchezLacombeModel,V,T,z,mix::SLMixingRule,r̄ = @f(rmix),∑z = sum(z))

Function used to dispatch on the different mixing rules available for Sanchez-Lacombe.

Example:

function mix_vε(model::SanchezLacombe,V,T,z,mix::SLKRule,r̄,Σz = sum(z))
    v = model.params.vol.values
    ε = model.params.epsilon.values
    r =  model.params.segment.values
    k = mix.k.values
    x = z ./ Σz
    ϕ = @. r * x / r̄
    εᵣ = sum(ε[i,j]*(1-k[i,j])*ϕ[i]*ϕ[j] for i ∈ @comps for j ∈ @comps)
    vᵣ = sum(v[i,j]*ϕ[i]*ϕ[j] for i ∈ @comps for j ∈ @comps)
    return vᵣ,εᵣ

Clapeyron.SLKRule — Type

SLKRule(components; userlocations=String[], verbose=false)

Input parameters

k: Pair Parameter (Float64) - Binary Interaction Parameter (no units)

Constant Kᵢⱼ mixing rule for Sanchez-Lacombe:

εᵢⱼ = √εᵢεⱼ*(1-kᵢⱼ)
vᵢⱼ = (vᵢ + vⱼ)/2
ϕᵢ = rᵢ*xᵢ/r̄
εᵣ = ΣΣϕᵢϕⱼεᵢⱼ
vᵣ = ΣΣϕᵢϕⱼvᵢⱼ

Clapeyron.SLk0k1lMixingRule — Type

SLKRule(components; userlocations=String[], verbose=false)

Input parameters

k0: Pair Parameter (Float64) - Binary Interaction Parameter (no units)

k1: Pair Parameter (Float64) - Binary Interaction Parameter (no units)

l: Pair Parameter (Float64) - Binary Interaction Parameter (no units)

Neau's Consistent k₀,k₁,l mixing rule for Sanchez-Lacombe:

εᵢⱼ = √εᵢεⱼ
vᵢⱼ = (1 - lᵢⱼ)(vᵢ + vⱼ)/2
ϕᵢ = rᵢ*xᵢ/r̄
εᵣ = ΣΣϕᵢϕⱼεᵢⱼ*(1 - k₀ᵢⱼ + (1 - δᵢⱼ)(Σϕₖk₁ᵢₖ + Σϕₖk₁ₖⱼ))
vᵣ = ΣΣϕᵢϕⱼvᵢⱼ

Where δᵢⱼ is i == j ? 1 : 0

References

Neau, E. (2002). A consistent method for phase equilibrium calculation using the Sanchez–Lacombe lattice–fluid equation-of-state. Fluid Phase Equilibria, 203(1–2), 133–140. doi:10.1016/s0378-3812(02)00176-0

Extended Corresponding states Model