Contents

Contents

Properties

Helper functions

Properties

Clapeyron.mean_ionic_activity_coefficient — Function

mean_ionic_activity_coefficient(model::ESElectrolyteModel,salts,p,T,m,zsolvent=[1.])

Calculate the mean ionic activity coefficient of selection of salts at a given pressure, temperature and molality. These are defined as:

γ± = φ±/φ±₀ * ∑zsolv/∑z

Example:

model = ePCSAFT(["water"],["sodium","chloride"])

salts = [("sodium chloride",("sodium"=>1,"chloride"=>1))]

p = 1e5
T = 298.15
m = [1.0]

γ± = mean_ionic_activity_coefficient(model,salts,p,T,m)

If multiple solvents are present, the composition of the solvent can be specified with the zsolvent keyword argument.

Clapeyron.mean_ionic_activity_coefficient_sat — Function

mean_ionic_activity_coefficient_sat(model::ESElectrolyteModel,salts,T,m,zsolvent=[1.])

Calculate the mean ionic activity coefficient of selection of salts at the saturation point at a certain temperature and molality. These are defined as:

γ± = φ±/φ±₀ * ∑zsolv/∑z

Example:

model = ePCSAFT(["water"],["sodium","chloride"])

salts = [("sodium chloride",("sodium"=>1,"chloride"=>1))]

T = 298.15
m = [1.0]

γ± = mean_ionic_activity_coefficient_sat(model,salts,T,m)

If multiple solvents are present, the composition of the solvent can be specified with the zsolvent keyword argument.

Clapeyron.osmotic_coefficient — Function

osmotic_coefficient(model::ESElectrolyteModel,salts,p,T,m,zsolvent=[1.])

Calculate the osmotic coefficient of selection of solvents at a given pressure, temperature and molality. These are defined as:

ϕ = -1/(∑νi*mi*Mw)*log(asolv)

Example:

model = ePCSAFT(["water"],["sodium","chloride"])

salts = [("sodium chloride",("sodium"=>1,"chloride"=>1))]

p = 1e5
T = 298.15
m = [1.0]

ϕ = osmotic_coefficient(model,salts,p,T,m)

If multiple solvents are present, the composition of the solvent can be specified with the zsolvent keyword argument.

Clapeyron.osmotic_coefficient_sat — Function

osmotic_coefficient_sat(model::ESElectrolyteModel,salts,T,m,zsolvent=[1.])

Calculate the osmotic coefficient of selection of solvents at the saturation point at a certain temperature and molality. These are defined as:

ϕ = -1/(∑νi*mi*Mw)*log(asolv)

Example:

model = ePCSAFT(["water"],["sodium","chloride"])

salts = [("sodium chloride",("sodium"=>1,"chloride"=>1))]

T = 298.15
m = [1.0]

ϕ = osmotic_coefficient(model,salts,T,m)

If multiple solvents are present, the composition of the solvent can be specified with the zsolvent keyword argument.

Helper functions

Clapeyron.molality_to_composition — Function

molality_to_composition(model::ElectrolyteModel,salts,m,zsolv=[1.])

Convert molality (mol/kg) to composition for a given model, salts, molality, and solvent composition.