Contents

Index

Volume-Temperature Based Properties

Missing docstring.

Missing docstring for pip. Check Documenter's build log for details.

Pressure-Temperature Based Properties

Clapeyron.volume — Function

volume(model::EoSModel,p,T,z=SA[1.0];phase=:unknown,threaded=true)

calculates the volume (m³) of the compound modelled by model at a certain pressure,temperature and moles.

phase is a Symbol that determines the initial volume root to look for:

If phase =:liquid it will return the volume of the phase using a liquid initial point.

If phase =:vapor it will return the volume of the phase using a gas initial point.

The default is phase =:unknown. with this, both liquid and volume roots will be calculated, and the phase with the least amount of energy is returned.

The calculation of both volume roots can be calculated in serial (threaded=false) or in parallel (threaded=true)

Automatic Differenciation functions

Missing docstring.

Missing docstring for ∂f∂T. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ∂f∂V. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ∂f. Check Documenter's build log for details.

Missing docstring.

Missing docstring for p∂p∂V. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ∂2f. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ∂2p. Check Documenter's build log for details.

Missing docstring.

Missing docstring for f_hess. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ∂²³f. Check Documenter's build log for details.