Contents

Contents

Index

Volume-Temperature Based Properties

Pressure-Temperature Based Properties

Single component properties

Automatic Differenciation functions

Initial guess functions

Index

Clapeyron.T_scale

Clapeyron.acentric_factor

Clapeyron.compressibility_factor

Clapeyron.crit_pure

Clapeyron.enthalpy_vap

Clapeyron.f_hess

Clapeyron.lb_volume

Clapeyron.p_scale

Clapeyron.pip

Clapeyron.p∂p∂V

Clapeyron.sat_pure

Clapeyron.second_virial_coefficient

Clapeyron.volume

Clapeyron.volume_virial

Clapeyron.x0_crit_pure

Clapeyron.x0_sat_pure

Clapeyron.x0_volume

Clapeyron.x0_volume_gas

Clapeyron.x0_volume_liquid

Clapeyron.∂2f

Clapeyron.∂2p

Clapeyron.∂f

Clapeyron.∂f∂T

Clapeyron.∂f∂V

Clapeyron.∂²³f

Volume-Temperature Based Properties

Clapeyron.second_virial_coefficient — Function

second_virial_coefficient(model::EoSModel, T, z=SA[1.])

Calculates the second virial coefficient B, defined as:

B = lim(V->∞)[ V^2/RT *  (∂Aᵣ∂V + V*∂²Aᵣ∂V²) ]

where Aᵣ is the residual helmholtz energy.

Clapeyron.pip — Function

pip(model::EoSModel,V,T,z=[1.0])

Phase identification parameter Π. as described in 1. If Π > 1, then the phase is clasified as a liquid or a liquid-like vapor, being a vapor or vapor-like liquid otherwise.

This identification parameter fails at temperatures and pressures well aboVe the critical point.

Calculated as:

Π = V*((∂²p/∂V∂T)/(∂p/∂T) - (∂²p/∂V²)/(∂p/∂V))

G. Venkatarathnama, L.R. Oellrich, Identification of the phase of a fluid using partial derivatives of pressure, volume,and temperature without reference to saturation properties: Applications in phase equilibria calculations, Fluid Phase Equilibria 301 (2011) 225–233

Pressure-Temperature Based Properties

Clapeyron.volume — Function

volume(model::EoSModel,p,T,z=SA[1.0];phase=:unknown,threaded=true)

calculates the volume (m³) of the compound modelled by model at a certain pressure,temperature and moles.

phase is a Symbol that determines the initial volume root to look for:

If phase =:liquid it will return the volume of the phase using a liquid initial point.

If phase =:vapor it will return the volume of the phase using a gas initial point.

The default is phase =:unknown. with this, both liquid and volume roots will be calculated, and the phase with the least amount of energy is returned.

The calculation of both volume roots can be calculated in serial (threaded=false) or in parallel (threaded=true)

Clapeyron.compressibility_factor — Function

compressibility_factor(model::EoSModel, p, T, z=SA[1.]; phase = :unknown,threaded=true)

Calculates the compressibility factor Z, defined as:

Z = p*V(p)/R*T

the keywords phase and threaded are passed to the volume solver.

Single component properties

Clapeyron.sat_pure — Function

sat_pure(model::EoSModel, T, V0 = x0_sat_pure(model,T))

Performs a single component saturation equilibrium calculation, at the specified temperature T, of one mol of pure sustance specified by model

Returns (p₀, Vₗ, Vᵥ) where p₀ is the saturation pressure (in Pa), Vₗ is the liquid saturation volume (in m³) and Vᵥ is the vapour saturation volume (in m³).

If the calculation fails, returns (NaN, NaN, NaN)

V0 is [log10(Vₗ₀),log10(Vᵥ₀)] , where Vₗ₀ and Vᵥ₀ are initial guesses for the liquid and vapour volumes.

Clapeyron.enthalpy_vap — Function

enthalpy_vap(model::EoSModel, T)

Calculates ΔH, the difference between saturated vapour and liquid enthalpies at temperature T, in J

Clapeyron.crit_pure — Function

crit_pure(model::EoSModel,x0=nothing)

Calculates the critical point of a single component modelled by model.

Returns (Tc, pc, Vc) where Tc is the critical temperature (in K), pc is the critical pressure (in Pa) and Vc is the critical volume (in m³)

Clapeyron.acentric_factor — Function

acentric_factor(model::EoSModel)

calculates the acentric factor using its definition:

ω = -log10(psatᵣ) -1, at Tᵣ = 0.7

To do so, it calculates the critical temperature (using crit_pure) and performs a saturation calculation (with sat_pure)

Automatic Differenciation functions

Clapeyron.∂f∂T — Function

∂f∂T(model,V,T,z=SA[1.0])

returns f and ∂f/∂T at constant total volume and composition, where f is the total helmholtz energy, given by eos(model,V,T,z)

Clapeyron.∂f∂V — Function

∂f∂V(model,V,T,z=SA[1.0])

returns f and ∂f/∂V at constant temperature and composition, where f is the total helmholtz energy, given by eos(model,V,T,z), and V is the total volume

Clapeyron.∂f — Function

∂f(model,V,T,z)

returns zeroth order (value) and first order derivative information of the total helmholtz energy (given by eos(model,V,T,z)). the result is given in two values:

grad_f,fval = ∂2f(model,V,T,z)

where:

fval   = f(V,T) = eos(model,V,T,z)

grad_f = [ ∂f/∂V; ∂f/∂T]

Where V is the total volume, T is the temperature and f is the total helmholtz energy.

Clapeyron.p∂p∂V — Function

p∂p∂V(model,V,T,z=SA[1.0])

returns p and ∂p/∂V at constant temperature, where p is the pressure = pressure(model,V,T,z) and V is the total Volume.

Clapeyron.∂2f — Function

∂2f(model,V,T,z)

returns zeroth order (value), first order and second order derivative information of the total helmholtz energy (given by eos(model,V,T,z)). the result is given in three values:

hess_f,grad_f,fval = ∂2f(model,V,T,z)

where: ``` fval = f(V,T) = eos(model,V,T,z)

grad_f = [ ∂f/∂V; ∂f/∂T]

hess_f = [ ∂²f/∂V²; ∂²f/∂V∂T ∂²f/∂V∂T; ∂²f/∂V²] ```

Where V is the total volume, T is the temperature and f is the total helmholtz energy.

Clapeyron.∂2p — Function

∂2p(model,V,T,z)

returns zeroth order (value), first order and second order derivative information of the pressure. the result is given in three values:

hess_p,grad_p,pval = ∂2p(model,V,T,z)

where: ``` pval = p(V,T) = pressure(model,V,T,z)

grad_p = [ ∂p/∂V; ∂p/∂T]

hess_p = [ ∂²p/∂V²; ∂²p/∂V∂T ∂²p/∂V∂T; ∂²p/∂V²] ```

Where V is the total volume, T is the temperature and p is the pressure.

Clapeyron.f_hess — Function

f_hess(model,V,T,z)

returns the second order volume (V) and temperature (T) derivatives of the total helmholtz energy (given by eos(model,V,T,z)). the result is given in a 2x2 SMatrix, in the form:

[ ∂²f/∂V² ∂²f/∂V∂T ∂²f/∂V∂T ∂²f/∂V²]

use this instead of the ∂2f if you only need second order information. ∂2f also gives zeroth and first order derivative information, but due to a bug in the used AD, it allocates more than necessary.

Clapeyron.∂²³f — Function

∂²³f(model,V,T,z=SA[1.0])

returns ∂²A/∂V² and ∂³A/∂V³, in a single ForwardDiff pass. used mainly in crit_pure objective function

Initial guess functions

Clapeyron.lb_volume — Function

lb_volume(model::EoSModel,z=SA[1.0])

Returns the lower bound volume.

It has different meanings depending on the Equation of State, but symbolizes the minimum allowable volume at a certain composition:

SAFT EoS: the packing volume

Cubic EoS, covolume (b) parameter

On empiric equations of state, the value is chosen to match the volume of the conditions at maximum pressure and minimum temperature , but the equation itself normally can be evaluated at lower volumes.

On SAFT and Cubic EoS, volumes lower than lb_volume will likely error.

The lower bound volume is used for guesses of liquid volumes at a certain pressure, saturated liquid volumes and critical volumes.

Clapeyron.T_scale — Function

T_scale(model::EoS,z=SA[1.0])

Represents a temperature scaling factor.

On any EoS based on Critical parameters (Cubic or Empiric EoS), the temperature scaling factor is chosen to be the critical temperature.

On SAFT or other molecular EoS, the temperature scaling factor is chosen to be a function of the potential depth ϵ.

Used as scaling factors in sat_pure and as input for solving crit_pure

Clapeyron.p_scale — Function

p_scale(model::SAFTModel,z=SA[1.0])

Represents a pressure scaling factor

On any EoS based on Critical parameters (Cubic or Empiric EoS), the pressure scaling factor is chosen to be a function of the critical pressure.

On SAFT or other molecular EoS, the temperature scaling factor is chosen to a function of ∑(zᵢϵᵢ(σᵢᵢ)³)

Used as scaling factors in sat_pure and as input for solving crit_pure

Clapeyron.x0_volume — Function

x0_volume(model::EoSModel,p,T,z; phase = :unknown)

Returns an initial guess of the volume at a pressure, temperature, composition and suggested phase.

If the suggested phase is :unkwown or :liquid, calls x0_volume_liquid.

If the suggested phase is :gas, calls x0_volume_gas.

Clapeyron.x0_volume_liquid — Function

x0_volume_liquid(model,T,z)

Returns an initial guess to the liquid volume, dependent on temperature and composition. by default is 1.25 times the lower bound volume.

Clapeyron.x0_volume_gas — Function

x0_volume_gas(model,p,T,z)

Returns an initial guess to the gas volume, depending of pressure, temperature and composition. by default uses a virial aproximation

Clapeyron.volume_virial — Function

volume_virial(model,p,T,z=SA[1.0])

Calculates an aproximation to the gas volume at specified pressure, volume and composition, by aproximating:

Z(v) ≈ 1 + B(T)/v 

where Z is the compressibility factor and B is the second virial coefficient. if B>0, (over the inversion temperature) returns NaN. If the solution to the problem is complex (Z = 1 + B/v implies solving a quadratic polynomial), returns -2*B.

Clapeyron.x0_sat_pure — Function

x0_sat_pure(model::EoSModel,T,z=SA[1.0])

Returns a 2-element vector corresponding to [log10(Vₗ),log10(Vᵥ)], where Vₗ and Vᵥ are the liquid and vapor initial guesses. Used in sat_pure.

Clapeyron.x0_crit_pure — Function

x0_crit_pure(model::SAFTModel,z=SA[1.0])

Returns a 2-element vector corresponding to [k,log10(Vc0)], where k is Tc0/T_scale(model,z)