Contents

Contents

Index

Estimation Tools

Index

Clapeyron.Estimation

Clapeyron.EstimationData

Clapeyron.ToEstimate

Clapeyron.objective_function

Clapeyron.return_model

Estimation Tools

Clapeyron.ToEstimate — Type

ToEstimate
ToEstimate(params_dict)

Input parameters: A dictionary with the following potential entries

params: The name of the parameter being fitted (Symbol)

indices: The index of the parameter being fitted (Integer or Tuple{Integer,Integer})

factor: Factor to multiply parameter being fitted to have it in the correct units (Float64)

symmetric: For PairParam, if the parameter is symmetric or asymmetric (Bool)

cross_assoc: For AssocParam, if the parameter is for cross-association (Bool)

recombine: For PairParam, if the combining rules must be applied for unlike interactions (Bool)

lower: Lower bound for the parameter (Float64)

upper: Upper bound for the parameter (Float64)

guess: Initial guess for the parameter (Float64)

Output:

A ToEstimate struct

Description

Turns the input parameter dictionary into a ToEstimate struct to be used within the parameter estimation.

Clapeyron.Estimation — Type

Estimation
Estimation(model,toestimate,filepaths,ignorefield,objective_form)

Input parameters:

model: The initial model containing the species we wish to parameterise

toestimate: The dictionary of parameters being fitted

filepaths or filepaths_weights: The location of the data files used to fit. Can also contain the weights of each dataset

ignorefield: Specify which EoSModel fields to ignore in the main model

objective_form: Specify the functional form of the objective function in the form objective_form(pred,exp)

Output:

Estimator object which contains the following:

model: The model whose parameters will be varied

initial_model: The initial model before parameterisation

toestimate: ToEstimate struct which contains all the information on the parameters

data: Vector of EstimationData structs where all the information on the data is stored

ignorefield: Vector of fields to ignore in the parameter estimation

objective_form: Function to evaluate the error measure for the objective function

The following objects are also output:

objective: The objective function which is used to fit the parameters

x0: Initial guesses for the parameters

upper: Upper bounds for the parameters

lower: Lower bounds for the parameters

Description

Produces the estimator and other useful objects used within parameter estimation

Clapeyron.EstimationData — Type

EstimationData
EstimationData(filepaths)

Input parameters:

filepaths or filepaths_weights: The filepath of the data used in parameter estimation. Optionally, a tuple containing the weights of each dataset.

Output:

An EstimationData struct with the following fields:

method: The property estimation method which is used to obtain predictions for a given input

inputs_name: The variable names for the inputs

outputs_name: The variable names for the outputs

inputs: Vector for each input

outputs: Vector for each output

weights: The weight for this particular dataset

Description

For a given input data set, produce an EstimationData struct.

Clapeyron.return_model — Function

return_model
return_model(estimation,model,params)

Input parameters:

estimation: The estimator object

model: The model whose parameters we are varying

params: The new parameters which we want to change

Output:

model: The new model with the updated parameters

Description

Based on the parameters provided and the estimator, a new model is produced from the input.

Clapeyron.objective_function — Function

objective_function
objective_function(estimation,params)

Input parameters:

estimation: The estimator object

params: The new parameters which we want to evaluate the objective function for

Output: The relate root mean square error given the data and parameters provided

Description

The objective function used within parameter estimation.