Contents

Contents

Index

Parsing Parameters from files

Parameter types

Index

Clapeyron.AssocParam

Clapeyron.GroupParam

Clapeyron.PairParam

Clapeyron.ParamOptions

Clapeyron.SingleParam

Clapeyron.SiteParam

Clapeyron.getparams

Parsing Parameters from files

Clapeyron.ParamOptions — Type

ParamOptions(;kwargs...)

Struct containing all the options related to parameter parsing:

userlocations::Vector{String} = String[]: List of used-defined locations to search.

usergrouplocations::Vector{String} = String[]: List of used-defined group locations to search.

asymmetricparams::Vector{String} = String[]: List of pair or association parameters that follow that param[i,j] ≠ param[j,i]

ignore_headers::Vector{String} = ["dipprnumber", "smiles"]: List of ignored headers.

ignore_missing_singleparams::Bool=false: Check if any single parameter has missing values. if set to false,getparams will fail when encountering a single parameter with missing values.

verbose::Bool = false: If true, show all operations done by getparams displayed in the terminal. this includes the warnings emmited by CSV.jl

species_columnreference::String ="species": column name to check for components. in pair and association params, it will check for #species#1 and #species#2, where #species# is the value of this option.

site_columnreference::String ="site": column name to check for sites in association params, it will check for #site#1 and #site#2, where #site# is the value of this option.

group_columnreference::String = "groups": column name to check for groups in group data.

normalisecomponents::Bool = true: If true, performs normalization of strings, on the CSV and input components

n_sites_columns::Dict{String,String} = Dict( "e" => "n_e","e1" => "n_e1","e2" => "n_e2","H" => "n_H"): dictionary to look number of sites. the number of sites is stored as columns in a single parameter csv file. for example, the number of sites of name e will be looked on the column n_e

return_sites::Bool = true: If set to false, association params will be ignored and sites will not be created, even if they exist in the list of locations.

Clapeyron.getparams — Function

params, sites = getparams(components,locations;kwargs...)

returns a Dict{String,ClapeyronParam} containing all the parameters found for the list of components in the available CSVs. locations are the locations relative to Clapeyron database. the available keywords are the ones used in ParamOptions

if return_sites is set to false, getparams will only return the found params.

Parameter types

Clapeyron.SingleParam — Type

SingleParam{T}

Struct designed to contain single parameters. Basically a vector with some extra info.

Creation:

julia> mw = SingleParam("molecular weight",["water","ammonia"],[18.01,17.03])
SingleParam{Float64}("molecular weight") with 2 components:
 "water" => 18.01
 "ammonia" => 17.03

julia> mw.values
2-element Vector{Float64}:
 18.01
 17.03

julia> mw.components
2-element Vector{String}:
 "water"
 "ammonia"

julia> mw2 = SingleParam(mw,"new name")
SingleParam{Float64}("new name") with 2 components:
 "water" => 18.01
 "ammonia" => 17.03

julia> has_oxigen = [true,false]; has_o = SingleParam(mw2,has_oxigen)
SingleParam{Bool}("new name") with 2 components:
 "water" => true
 "ammonia" => false

Example usage in models:

function molecular_weight(model,molar_frac)
    mw = model.params.mw.values
    res = zero(eltype(molarfrac))
    for i in @comps #iterating through all components
        res += molar_frac[i]*mw[i]
    end
    return res
end

Clapeyron.PairParam — Type

PairParam{T}

Struct designed to contain pair data. used a matrix as underlying data storage.

Creation:

julia> kij = PairParam("interaction params",["water","ammonia"],[0.1 0.0;0.1 0.0])
PairParam{Float64}["water", "ammonia"]) with values:
2×2 Matrix{Float64}:
 0.1  0.0
 0.1  0.0

julia> kij.values
2×2 Matrix{Float64}:
 0.1  0.0
 0.1  0.0

julia> kij.diagvalues
2-element view(::Vector{Float64}, 
1:3:4) with eltype Float64:
 0.1
 0.0

Example usage in models:

```julia #lets compute ∑xᵢxⱼkᵢⱼ function alpha(model,x) kij = model.params.kij.values ki = model.params.kij.diagvalues res = zero(eltype(molarfrac)) for i in @comps @show ki[i] for j in @comps res += x[i]x[j]kij[i,j] end end return res end

Clapeyron.AssocParam — Type

AssocParam{T}

Struct holding association parameters.

Clapeyron.GroupParam — Type

GroupParam

Struct holding group parameters.

Clapeyron.SiteParam — Type

SiteParam

Struct holding site parameters.