Contents

Contents

Index

Parsing Parameters from Files

Creating Files from Parameters

Parameter types

Combining Rules

Group Combining Rules

Model Splitting

Model Citing

Index

Clapeyron.AssocOptions

Clapeyron.AssocParam

Clapeyron.GroupParam

Clapeyron.PairParam

Clapeyron.ParamOptions

Clapeyron.SingleParam

Clapeyron.SiteParam

Clapeyron.ParamTable

Clapeyron.cite

Clapeyron.cleartemp!

Clapeyron.doi

Clapeyron.epsilon_HudsenMcCoubrey

Clapeyron.epsilon_LorentzBerthelot

Clapeyron.getparams

Clapeyron.group_matrix

Clapeyron.group_pairmean

Clapeyron.group_sum

Clapeyron.index_reduction

Clapeyron.is_splittable

Clapeyron.kij_mix

Clapeyron.lambda_LorentzBerthelot

Clapeyron.lambda_squarewell

Clapeyron.pair_mix

Clapeyron.sigma_LorentzBerthelot

Clapeyron.split_model

Parsing Parameters from Files

Clapeyron.ParamOptions — Type

ParamOptions(;kwargs...)

Struct containing all the options related to parameter parsing:

userlocations::Vector{String} = String[]: List of used-defined locations to search.

usergrouplocations::Vector{String} = String[]: List of used-defined group locations to search.

asymmetricparams::Vector{String} = String[]: List of pair or association parameters that follow that param[i,j] ≠ param[j,i]

ignore_headers::Vector{String} = ["dipprnumber", "smiles"]: List of ignored headers.

ignore_missing_singleparams::Vector{String} = String[]: List of parameters where checking for missing single parameter values are ignored.

verbose::Bool = false: If true, show all operations done by getparams displayed in the terminal. this includes the warnings emmited by CSV.jl

species_columnreference::String ="species": column name to check for components. in pair and association params, it will check for #species#1 and #species#2, where #species# is the value of this option.

site_columnreference::String ="site": column name to check for sites in association params, it will check for #site#1 and #site#2, where #site# is the value of this option.

group_columnreference::String = "groups": column name to check for groups in group data.

normalisecomponents::Bool = true: If true, performs normalization of strings, on the CSV and input components

n_sites_columns::Dict{String,String} = Dict( "e" => "n_e","e1" => "n_e1","e2" => "n_e2","H" => "n_H"): dictionary to look number of sites. the number of sites is stored as columns in a single parameter csv file. for example, the number of sites of name e will be looked on the column n_e

return_sites::Bool = true: If set to false, association params will be ignored and sites will not be created, even if they exist in the list of locations.

component_delimiter::String = "~|~": When there are multiple component names to match, seperate them by this delimiter.

Clapeyron.getparams — Function

params, sites = getparams(components,locations;kwargs...)

returns a Dict{String,ClapeyronParam} containing all the parameters found for the list of components in the available CSVs. locations are the locations relative to Clapeyron database. the available keywords are the ones used in ParamOptions

if return_sites is set to false, getparams will only return the found params.

Creating Files from Parameters

Clapeyron.ParamTable — Function

ParamTable(type::Symbol,table,location = nothing,name = nothing, options = ParamOptions())

Creates a clapeyron CSV file and returns the location of that file. the type determines the table type:

:single creates a table with single parameters

:pair creates a table with pair parameters

:assoc creates a table with association parameters

:group creates a table with association parameters

By default, the name is generated randomly, and the table is stored as a temporary scratch space (provided by Scratch.jl). You can clean said scratch space by using Clapeyron.cleartemp!().

Examples:

julia> data = (species = ["water"],Mw = [18.03]) #it could be a Dict, a named tuple, or any Tables.jl compatible table
(species = ["water"], Mw = [18.9])

julia> file = ParamTable(:single,data ,name="water_new_mw")
"C:\Users\user\.julia\scratchspaces\7c7805af-46cc-48c9-995b-ed0ed2dc909a\ParamTables\singledata_water_new_mw.csv"

julia> model = PCSAFT(["water","methanol"],userlocations = [file])
PCSAFT{BasicIdeal} with 2 components:
 "water"
 "methanol"
Contains parameters: Mw, segment, sigma,
epsilon, epsilon_assoc, bondvol

julia> model.params.Mw
SingleParam{Float64}("Mw") with 2 components:
 "water" => 18.9
 "methanol" => 32.042

Clapeyron.cleartemp! — Function

cleartemp!()

Deletes all files in the temporary Clapeyron scratch space, used to store the csvs created by ParamTable.

Parameter types

Clapeyron.SingleParam — Type

SingleParam{T}

Struct designed to contain single parameters. Basically a vector with some extra info.

Creation:

julia> mw = SingleParam("molecular weight",["water","ammonia"],[18.01,17.03])
SingleParam{Float64}("molecular weight") with 2 components:
 "water" => 18.01
 "ammonia" => 17.03
julia> mw.values
2-element Vector{Float64}:
 18.01
 17.03
julia> mw.components
2-element Vector{String}:
 "water"
 "ammonia"
julia> mw2 = SingleParam(mw,"new name")
SingleParam{Float64}("new name") with 2 components:
 "water" => 18.01
 "ammonia" => 17.03
julia> has_oxigen = [true,false]; has_o = SingleParam(mw2,has_oxigen)
SingleParam{Bool}("new name") with 2 components:
 "water" => true
 "ammonia" => false

Example usage in models:

function molecular_weight(model,molar_frac)
    mw = model.params.mw.values
    res = zero(eltype(molarfrac))
    for i in @comps #iterating through all components
        res += molar_frac[i]*mw[i]
    end
    return res
end

Clapeyron.PairParam — Type

PairParam{T}

Struct designed to contain pair data. used a matrix as underlying data storage.

Creation:

julia> kij = PairParam("interaction params",["water","ammonia"],[0.1 0.0;0.1 0.0])
PairParam{Float64}["water", "ammonia"]) with values:
2×2 Matrix{Float64}:
 0.1  0.0
 0.1  0.0
julia> kij.values
2×2 Matrix{Float64}:
 0.1  0.0
 0.1  0.0
julia> kij.diagvalues
2-element view(::Vector{Float64}, 
1:3:4) with eltype Float64:
 0.1
 0.0

Example usage in models:

#lets compute ∑xᵢxⱼkᵢⱼ
function alpha(model,x)
    kij = model.params.kij.values
    ki = model.params.kij.diagvalues
    res = zero(eltype(molarfrac))
    for i in @comps 
        @show ki[i] #diagonal values
        for j in @comps 
            res += x[i]*x[j]*kij[i,j]
        end
    end
    return res
end

Clapeyron.AssocParam — Type

AssocParam{T}

Struct holding association parameters.

Clapeyron.GroupParam — Type

GroupParam

Struct holding group parameters.contains:

components: a list of all components

groups: a list of groups names for each component

i_groups: a list containing the number of groups for each component

n_groups: a list of the group multiplicity of each group corresponding to each group in i_groups

flattenedgroups: a list of all unique groups–the parameters correspond to this list

n_flattenedgroups: the group multiplicities corresponding to each group in flattenedgroups

i_flattenedgroups: an iterator that goes through the indices for each flattened group

You can create a group param by passing a `Vector{Tuple{String, Vector{Pair{String, Int64}}}}. For example:

julia> grouplist = [
           ("ethanol", ["CH3"=>1, "CH2"=>1, "OH"=>1]), 
           ("nonadecanol", ["CH3"=>1, "CH2"=>18, "OH"=>1]),
           ("ibuprofen", ["CH3"=>3, "COOH"=>1, "aCCH"=>1, "aCCH2"=>1, "aCH"=>4])];

julia> groups = GroupParam(grouplist)
GroupParam with 3 components:
 "ethanol": "CH3" => 1, "CH2" => 1, "OH" => 1
 "nonadecanol": "CH3" => 1, "CH2" => 18, "OH" => 1    
 "ibuprofen": "CH3" => 3, "COOH" => 1, "aCCH" => 1, "aCCH2" => 1, "aCH" => 4

julia> groups.flattenedgroups
7-element Vector{String}:
 "CH3"
 "CH2"
 "OH"
 "COOH"
 "aCCH"
 "aCCH2"
 "aCH"

julia> groups.i_groups
3-element Vector{Vector{Int64}}:
 [1, 2, 3]
 [1, 2, 3]
 [1, 4, 5, 6, 7]

julia> groups.n_groups
3-element Vector{Vector{Int64}}:
 [1, 1, 1]
 [1, 18, 1]
 [3, 1, 1, 1, 4]

julia> groups.n_flattenedgroups
 3-element Vector{Vector{Int64}}:
 [1, 1, 1, 0, 0, 0, 0]
 [1, 18, 1, 0, 0, 0, 0]
 [3, 0, 0, 1, 1, 1, 4]

if you have CSV with group data, you can also pass those, to automatically query the missing groups in your input vector:

julia> grouplist = [
           "ethanol", 
           ("nonadecanol", ["CH3"=>1, "CH2"=>18, "OH"=>1]),
           ("ibuprofen", ["CH3"=>3, "COOH"=>1, "aCCH"=>1, "aCCH2"=>1, "aCH"=>4])];

           julia> groups = GroupParam(grouplist, ["SAFT/SAFTgammaMie/SAFTgammaMie_groups.csv"])
           GroupParam with 3 components:
            "ethanol": "CH2OH" => 1, "CH3" => 1
            "nonadecanol": "CH3" => 1, "CH2" => 18, "OH" => 1    
            "ibuprofen": "CH3" => 3, "COOH" => 1, "aCCH" => 1, "aCCH2" => 1, "aCH" => 4

In this case, SAFTGammaMie files support the second order group CH2OH.

Clapeyron.SiteParam — Type

SiteParam

Struct holding site parameters. Is built by parsing all association parameters in the input CSV files. It has the following fields:

components: a list of all components (or groups in Group Contribution models)

sites: a list containing a list of all sites corresponding to each component (or group) in the components field

n_sites: a list of the site multiplicities corresponding to each site in flattenedsites

flattenedsites: a list of all unique sites

i_sites: an iterator that goes through the indices corresponding to each site in flattenedsites

n_flattenedsites: the site multiplicities corresponding to each site in flattenedsites

i_flattenedsites: an iterator that goes through the indices for each flattened site

Let's explore the sites in a 3-component SAFTGammaMie model:

julia> model3 = SAFTgammaMie([    
                "ethanol",
                ("nonadecanol", ["CH3"=>1, "CH2"=>18, "OH"=>1]),     
                ("ibuprofen", ["CH3"=>3, "COOH"=>1, "aCCH"=>1, "aCCH2"=>1, "aCH"=>4])
                               ])

SAFTgammaMie{BasicIdeal} with 3 components:
 "ethanol"
 "nonadecanol"
 "ibuprofen"
Contains parameters: segment, shapefactor, lambda_a, lambda_r, sigma, epsilon, epsilon_assoc, bondvol 

julia> model3.sites
SiteParam with 8 sites:
 "CH2OH": "H" => 1, "e1" => 2     
 "CH3": (no sites)
 "CH2": (no sites)
 "OH": "H" => 1, "e1" => 2        
 "COOH": "e2" => 2, "H" => 1, "e1" => 2
 "aCCH": (no sites)
 "aCCH2": (no sites)
 "aCH": (no sites)

julia> model3.sites.flattenedsites
3-element Vector{String}:
 "H"
 "e1"
 "e2"

julia> model3.sites.i_sites       
8-element Vector{Vector{Int64}}:
 [1, 2]
 []
 []
 [1, 2]
 [1, 2, 3]
 []
 []
 []

julia> model3.sites.n_sites       
8-element Vector{Vector{Int64}}:
 [1, 2]
 []
 []
 [1, 2]
 [2, 1, 2]
 []
 []
 []

Clapeyron.AssocOptions — Type

AssocOptions(;rtol = 1e-12,atol = 1e-12,max_iters = 1000,dampingfactor = 0.5,combining =:nocombining)

Struct containing iteration parameters for the solver of association sites.

the combining option controls the type of combining rule applied to the association strength:

sparse_nocombining (default). Does not perform any combination rules over the association strength, and returns a sparse matrix.

dense_nocombining. Does not perform any combination rules over the association strength, and returns a dense matrix.

elliott combining rule: Δ(i,j,a,b) = √(Δ(i,i,a,b)*Δ(j,j,a,b)). Returns a dense matrix.

Association Scheme matters

Elliott's rule requires that both Δ(i,i,a,b) and Δ(j,j,a,b) are non-zero, that means that components that don't self associate will not be combined.

Combining Rules

Clapeyron.kij_mix — Function

kij_mix(f,p::ClapeyronParam,k::PairParam)::PairParam
kij_mix(f,p::ClapeyronParam)::PairParam

General combining rule for pair parameter with a kᵢⱼ interaction parameter. returns a pair parameter with non diagonal entries equal to:

pᵢⱼ = f(pᵢ,pⱼ,kᵢⱼ)

Where f is a 'combining' function that follows the rules:

f(pᵢ,pⱼ,0) = f(pⱼ,pᵢ,0)
f(pᵢ,pᵢ,0) = pᵢ

and k must follow:

kᵢᵢ = 0 

Ignores non-diagonal entries already set.

If a Single Parameter is passed as input, it will be converted to a Pair Parameter with pᵢᵢ = pᵢ.

Clapeyron.pair_mix — Function

pair_mix(g,P::ClapeyronParam,Q::ClapeyronParam)::PairParam
pair_mix(g,P::ClapeyronParam,Q::ClapeyronParam)::PairParam

General combining rule for a pair and a single parameter. returns a pair parameter P with non diagonal entries equal to:

Pᵢⱼ = g(Pᵢ,Pⱼ,Qᵢ,Qⱼ,Qᵢⱼ)

Where f is a 'combining' function that follows the rules:

Pᵢⱼ = Pⱼᵢ = g(Pᵢ,Pⱼ,Qᵢ,Qⱼ,Qᵢⱼ) = g(Pⱼ,Pᵢ,Qⱼ,Qᵢ,Qᵢⱼ)
g(Pᵢ,Pᵢ,Qᵢ,Qᵢ,Qᵢ) = Pᵢ

it is a more general form of kij_mix, where kij_mix(f,P,Q) == pair_mix(g,P,Q) is correct if:

f(Pᵢ,Pⱼ,Qᵢⱼ) = g(Pᵢ,Pⱼ,_,_,Qᵢⱼ)

Clapeyron.sigma_LorentzBerthelot — Function

sigma_LorentzBerthelot(σ::ClapeyronParam,ζ::PairParam)::PairParam
sigma_LorentzBerthelot(σ::ClapeyronParam)::PairParam

Combining rule for a single or pair parameter. returns a pair parameter with non diagonal entries equal to:

σᵢⱼ = (1 - ζᵢⱼ)*(σᵢ + σⱼ)/2

If ζᵢⱼ is not defined, the definition is reduced to a simple arithmetic mean:

σᵢⱼ = (σᵢ + σⱼ)/2

Ignores non-diagonal entries already set.

If a Single Parameter is passed as input, it will be converted to a Pair Parameter with σᵢᵢ = σᵢ.

Clapeyron.epsilon_LorentzBerthelot — Function

epsilon_LorentzBerthelot(ϵ::ClapeyronParam,k::PairParam)::PairParam
epsilon_LorentzBerthelot(ϵ::ClapeyronParam)::PairParam

Combining rule for a single or pair parameter. returns a pair parameter with non diagonal entries equal to:

ϵᵢⱼ = (1 - kᵢⱼ)*√(ϵᵢϵⱼ)

If kᵢⱼ is not defined, the definition is reduced to a simple geometric mean:

ϵᵢⱼ = √(ϵᵢϵⱼ)

Ignores non-diagonal entries already set.

If a Single Parameter is passed as input, it will be converted to a Pair Parameter with ϵᵢᵢ = ϵᵢ.

Clapeyron.epsilon_HudsenMcCoubrey — Function

epsilon_HudsenMcCoubrey(ϵ::ClapeyronParam,σ::PairParam)::PairParam
epsilon_HudsenMcCoubrey(ϵ::ClapeyronParam)::PairParam

Combining rule for a single or pair parameter. returns a pair parameter with non diagonal entries equal to:

ϵᵢⱼ = √(ϵᵢϵⱼ)*(σᵢᵢ^3 * σⱼⱼ^3)/σᵢⱼ^6 

If σᵢⱼ is not defined, the definition is reduced to a simple geometric mean:

ϵᵢⱼ = √(ϵᵢϵⱼ)

Ignores non-diagonal entries already set.

If a Single Parameter is passed as input, it will be converted to a Pair Parameter with ϵᵢᵢ = ϵᵢ.

Clapeyron.lambda_LorentzBerthelot — Function

lambda_LorentzBerthelot(λ::ClapeyronParam,k = 3)::PairParam

Combining rule for a single or pair parameter. returns a pair parameter with non diagonal entries equal to:

λᵢⱼ = k + √((λᵢᵢ - k)(λⱼⱼ - k))

with k = 0 the definition is reduced to a simple geometric mean:

ϵᵢⱼ = √(ϵᵢϵⱼ)

Ignores non-diagonal entries already set.

If a Single Parameter is passed as input, it will be converted to a Pair Parameter with λᵢᵢ = λᵢ.

Clapeyron.lambda_squarewell — Function

lambda_squarewell(λ::ClapeyronParam,σ::PairParam)::PairParam

Combining rule for a single or pair parameter. returns a pair parameter with non diagonal entries equal to:

λᵢⱼ = (σᵢᵢλᵢᵢ + σⱼⱼλⱼⱼ)/(σᵢᵢ + σⱼⱼ)

Ignores non-diagonal entries already set.

If a Single Parameter is passed as input, it will be converted to a Pair Parameter with λᵢᵢ = λᵢ.

Group Combining Rules

Clapeyron.group_sum — Function

group_sum(groups::GroupParam,P::SingleParameter)

Given a GroupParam and a Single Parameter P for group data, it will return a single parameter p of component data, where:

pᵢ = ∑Pₖνᵢₖ

where νᵢₖ is the number of groups k at component i.

group_sum(groups::GroupParam,P::AbstractVector)

Given a GroupParam and a Vector P for group data, it will return a Vector p of component data, where:

pᵢ = ∑Pₖνᵢₖ

where νᵢₖ is the number of groups k at component i.

group_sum(groups::GroupParam,::Nothing)

Given a GroupParam, it will return a vector p of component data, where:

pᵢ = ∑νᵢₖ

where νᵢₖ is the number of groups k at component i.

Clapeyron.group_pairmean — Function

group_pairmean(groups::GroupParam,param::PairParam)
group_pairmean(f,groups::GroupParam,param::SingleParam)

Given a GroupParamand a parameter P it will return a single parameter p of component data, where:

pᵢ = ∑νᵢₖ(∑(νᵢₗ*P(i,j))) / ∑νᵢₖ(∑νᵢₗ)

where νᵢₖ is the number of groups k at component i and P(i,j) depends on the type of P:

if P is a single paremeter, then P(i,j) = f(P[i],P[j])

if P is a pair paremeter, then P(i,j) = p[i,j]

Clapeyron.group_matrix — Function

group_matrix(groups::GroupParam)

returns a matrix of size (k,i) with values νₖᵢ. when multiplied with a molar amount, it returns the amount of moles of each group.

Model Splitting

Clapeyron.split_model — Function

split_model(model::EoSModel)

Takes in a model for a multi-component system and returns a vector of model for each pure system.

Example:

julia> gerg2 = GERG2008(["propane","pentane"])
GERG008 model with 2 components:
"propane"
"pentane"

julia> split_model(gerg2)
2-element Vector{GERG2008}:
 GERG2008("propane")
 GERG2008("pentane")

Clapeyron.is_splittable — Function

is_splittable(model)::Bool

Trait to determine if a EoSModel should be splitted by itself or can be simply filled into a vector. This is useful in the case of models without any parameters, as those models are impossible by definition to split, because they don't have any underlying data.

The Default is is_splittable(model) = true.

Clapeyron.index_reduction — Function

index_reduction(model::EoSModel,z,zmin = sum(z)*4*eps(eltype(z)))

Removes any component with composition z[i] < zmin. returns a reduced model model_r and a vector of indices idx_r, such as:

model_r,idx_r = index_reduction(model,z)
eos(model,V,T,z) ≈ eos(model_r,V,T,z[idx_r])

if the model does not have empty compositions, it will just return the input model

Model Citing

Clapeyron.doi — Function

doi(model)

Returns a Vector of strings containing the top-level bibliographic references of the model, in DOI format.

julia> umr = UMRPR(["water"],idealmodel = WalkerIdeal);Clapeyron.doi(umr)
1-element Vector{String}:
 "10.1021/I160057A011"

Clapeyron.cite — Function

cite(model)

Returns a Vector of strings containing all bibliographic references of the model, in DOI format. this includes any nested models.

julia> umr = UMRPR(["water"],idealmodel = WalkerIdeal);Clapeyron.cite(umr) #should cite UMRPR, UNIFAC, WalkerIdeal
4-element Vector{String}:
 "10.1021/I160057A011"
 "10.1021/ie049580p"
 "10.1021/i260064a004"
 "10.1021/acs.jced.0c00723"

This list will displayed by each EoSModel on future versions. you can enable/disable this by setting ENV["CLAPEYRON_SHOW_REFERENCES"] = "TRUE"/"FALSE"