Index

Clapeyron.NRTL

Clapeyron.PSRKUNIFAC

Clapeyron.UNIFAC

Clapeyron.UNIQUAC

Clapeyron.VTPRUNIFAC

Clapeyron.Wilson

Clapeyron.ogUNIFAC

Activity Models

There are two alternatives on the definition of an activity model:

Defining an excess gibbs energy function

Defining an activity coefficient function

those two can be converted between one form to another via:

$\gamma_{i} = \frac{\partial{G^E}}{\partial{n_i}}$

$\ln{\gamma_{i}} = \frac{1}{RT}\frac{\partial{G^E}}{\partial{n_i}}$

When defining one form, the other is derived automatically.

Those functions can also be derived from any arbitrary equation of state:

$\frac{\partial{G^E}}{\partial{n_i}}= \mu_i - \mu^{0}_i$

Where $\mu_i$ and $\mu^{0}_i$ are the mixture and pure chemical potentials of component i. in this case, those potentials are dependent of the pressure. whereas activity models are usually only temperature and composition dependent.

Clapeyron.Wilson — Type

Wilson <: ActivityModel

Wilson(components::Vector{String};
puremodel = PR,
userlocations = String[], 
pure_userlocations = String[],
verbose = false)

Input parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

ZRA: Single Parameter (Float64) - Rackett Compresibility factor

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

g: Pair Parameter (Float64, asymetrical, defaults to 0) - Interaction Parameter

Input models

puremodel: model to calculate pure pressure-dependent properties

Description

Wilson activity model, with Rackett correlation for liquid volume:

Gᴱ = nRT∑xᵢlog(∑xⱼjΛᵢⱼ)
Λᵢⱼ = exp(-gᵢⱼ/T)*Vⱼ/Vᵢ
Vᵢ = (RTcᵢ/Pcᵢ)(0.29056 - 0.08775ZRAᵢ)^(1 + (1-T/Tcᵢ)^2/7)

References

Wilson, G. M. (1964). Vapor-liquid equilibrium. XI. A new expression for the excess free energy of mixing. Journal of the American Chemical Society, 86(2), 127–130. doi:10.1021/ja01056a002

Clapeyron.NRTL — Type

NRTL <: ActivityModel

function NRTL(components::Vector{String};
puremodel=PR,
userlocations=String[],
pure_userlocations = String[],
verbose=false)

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

a: Pair Parameter (Float64, asymetrical, defaults to 0) - Interaction Parameter

b: Pair Parameter (Float64, asymetrical, defaults to 0) - Interaction Parameter

c: Pair Parameter (Float64, asymetrical, defaults to 0) - Interaction Parameter

Input models

puremodel: model to calculate pure pressure-dependent properties

Description

NRTL (Non Random Two Fluid) activity model:

Gᴱ = nRT∑[xᵢ(∑τⱼᵢGⱼᵢxⱼ)/(∑Gⱼᵢxⱼ)]
Gᵢⱼ exp(-cᵢⱼτᵢⱼ)
τᵢⱼ = aᵢⱼ + bᵢⱼ/T

References

Renon, H., & Prausnitz, J. M. (1968). Local compositions in thermodynamic excess functions for liquid mixtures. AIChE journal. American Institute of Chemical Engineers, 14(1), 135–144. doi:10.1002/aic.690140124

Clapeyron.UNIQUAC — Type

UNIQUACModel <: ActivityModel

UNIQUAC(components::Vector{String};
puremodel = PR,
userlocations = String[], 
pure_userlocations = String[],
verbose = false)

Input parameters

a: Pair Parameter (Float64, asymetrical, defaults to 0) - Binary Interaction Energy Parameter

r: Single Parameter (Float64) - Normalized Van der Vals volume

q: Single Parameter (Float64) - Normalized Surface Area

q_p: Single Parameter (Float64) - Modified Normalized Surface Area

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

Input models

puremodel: model to calculate pure pressure-dependent properties

UNIQUAC (Universal QuasiChemical Activity Coefficients) activity model:

Gᴱ = nRT(gᴱ(comb) + gᴱ(res))
gᴱ(comb) = ∑[xᵢlog(Φᵢ/xᵢ) + 5qᵢxᵢlog(θᵢ/Φᵢ)]
gᴱ(res) = -∑xᵢqᵖᵢlog(∑θᵖⱼτⱼᵢ)
θᵢ = qᵢxᵢ/∑qᵢxᵢ
θᵖ = qᵖᵢxᵢ/∑qᵖᵢxᵢ
Φᵢ = rᵢxᵢ/∑rᵢxᵢ
τᵢⱼ = exp(-aᵢⱼ/T)

References

Abrams, D. S., & Prausnitz, J. M. (1975). Statistical thermodynamics of liquid mixtures: A new expression for the excess Gibbs energy of partly or completely miscible systems. AIChE journal. American Institute of Chemical Engineers, 21(1), 116–128. doi:10.1002/aic.690210115

Clapeyron.ogUNIFAC — Type

ogUNIFACModel <: UNIFACModel

ogUNIFAC(components::Vector{String};
puremodel = PR, 
userlocations = String[],
pure_userlocations = String[],
verbose = false)

Input parameters

R: Single Parameter (Float64) - Normalized group Van der Vals volume

Q: Single Parameter (Float64) - Normalized group Surface Area

A: Pair Parameter (Float64, asymetrical, defaults to 0) - Binary group Interaction Energy Parameter

Input models

puremodel: model to calculate pure pressure-dependent properties

UNIFAC (UNIQUAC Functional-group Activity Coefficients) activity model.

Original formulation.

The Combinatorial part corresponds to an GC-averaged modified UNIQUAC model. The residual part iterates over groups instead of components.

Gᴱ = nRT(gᴱ(comb) + gᴱ(res))

Combinatorial part:

gᴱ(comb) = ∑[xᵢlog(Φᵢ/xᵢ) + 5qᵢxᵢlog(θᵢ/Φᵢ)]
θᵢ = qᵢxᵢ/∑qᵢxᵢ
Φᵢ = rᵢxᵢ/∑rᵢxᵢ
rᵢ = ∑Rₖνᵢₖ for k ∈ groups
qᵢ = ∑Qₖνᵢₖ for k ∈ groups

Residual Part:

gᴱ(residual) = -v̄∑XₖQₖlog(∑ΘₘΨₘₖ)
v̄ = ∑∑xᵢνᵢₖ for k ∈ groups,  for i ∈ components
Xₖ = (∑xᵢνᵢₖ)/v̄ for i ∈ components 
Θₖ = QₖXₖ/∑QₖXₖ
Ψₖₘ = exp(-(Aₖₘ/T)

References

Fredenslund, A., Gmehling, J., Michelsen, M. L., Rasmussen, P., & Prausnitz, J. M. (1977). Computerized design of multicomponent distillation columns using the UNIFAC group contribution method for calculation of activity coefficients. Industrial & Engineering Chemistry Process Design and Development, 16(4), 450–462. doi:10.1021/i260064a004

Clapeyron.UNIFAC — Type

UNIFACModel <: ActivityModel

UNIFAC(components::Vector{String};
puremodel = PR,
userlocations = String[], 
pure_userlocations = String[],
verbose = false)

Input parameters

R: Single Parameter (Float64) - Normalized group Van der Vals volume

Q: Single Parameter (Float64) - Normalized group Surface Area

A: Pair Parameter (Float64, asymetrical, defaults to 0) - Binary group Interaction Energy Parameter

B: Pair Parameter (Float64, asymetrical, defaults to 0) - Binary group Interaction Energy Parameter

C: Pair Parameter (Float64, asymetrical, defaults to 0) - Binary group Interaction Energy Parameter

Input models

puremodel: model to calculate pure pressure-dependent properties

Description

UNIFAC (UNIQUAC Functional-group Activity Coefficients) activity model.

Modified UNIFAC (Dortmund) implementation.

The Combinatorial part corresponds to an GC-averaged modified UNIQUAC model. The residual part iterates over groups instead of components.

Gᴱ = nRT(gᴱ(comb) + gᴱ(res))

Combinatorial part:

gᴱ(comb) = ∑[xᵢlog(Φ'ᵢ) + 5qᵢxᵢlog(θᵢ/Φᵢ)]
θᵢ = qᵢxᵢ/∑qᵢxᵢ
Φᵢ = rᵢxᵢ/∑rᵢxᵢ
Φ'ᵢ = rᵢ^(0.75)/∑xᵢrᵢ^(0.75)
rᵢ = ∑Rₖνᵢₖ for k ∈ groups
qᵢ = ∑Qₖνᵢₖ for k ∈ groups

Residual Part:

gᴱ(residual) = -v̄∑XₖQₖlog(∑ΘₘΨₘₖ)
v̄ = ∑∑xᵢνᵢₖ for k ∈ groups,  for i ∈ components
Xₖ = (∑xᵢνᵢₖ)/v̄ for i ∈ components 
Θₖ = QₖXₖ/∑QₖXₖ
Ψₖₘ = exp(-(Aₖₘ + BₖₘT + CₖₘT²)/T)

References

Fredenslund, A., Gmehling, J., Michelsen, M. L., Rasmussen, P., & Prausnitz, J. M. (1977). Computerized design of multicomponent distillation columns using the UNIFAC group contribution method for calculation of activity coefficients. Industrial & Engineering Chemistry Process Design and Development, 16(4), 450–462. doi:10.1021/i260064a004

Weidlich, U.; Gmehling, J. A modified UNIFAC model. 1. Prediction of VLE, hE, and.gamma..infin. Ind. Eng. Chem. Res. 1987, 26, 1372–1381.

Clapeyron.PSRKUNIFAC — Type

PSRKUNIFACModel <: UNIFACModel

PSRKUNIFAC(components::Vector{String};
puremodel = PR,
userlocations = String[], 
pure_userlocations = String[],
verbose = false)

Input parameters

R: Single Parameter (Float64) - Normalized group Van der Vals volume

Q: Single Parameter (Float64) - Normalized group Surface Area

A: Pair Parameter (Float64, asymetrical, defaults to 0) - Binary group Interaction Energy Parameter

B: Pair Parameter (Float64, asymetrical, defaults to 0) - Binary group Interaction Energy Parameter

C: Pair Parameter (Float64, asymetrical, defaults to 0) - Binary group Interaction Energy Parameter

Input models

puremodel: model to calculate pure pressure-dependent properties

Description

UNIFAC (UNIQUAC Functional-group Activity Coefficients) activity model.

Modified UNIFAC (Dortmund) implementation, with parameters tuned to the Predictive Soave-Redlich-Kwong (PSRK) EoS.

The Combinatorial part corresponds to an GC-averaged modified UNIQUAC model. The residual part iterates over groups instead of components.

Gᴱ = nRT(gᴱ(comb) + gᴱ(res))

Combinatorial part:

gᴱ(comb) = ∑[xᵢlog(Φ'ᵢ) + 5qᵢxᵢlog(θᵢ/Φᵢ)]
θᵢ = qᵢxᵢ/∑qᵢxᵢ
Φᵢ = rᵢxᵢ/∑rᵢxᵢ
Φ'ᵢ = rᵢ^(0.75)/∑xᵢrᵢ^(0.75)
rᵢ = ∑Rₖνᵢₖ for k ∈ groups
qᵢ = ∑Qₖνᵢₖ for k ∈ groups

Residual Part:

gᴱ(residual) = -v̄∑XₖQₖlog(∑ΘₘΨₘₖ)
v̄ = ∑∑xᵢνᵢₖ for k ∈ groups,  for i ∈ components
Xₖ = (∑xᵢνᵢₖ)/v̄ for i ∈ components 
Θₖ = QₖXₖ/∑QₖXₖ
Ψₖₘ = exp(-(Aₖₘ + BₖₘT + CₖₘT²)/T)

References

Fredenslund, A., Gmehling, J., Michelsen, M. L., Rasmussen, P., & Prausnitz, J. M. (1977). Computerized design of multicomponent distillation columns using the UNIFAC group contribution method for calculation of activity coefficients. Industrial & Engineering Chemistry Process Design and Development, 16(4), 450–462. doi:10.1021/i260064a004

Weidlich, U.; Gmehling, J. A modified UNIFAC model. 1. Prediction of VLE, hE, and.gamma..infin. Ind. Eng. Chem. Res. 1987, 26, 1372–1381.

Horstmann, S., Jabłoniec, A., Krafczyk, J., Fischer, K., & Gmehling, J. (2005). PSRK group contribution equation of state: comprehensive revision and extension IV, including critical constants and α-function parameters for 1000 components. Fluid Phase Equilibria, 227(2), 157–164. doi:10.1016/j.fluid.2004.11.002

Clapeyron.VTPRUNIFAC — Type

VTPRUNIFACModel <: UNIFACModel

VTPRUNIFAC(components::Vector{String};
puremodel = PR,
userlocations = String[], 
pure_userlocations = String[],
verbose = false)

Input parameters

Q: Single Parameter (Float64) - Normalized group Surface Area

A: Pair Parameter (Float64, asymetrical, defaults to 0) - Binary group Interaction Energy Parameter

B: Pair Parameter (Float64, asymetrical, defaults to 0) - Binary group Interaction Energy Parameter

C: Pair Parameter (Float64, asymetrical, defaults to 0) - Binary group Interaction Energy Parameter

Input models

puremodel: model to calculate pure pressure-dependent properties

Description

UNIFAC (UNIQUAC Functional-group Activity Coefficients) activity model.

Modified UNIFAC (Dortmund) implementation, only residual part, activity model used for the Volume-Translated Peng-Robinson (VTPR) EoS.

The residual part iterates over groups instead of components.

Gᴱ = nRT(gᴱ(res))
gᴱ(res) = -v̄∑XₖQₖlog(∑ΘₘΨₘₖ)
v̄ = ∑∑xᵢνᵢₖ for k ∈ groups,  for i ∈ components
Xₖ = (∑xᵢνᵢₖ)/v̄ for i ∈ components 
Θₖ = QₖXₖ/∑QₖXₖ
Ψₖₘ = exp(-(Aₖₘ + BₖₘT + CₖₘT²)/T)

References

Fredenslund, A., Gmehling, J., Michelsen, M. L., Rasmussen, P., & Prausnitz, J. M. (1977). Computerized design of multicomponent distillation columns using the UNIFAC group contribution method for calculation of activity coefficients. Industrial & Engineering Chemistry Process Design and Development, 16(4), 450–462. doi:10.1021/i260064a004

Weidlich, U.; Gmehling, J. A modified UNIFAC model. 1. Prediction of VLE, hE, and.gamma..infin. Ind. Eng. Chem. Res. 1987, 26, 1372–1381.

Ahlers, J., & Gmehling, J. (2001). Development of an universal group contribution equation of state. Fluid Phase Equilibria, 191(1–2), 177–188. doi:10.1016/s0378-3812(01)00626-4