Index

Clapeyron.BasicIdeal

Clapeyron.IAPWS95Ideal

Clapeyron.JobackIdeal

Clapeyron.LJRefIdeal

Clapeyron.MonomerIdeal

Clapeyron.ReidIdeal

Clapeyron.WalkerIdeal

Correlation Models

Correlation models are, as their name says, fitted equations that express one property of a compound. they meant to be used in conjunction with other models (like Activity models that require a saturated liquid volume), or via a CompositeModel. because they only overload one property, the way to define a correlation is different than normal EoSModels

Saturation Correlations

Saturation Correlations are any EoSModel that are subtypes of SaturationModel. return psat(T) and the upper limit (Tc,Pc) pair. To define saturation correlations, you need to overload:

function crit_pure(model::MySaturationModel <: SaturationModel)
    ...
    return (Tc,Pc,NaN)
end

function Clapeyron.saturation_pressure_impl(model::MySaturationModel <: SaturationModel,T,::SaturationCorrelation)
    ...
    return (psat,NaN,NaN)
end

Saturation Models and Types

Clapeyron.SaturationModel — Type

SaturationModel <: EoSModel

Abstract type for Saturation correlation models.

Clapeyron.SaturationCorrelation — Type

SaturationCorrelation <: SaturationMethod

saturation method used for dispatch on saturation correlations.

Clapeyron.LeeKeslerSat — Type

LeeKeslerSat <: SaturationModel

LeeKeslerSat(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

pc: Single Parameter (Float64) - Critical Pressure [Pa]

w: Single Parameter (Float64) - acentric factor

Model Parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

pc: Single Parameter (Float64) - Critical Pressure [Pa]

acentricfactor: Single Parameter (Float64) - acentric factor

Description

Lee-Kesler correlation for saturation pressure:

psat(T) = f₀ + ω•f₁
Tr = T/Tc
f₀ = 5.92714 - 6.09648/Tr - 1.28862•log(Tr) + 0.169347•Tr⁶
f₁ = 15.2518 - 15.6875/Tr - 13.4721•log(Tr) + 0.43577•Tr⁶

References

Lee, B. I., & Kesler, M. G. (1975). A generalized thermodynamic correlation based on three-parameter corresponding states. AIChE journal. American Institute of Chemical Engineers, 21(3), 510–527. doi:10.1002/aic.690210313

Clapeyron.DIPPR101Sat — Type

DIPPR101Sat <: SaturationModel

DIPPR101Sat(components::Vector{String};
userlocations::Vector{String}=String[],
verbose::Bool=false)

Input Parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

pc: Single Parameter (Float64) - Critical Pressure [Pa]

A: Single Parameter (Float64)

B: Single Parameter (Float64)

C: Single Parameter (Float64)

D: Single Parameter (Float64)

E: Single Parameter (Float64)

Tmin: Single Parameter (Float64) - mininum Temperature range [K]

Tmax: Single Parameter (Float64) - maximum Temperature range [K]

Model Parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

A: Single Parameter (Float64)

B: Single Parameter (Float64)

C: Single Parameter (Float64)

D: Single Parameter (Float64)

E: Single Parameter (Float64)

Tmin: Single Parameter (Float64) - mininum Temperature range [K]

Tmax: Single Parameter (Float64) - maximum Temperature range [K]

Description

DIPPR 101 Equation for saturation pressure:

psat(T) = exp(A + B/T + C•log(T) + D•T^E)

References

Design Institute for Physical Properties, 1996. DIPPR Project 801 DIPPR/AIChE

Liquid Volume Correlations

Liquid Volume Correlations are any EoSModel that are subtypes of LiquidVolumeModel. They return volume(model,p,T,z, phase = :liquid).

Virial Models

Virial models are defined in terms of the second virial coefficient, B(T,z). the reduced residual helmholtz energy is defined as:

$\frac{A_\mathrm{res}}{Nk_\mathrm{B}T} = \frac{B}{V}$,

To implement a virial model, it is necessary to overload Clapeyron.second_virial_coefficient_impl(model::<:SecondVirialModel,T,z).

Clapeyron.AbbottVirial — Type

AbbottVirial <: SecondVirialModel
AbbottVirial(components;
            idealmodel=BasicIdeal,
            userlocations=String[],
            ideal_userlocations=String[],
            verbose=false)

Input parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

w: Single Parameter (Float64) - Acentric Factor

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

Model Parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

acentricfactor: Single Parameter (Float64) - Acentric Factor

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

Input models

idealmodel: Ideal Model

Description

Virial model using Corresponding State Principles:

B = ∑xᵢxⱼBᵢⱼ
Bᵢⱼ = BrᵢⱼRTcᵢⱼ/Pcᵢⱼ
Brᵢⱼ = B₀ + ωᵢⱼB₁
B₀ = 0.083 + 0.422/Trᵢⱼ^1.6
B₁ = 0.139 - 0.172/Trᵢⱼ^4.2
Trᵢⱼ = T/Tcᵢⱼ
Tcᵢⱼ = √TcᵢTcⱼ
Pcᵢⱼ = (Pcᵢ + Pcⱼ)/2
ωᵢⱼ = (ωᵢ + ωⱼ)/2

References

Smith, H. C. Van Ness Joseph M. Introduction to Chemical Engineering Thermodynamics 4E 1987.

Clapeyron.TsonopoulosVirial — Type

TsonopoulosVirial <: SecondVirialModel
TsonopoulosVirial(components;
        idealmodel=BasicIdeal,
        userlocations=String[],
        ideal_userlocations=String[],
        verbose=false)

Input parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

w: Single Parameter (Float64) - Acentric Factor

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

Model Parameters

Tc: Single Parameter (Float64) - Critical Temperature [K]

Pc: Single Parameter (Float64) - Critical Pressure [Pa]

acentricfactor: Single Parameter (Float64) - Acentric Factor

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

Input models

idealmodel: Ideal Model

Description

Virial model using Corresponding State Principles:

B = ∑xᵢxⱼBᵢⱼ
Bᵢⱼ = BrᵢⱼRTcᵢⱼ/Pcᵢⱼ
Brᵢⱼ = B₀ + ωᵢⱼB₁
B₀ = 0.1445 - 0.330/Trᵢⱼ - 0.1385/Trᵢⱼ^2 - 0.0121/Trᵢⱼ^3 - 0.000607/Trᵢⱼ^8
B₁ = 0.0637 + 0.331/Trᵢⱼ - 0.423/Trᵢⱼ^2 - 0.423/Trᵢⱼ^3 - 0.008/Trᵢⱼ^8
Trᵢⱼ = T/Tcᵢⱼ
Tcᵢⱼ = √TcᵢTcⱼ
Pcᵢⱼ = (Pcᵢ + Pcⱼ)/2
ωᵢⱼ = (ωᵢ + ωⱼ)/2

References

Tsonopoulos, C. (1974). An empirical correlation of second virial coefficients. AIChE Journal. American Institute of Chemical Engineers, 20(2), 263–272. doi:10.1002/aic.690200209

Clapeyron.EoSVirial2 — Type

EoSVirial2 <: SecondVirialModel
EoSVirial2(model;idealmodel = idealmodel(model))

Input models

model: Model providing second virial coefficient is obtained

idealmodel: Ideal Model

Description

Virial model, that just calls the second virial coefficient of the underlying model.

B(T,z) = B(model,T,z)