Index

Clapeyron.ADPCSAFT

Clapeyron.BACKSAFT

Clapeyron.CKSAFT

Clapeyron.COFFEE

Clapeyron.CPA

Clapeyron.CPPCSAFT

Clapeyron.DAPT

Clapeyron.GEPCSAFT

Clapeyron.HeterogcPCPSAFT

Clapeyron.HomogcPCPSAFT

Clapeyron.LJSAFT

Clapeyron.PCPSAFT

Clapeyron.PCSAFT

Clapeyron.QPCPSAFT

Clapeyron.SAFTVRMie

Clapeyron.SAFTVRMie15

Clapeyron.SAFTVRQMie

Clapeyron.SAFTVRSMie

Clapeyron.SAFTVRSW

Clapeyron.SAFTgammaMie

Clapeyron.gcsPCSAFT

Clapeyron.ogSAFT

Clapeyron.pharmaPCSAFT

Clapeyron.sCKSAFT

Clapeyron.sPCSAFT

Clapeyron.softSAFT

Clapeyron.softSAFT2016

Clapeyron.solidsoftSAFT

Clapeyron.structSAFTgammaMie

Clapeyron.sCPA

SAFT Models

Clapeyron.ogSAFT — Type

ogSAFTModel <: SAFTModel

ogSAFT(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

(original) Statistical Associating Fluid Theory (og-SAFT) Equation of State

References

Chapman, W. G., Gubbins, K. E., Jackson, G., & Radosz, M. (1989). SAFT: Equation-of-state solution model for associating fluids. Fluid Phase Equilibria, 52, 31–38. doi:10.1016/0378-3812(89)80308-5

Chapman, W. G., Gubbins, K. E., Jackson, G., & Radosz, M. (1990). New reference equation of state for associating liquids. Industrial & Engineering Chemistry Research, 29(8), 1709–1721. doi:10.1021/ie00104a021

Clapeyron.CKSAFT — Type

CKSAFTModel <: SAFTModel

CKSAFT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

vol: Single Parameter (Float64) - Segment Volume [dm^3]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

c: Single Parameter (Float64) - Dispersion T-dependent parameter (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

c: Single Parameter (Float64) - Dispersion T-dependent parameter (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Chen and Kreglewski SAFT (CK-SAFT)

References

Huang, S. H., & Radosz, M. (1990). Equation of state for small, large, polydisperse, and associating molecules. Industrial & Engineering Chemistry Research, 29(11), 2284–2294. doi:10.1021/ie00107a014

Huang, S. H., & Radosz, M. (1991). Equation of state for small, large, polydisperse, and associating molecules: extension to fluid mixtures. Industrial & Engineering Chemistry Research, 30(8), 1994–2005. doi:10.1021/ie00056a050

Clapeyron.sCKSAFT — Type

sCKSAFTModel <: CKSAFTModel

sCKSAFT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

vol: Single Parameter (Float64) - Segment Volume [dm^3]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Simplified Chen and Kreglewski SAFT (sCK-SAFT)

References

Huang, S. H., & Radosz, M. (1990). Equation of state for small, large, polydisperse, and associating molecules. Industrial & Engineering Chemistry Research, 29(11), 2284–2294. doi:10.1021/ie00107a014

Huang, S. H., & Radosz, M. (1991). Equation of state for small, large, polydisperse, and associating molecules: extension to fluid mixtures. Industrial & Engineering Chemistry Research, 30(8), 1994–2005. doi:10.1021/ie00056a050

Fu, Y.-H., & Sandler, S. I. (1995). A simplified SAFT equation of state for associating compounds and mixtures. Industrial & Engineering Chemistry Research, 34(5), 1897–1909. doi:10.1021/ie00044a042

Clapeyron.BACKSAFT — Type

BACKSAFTModel <: SAFTModel

BACKSAFT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

vol: Single Parameter (Float64) - Segment Volume [dm^3]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K/mol]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

c: Single Parameter (Float64) - Adjustable parameter (no units)

alpha: Single Parameter (Float64) - Non-spherical deviation (no units)

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

c: Single Parameter (Float64) - Adjustable parameter (no units)

alpha: Single Parameter (Float64) - Non-spherical deviation (no units)

Input models

idealmodel: Ideal Model

Description

BACKSAFT

References

Mi, J.-G., Chen, J., Gao, G.-H., & Fei, W.-Y. (2002). Equation of state extended from SAFT with improved results for polar fluids across the critical point. Fluid Phase Equilibria, 201(2), 295–307. doi:10.1016/s0378-3812(02)00093-6

Clapeyron.PCSAFT — Type

PCSAFTModel <: SAFTModel

PCSAFT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Perturbed-Chain SAFT (PC-SAFT)

References

Gross, J., & Sadowski, G. (2001). Perturbed-chain SAFT: An equation of state based on a perturbation theory for chain molecules. Industrial & Engineering Chemistry Research, 40(4), 1244–1260. doi:10.1021/ie0003887

Gross, J., & Sadowski, G. (2002). Application of the perturbed-chain SAFT equation of state to associating systems. Industrial & Engineering Chemistry Research, 41(22), 5510–5515. doi:10.1021/ie010954d

Clapeyron.PCPSAFT — Type

PCPSAFTModel <: PCSAFTModel

const PPCSAFT = PCPSAFT

PCPSAFT(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

dipole: Single Parameter (Float64) - Dipole moment [D]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

dipole: Single Parameter (Float64) (optional) - Dipole moment [D]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Perturbed-Chain Polar SAFT (PCP-SAFT)

References

Gross, J., & Vrabec, J. (2005). An equation-of-state contribution for polar components: Dipolar molecules. AIChE Journal, 52(3), 856-1282. doi:10.1002/aic.10683

Clapeyron.QPCPSAFT — Type

QPCPSAFTModel <: PCPSAFTModel

QPCPSAFT(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

dipole: Single Parameter (Float64) - Dipole moment [D]

quadrupole: Single Parameter (Float64) - Quadrupole moment [DA°]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

dipole: Single Parameter (Float64) - Dipole moment [D]

quadrupole: Single Parameter (Float64) - Quadrupole moment [DA°]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Polar Perturbed-Chain SAFT, including Quadrupolar interaction contributions (Q-PCP-SAFT)

References

Gross, J. (2005). An equation-of-state contribution for polar components: Quadrupolar molecules. AIChE Journal, 51(9), 2556-2568. doi:10.1002/aic.10502

Gross, J., & Vrabec, J. (2008). Vapor−Liquid Equilibria Simulation and an Equation of State Contribution for Dipole−Quadrupole Interactions. J. Phys. Chem. B, 112(1), 51-60. doi:10.1021/jp072619u

Clapeyron.HomogcPCPSAFT — Type

HomogcPCPSAFTModel <: PCPSAFTModel

HomogcPCPSAFT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

m: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

dipole: Single Parameter (Float64) - Dipole moment [D]

k: Pair Parameter (Float64) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

dipole: Single Parameter (Float64) - Dipole moment [D]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Homosegmented Group-contribution Polar Perturbed-Chain SAFT (Homo-gc-PCP-SAFT)

References

Sauer, E., Stavrou, M., Gross, J. (2014). Comparison between a Homo- and a Heterosegmented Group Contribution Approach Based on the Perturbed-Chain Polar Statistical Associating Fluid Theory Equation of State. Industrial & Engineering Chemistry Research, 53(38), 14854-14864. doi:10.1021/ie502203w

Clapeyron.HeterogcPCPSAFT — Type

gcPCPSAFTModel <: PCSAFTModel

HeterogcPCPSAFT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

m: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

dipole: Single Parameter (Float64) - Dipole moment [D]

k: Pair Parameter (Float64) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

dipole: Single Parameter (Float64) - Dipole moment [D]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Heterosegmented Group-contribution Polar Perturbed-Chain SAFT (Hetero-gc-PCP-SAFT)

References

Gross, J., Spuhl, O., Tumakaka, F. & Sadowski, G. (2003). Modeling Copolymer Systems Using the Perturbed-Chain SAFT Equation of State. Industrial & Engineering Chemistry Research, 42, 1266-1274. doi:10.1021/ie020509y

Sauer, E., Stavrou, M. & Gross, J. (2014). Comparison between a Homo- and a Heterosegmented Group Contribution Approach Based on the Perturbed-Chain Polar Statistical Associating Fluid Theory Equation of State. Industrial & Engineering Chemistry Research, 53(38), 14854–14864. doi:10.1021/ie502203w

List of available groups

Name	Description

CH3	Methyl

CH2	Methylene

CH

C

CH2=	Terminal alkene

CH=

=C<

C#CH	Terminal alkyne

cCH2_pen	Cyclic pentane group

cCH_pen

cCH2_hex	Cyclic hexane group

cCH_hex

aCH	Aromatic group

aCH

OH	Hydroxyl group

NH2	Amine group

Clapeyron.sPCSAFT — Type

sPCSAFT <: PCSAFTModel

sPCSAFT(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Simplified Perturbed-Chain SAFT (sPC-SAFT)

References

von Solms, N., Michelsen, M. L., & Kontogeorgis, G. M. (2003). Computational and physical performance of a modified PC-SAFT equation of state for highly asymmetric and associating mixtures. Industrial & Engineering Chemistry Research, 42(5), 1098–1105. doi:10.1021/ie020753p

Clapeyron.gcsPCSAFT — Type

gcsPCSAFT <: PCSAFTModel
gcsPCSAFT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

m: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Group-contribution version of Simplified Perturbed-Chain SAFT (sPC-SAFT)

References

Tihic, A., Kontogeorgis, G.M., von Solms, N., Michelsen, M.L. (2008). A predictive group-contribution simplified PC-SAFT equation of state: Application to polymer systems. Industrial & Engineering Chemistry Research, 47(15), 5092-5101. doi:10.1021/ie0710768

Clapeyron.pharmaPCSAFT — Type

pharmaPCSAFTModel <: PCSAFTModel

pharmaPCSAFT(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Constant binary Interaction Paramater (no units)

kT: Pair Parameter (Float64) - T-dependent inary Interaction Paramater [K^-1]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

k: Pair Parameter (Float64) (optional) - Constant binary Interaction Paramater (no units)

kT: Pair Parameter (Float64) - T-dependent inary Interaction Paramater [K^-1]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Perturbed-Chain SAFT (PC-SAFT), with T dependent kij and water correlation [2] for segment diameter. For using the water's sigma correlation, water08 should be selected instead of water.

References

Paus, R., Ji, Y., Vahle, L., & Sadowski, G. (2015). Predicting the solubility advantage of amorphous pharmaceuticals: A novel thermodynamic approach. Molecular Pharmaceutics, 12(8), 2823–2833. doi:10.1021/mp500824d

Cameretti, L. F., & Sadowski, G. (2008). Modeling of aqueous amino acid and polypeptide solutions with PC-SAFT. Genie Des Procedes [Chemical Engineering and Processing], 47(6), 1018–1025. doi:10.1016/j.cep.2007.02.034

Clapeyron.CPPCSAFT — Type

CPPCSAFTModel <: PCSAFTModel

CPPCSAFT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

delta: Single Parameter (Float64) - Critical volume displacement (no units)

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Pair Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

delta: Single Parameter (Float64) - Critical volume displacement (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Critical Point-Based PC-SAFT (CP-PC-SAFT)

References

Polishuk, I. (2014). Standardized critical point-based numerical solution of statistical association fluid theory parameters: The perturbed chain-statistical association fluid theory equation of state revisited. Industrial & Engineering Chemistry Research, 53(36), 14127–14141. doi:10.1021/ie502633e

Clapeyron.GEPCSAFT — Type

GEPCSAFT <: SAFTModel

GEPCSAFT(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

activity: Activity model

Description

Perturbed-Chain SAFT (PC-SAFT), with Gᴱ mixing rule.

References

Walker, P. J. (2022). Toward advanced, predictive mixing rules in SAFT equations of state. Industrial & Engineering Chemistry Research. doi:10.1021/acs.iecr.2c03464

Clapeyron.ADPCSAFT — Type

ADPCSAFTModel <: SAFTModel
ADPCSAFT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

r_c: Single Parameter (Float64)

c1: Single Parameter (Float64)

c2: Single Parameter (Float64)

c3: Single Parameter (Float64)

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

theta_c: Association Parameter (Float64)

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

r_c: Single Parameter (Float64)

c1: Single Parameter (Float64)

c2: Single Parameter (Float64)

c3: Single Parameter (Float64)

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

theta_c: Association Parameter (Float64)

Input models

idealmodel: Ideal Model

Description

modified Perturbed-Chain SAFT (PC-SAFT) with an association dependent hard sphere diameter.Currently only works for water.

numerically unstable

Due to its functional form, DAPT is numerically unstable. Please use big Floats for most calculations.

References

Marshall, B. D. (2021). A modified perturbed chain‐statistical associating fluid theory equation of state for water which includes an association dependent hard sphere diameter. AIChE Journal. American Institute of Chemical Engineers, 67(10). doi:10.1002/aic.17342

Clapeyron.LJSAFT — Type

LJSAFTModel <: SAFTModel

LJSAFT(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

b: Single Parameter (Float64) - Segment Volume [dm^3/mol]

T_tilde: Single Parameter (Float64) - Lennard-Jones attraction parameter [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater for energy(no units)

zeta: Pair Parameter (Float64) - Binary Interaction Paramater for volume (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

b: Pair Parameter (Float64) - Mixed segment covolume [dm^3/mol]

T_tilde: Pair Parameter (Float64) - Mixed Lennard-Jones attraction parameter [K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Perturbed-Chain SAFT (PC-SAFT)

References

Kraska, T., & Gubbins, K. E. (1996). Phase equilibria calculations with a modified SAFT equation of state. 1. Pure alkanes, alkanols, and water. Industrial & Engineering Chemistry Research, 35(12), 4727–4737. doi:10.1021/ie9602320

Clapeyron.softSAFT — Type

softSAFTModel <: SAFTModel

softSAFT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Soft SAFT, with Lennard-Jones function from Johnson et al. (1993)

References

Johnson, J. K., Zollweg, J. A., & Gubbins, K. E. (1993). The Lennard-Jones equation of state revisited. Molecular physics, 78(3), 591–618. doi:10.1080/00268979300100411

FELIPE J. BLAS and LOURDES F. VEGA. (1997). Thermodynamic behaviour of homonuclear and heteronuclear Lennard-Jones chains with association sites from simulation and theory. Molecular physics, 92(1), 135–150. doi:10.1080/002689797170707

Clapeyron.softSAFT2016 — Type

softSAFT2016Model <: softSAFTModel

softSAFT2016(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Soft SAFT, with Lennard-Jones function from Thol et al. (2016)

References

FELIPE J. BLAS and LOURDES F. VEGA. (1997). Thermodynamic behaviour of homonuclear and heteronuclear Lennard-Jones chains with association sites from simulation and theory. Molecular physics, 92(1), 135–150. doi:10.1080/002689797170707

Thol, M., Rutkai, G., Köster, A., Lustig, R., Span, R., & Vrabec, J. (2016). Equation of state for the Lennard-Jones fluid. Journal of physical and chemical reference data, 45(2), 023101. doi:10.1063/1.4945000

Clapeyron.solidsoftSAFT — Type

solidsoftSAFTModel <: SAFTModel

solidsoftSAFT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Soft SAFT equation of state for the solid phase.

References

Johnson, J. K., Zollweg, J. A., & Gubbins, K. E. (1993). Modelling the solid–liquid–vapour phase behaviour ofn-alkanes in a TPT-1framework. Molecular physics, 78(3), 591–618. doi:10.1080/00268979300100411

FELIPE J. BLAS and LOURDES F. VEGA. (1997). Thermodynamic behaviour of homonuclear and heteronuclear Lennard-Jones chains with association sites from simulation and theory. Molecular physics, 92(1), 135–150. doi:10.1080/002689797170707

Ramírez-Carpio, V., Galindo, A., & Gil-Villegas, A. (2023). Modelling the solid–liquid–vapour phase behaviour of n -alkanes in a TPT-1 framework. Molecular Physics, 121(19–20). doi:10.1080/00268976.2023.2204150

Clapeyron.SAFTVRSW — Type

SAFTVRSWModel <: SAFTModel

SAFTVRSW(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

lambda: Single Parameter (Float64) - Soft Well range parameter (no units)

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

l: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

lambda: Pair Parameter (Float64) - Mixed Soft Well range parameter (no units)

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

SAFT, Variable Range (VR) ,Square Well (SW)

References

Gil-Villegas, A., Galindo, A., Whitehead, P. J., Mills, S. J., Jackson, G., & Burgess, A. N. (1997). Statistical associating fluid theory for chain molecules with attractive potentials of variable range. The Journal of chemical physics, 106(10), 4168–4186. doi:10.1063/1.473101

Clapeyron.SAFTVRMie — Type

SAFTVRMieModel <: SAFTModel

SAFTVRMie(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

SAFT-VR with Mie potential

References

Lafitte, T., Apostolakou, A., Avendaño, C., Galindo, A., Adjiman, C. S., Müller, E. A., & Jackson, G. (2013). Accurate statistical associating fluid theory for chain molecules formed from Mie segments. The Journal of Chemical Physics, 139(15), 154504. doi:10.1063/1.4819786

Dufal, S., Lafitte, T., Haslam, A. J., Galindo, A., Clark, G. N. I., Vega, C., & Jackson, G. (2015). The A in SAFT: developing the contribution of association to the Helmholtz free energy within a Wertheim TPT1 treatment of generic Mie fluids. Molecular Physics, 113(9–10), 948–984. doi:10.1080/00268976.2015.1029027

Clapeyron.SAFTVRMie15 — Type

SAFTVRMie15Model <: SAFTVRMieModel

SAFTVRMie(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

SAFT-VR with Mie potential and the Mie association kernel

References

Lafitte, T., Apostolakou, A., Avendaño, C., Galindo, A., Adjiman, C. S., Müller, E. A., & Jackson, G. (2013). Accurate statistical associating fluid theory for chain molecules formed from Mie segments. The Journal of Chemical Physics, 139(15), 154504. doi:10.1063/1.4819786

Dufal, S., Lafitte, T., Haslam, A. J., Galindo, A., Clark, G. N. I., Vega, C., & Jackson, G. (2015). The A in SAFT: developing the contribution of association to the Helmholtz free energy within a Wertheim TPT1 treatment of generic Mie fluids. Molecular Physics, 113(9–10), 948–984. doi:10.1080/00268976.2015.1029027

Clapeyron.SAFTVRQMie — Type

SAFTVRQMieModel <: SAFTVRMieModel

SAFTVRQMie(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

Model Parameters

Mw: Pair Parameter (Float64) - Mixed Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

Input models

idealmodel: Ideal Model

Description

Quantum-Corrected SAFT-VR Mie. In particular, it uses the second order Feynman–Hibbs corrections to the Mie Potential

References

Aasen, A., Hammer, M., Ervik, Å., Müller, E. A., & Wilhelmsen, Ø. (2019). Equation of state and force fields for Feynman–Hibbs-corrected Mie fluids. I. Application to pure helium, neon, hydrogen, and deuterium. The Journal of Chemical Physics, 151(6), 064508. doi:10.1063/1.5111364

Aasen, A., Hammer, M., Müller, E. A., & Wilhelmsen, Ø. (2020). Equation of state and force fields for Feynman-Hibbs-corrected Mie fluids. II. Application to mixtures of helium, neon, hydrogen, and deuterium. The Journal of Chemical Physics, 152(7), 074507. doi:10.1063/1.5136079

Clapeyron.SAFTVRSMie — Type

SAFTVRSMieModel <: SAFTModel

SAFTVRSMie(components;
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

SAFT-VR with Mie potential for the solid phase using WCA perturbation theory.

References

Jalani, Y., Ramrattana, N., Walker, P.J., Riedemann, A., Galindo, A., Mater, O. K., & Müller, E. A. (2024). SAFT-VR Mie Equation of State for the Solid and Fluid Phases. (in preparation)

Clapeyron.SAFTgammaMie — Type

SAFTgammaMie <: SAFTModel

SAFTgammaMie(components; idealmodel = BasicIdeal, userlocations = String[], groupuserlocations = String[], idealuserlocations = String[], epsilonmixing = :default, referencestate = nothing, verbose = false, assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

vst: Single Parameter (Float64) - Number of segments (no units)

S: Single Parameter (Float64) - Shape factor for segment (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Model Parameters

segment: Single Parameter (Float64) - Number of segments (no units)

shapefactor: Single Parameter (Float64) - Shape factor for segment (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

SAFT-γ-Mie EoS

Info

You can choose between the Hudsen-McCoubrey combining rule (√(ϵᵢ*ϵⱼ)*(σᵢ^3 * σⱼ^3)/σᵢⱼ^6) or the default rule (√(ϵᵢ*ϵⱼ*(σᵢ^3 * σⱼ^3))/σᵢⱼ^3) by passing the epsilon_mixing argument. with arguments :default or :hudsen_mccoubrey

References

Papaioannou, V., Lafitte, T., Avendaño, C., Adjiman, C. S., Jackson, G., Müller, E. A., & Galindo, A. (2014). Group contribution methodology based on the statistical associating fluid theory for heteronuclear molecules formed from Mie segments. The Journal of Chemical Physics, 140(5), 054107. doi:10.1063/1.4851455

Dufal, S., Papaioannou, V., Sadeqzadeh, M., Pogiatzis, T., Chremos, A., Adjiman, C. S., … Galindo, A. (2014). Prediction of thermodynamic properties and phase behavior of fluids and mixtures with the SAFT-γ Mie group-contribution equation of state. Journal of Chemical and Engineering Data, 59(10), 3272–3288. doi:10.1021/je500248h

List of available groups

Name	Description

CH3	Methyl

CH2	Methylene

CH

C

aCH	Aromatic CH

aCCH2

aCCH

CH2=	Alkene group

CH=

cCH2	Cyclic alkane group

COOH	Carboxylic acid group

COO	Ester group

OH	Hydroxyl

CH2OH	Methylene hydroxyl group

CHOH

NH2	Amine group

NH

N

cNH

cN

CH=

aCCH3

aCOH

cCH

cCHNH

cCHN

aCCOaC

aCCOOH

aCNHaC

CH3CO

eO	End ether oxygen

cO	Center ether oxygen

Clapeyron.structSAFTgammaMie — Type

structSAFTgammaMie <: SAFTgammaMieModel

structSAFTgammaMie(components; 
idealmodel = BasicIdeal,
userlocations = String[],
group_userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
epsilon_mixing = :default,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

vst: Single Parameter (Float64) - Number of segments (no units)

S: Single Parameter (Float64) - Shape factor for segment (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Model Parameters

segment: Single Parameter (Float64) - Number of segments (no units)

shapefactor: Single Parameter (Float64) - Shape factor for segment (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

lambda_a: Pair Parameter (Float64) - Atractive range parameter (no units)

lambda_r: Pair Parameter (Float64) - Repulsive range parameter (no units)

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

s-SAFT-γ-Mie EoS

References

Shaahmadi,, F., Hurter, R.M., Burger, A.J., Cripwell, J.T. (2021). Improving the SAFT-γ Mie equation of state to account for functional group interactions in a structural (s-SAFT-γ Mie) framework: Linear and branched alkanes. The Journal of Chemical Physics, 154, 244102. doi:10.1063/5.0048315

Schulze-Hulbe, A., Shaahmadi, F., Burger, A.J., Cripwell, J.T. (2022). Extending the Structural (s)-SAFT-γ Mie Equation of State to Primary Alcohols. Industrial & Engineering Chemistry Research, 61 (33), 12208-12228. doi:10.1021/acs.iecr.2c00198

Clapeyron.DAPT — Type

DAPTModel <: SAFTModel
DAPT(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

r_c: Single Parameter (Float64)

lambda: Single Parameter (Float64)

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

theta_c: Association Parameter (Float64)

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

r_c: Single Parameter (Float64)

lambda: Single Parameter (Float64)

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

theta_c: Association Parameter (Float64)

Input models

idealmodel: Ideal Model

Description

Doubly-Associated Perturbation Theory model. Currently only works for water.

numerically unstable

Due to its functional form, DAPT is numerically unstable. Please use big Floats for most calculations.

References

Marshall, B. D. (2019). A doubly associated reference perturbation theory for water. Fluid Phase Equilibria, 500(112252), 112252. doi:10.1016/j.fluid.2019.112252

CPA Models

Clapeyron.CPA — Type

CPAModel <: EoSModel

function CPA(components;
    radial_dist::Symbol = :CS,
    idealmodel = BasicIdeal,
    cubicmodel = RK,
    alpha = sCPAAlpha,
    mixing = vdW1fRule,
    activity = nothing,
    translation = NoTranslation,
    userlocations = String[],
    ideal_userlocations = String[],
    alpha_userlocations = String[],
    activity_userlocations = String[],
    mixing_userlocations = String[],
    translation_userlocations = String[],
    reference_state = nothing,
    verbose = false,
    assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

Tc: Single Parameter (Float64) - Critical Temperature [K]

a: Single Parameter (Float64) - Atraction parameter [m^6*Pa/mol]

b: Single Parameter (Float64) - Covolume [m^3/mol]

c1: Single Parameter (Float64) - α-function constant Parameter (no units)

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

l: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

Tc: Single Parameter (Float64) - Critical Temperature [K]

a: Pair Parameter (Float64) - Mixed Atraction Parameter [m^6*Pa/mol]

b: Pair Parameter (Float64) - Mixed Covolume [m^3/mol]

c1: Single Parameter (Float64) - α-function constant Parameter (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [J]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Input models

idealmodel: Ideal Model

cubicmodel: Cubic Model

Description

Cubic Plus Association (CPA) EoS. Consists in the addition of a cubic part and an association part:

a_res(model::CPA) = a_res(model::Cubic) + a_assoc(model)

The radial_dist argument can be used to choose between a Carnahan-Starling form (CS, default) or the Kontogeorgis (KG) term, more widely known as s-CPA.

References

Kontogeorgis, G. M., Michelsen, M. L., Folas, G. K., Derawi, S., von Solms, N., & Stenby, E. H. (2006). Ten years with the CPA (cubic-plus-association) equation of state. Part 1. Pure compounds and self-associating systems. Industrial & Engineering Chemistry Research, 45(14), 4855–4868. doi:10.1021/ie051305v

Clapeyron.sCPA — Function

sCPAModel <: CPAModel

function sCPA(components;
    idealmodel = BasicIdeal,
    radial_dist::Symbol = :KG,
    cubicmodel = RK,
    alpha = sCPAAlpha,
    mixing = vdW1fRule,
    activity = nothing,
    translation = NoTranslation,
    userlocations = String[],
    ideal_userlocations = String[],
    alpha_userlocations = String[],
    activity_userlocations = String[],
    mixing_userlocations = String[],
    translation_userlocations = String[],
    reference_state = nothing,
    verbose = false,
    assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

Tc: Single Parameter (Float64) - Critical Temperature [K]

a: Single Parameter (Float64) - Atraction parameter [m^6*Pa/mol]

b: Single Parameter (Float64) - Covolume [m^3/mol]

c1: Single Parameter (Float64) - α-function constant Parameter (no units)

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

a: Pair Parameter (Float64) - Mixed Atraction Parameter [m^6*Pa/mol]

b: Pair Parameter (Float64) - Mixed Covolume [m^3/mol]

c1: Single Parameter (Float64) - α-function constant Parameter (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [J]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Input models

idealmodel: Ideal Model

cubicmodel: Cubic Model

Description

Simplified Cubic Plus Association (s-CPA) EoS. Consists in the addition of a cubic part and an association part:

a_res(model::CPA) = a_res(model::Cubic) + a_assoc(model)

The radial_dist argument can be used to choose between a Carnahan-Starling form (CS, default) or the Kontogeorgis (KG) term, more widely known as s-CPA. using sCPA(components, radial_dist =: CS) is equivalent to using the original CPA

References

Kontogeorgis, G. M., Michelsen, M. L., Folas, G. K., Derawi, S., von Solms, N., & Stenby, E. H. (2006). Ten years with the CPA (cubic-plus-association) equation of state. Part 1. Pure compounds and self-associating systems. Industrial & Engineering Chemistry Research, 45(14), 4855–4868. doi:10.1021/ie051305v

Miscellaneous Models

Clapeyron.COFFEE — Type

COFFEEModel <: SAFTVRMieModel

COFFEE(components; 
idealmodel = BasicIdeal,
userlocations = String[],
ideal_userlocations = String[],
reference_state = nothing,
verbose = false,
assoc_options = AssocOptions())

Input parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Single Parameter (Float64) - Segment Diameter [A°]

epsilon: Single Parameter (Float64) - Reduced dispersion energy [K]

k: Pair Parameter (Float64) (optional) - Binary Interaction Paramater (no units)

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume [m^3]

Model Parameters

Mw: Single Parameter (Float64) - Molecular Weight [g/mol]

segment: Single Parameter (Float64) - Number of segments (no units)

sigma: Pair Parameter (Float64) - Mixed segment Diameter [m]

epsilon: Pair Parameter (Float64) - Mixed reduced dispersion energy[K]

epsilon_assoc: Association Parameter (Float64) - Reduced association energy [K]

bondvol: Association Parameter (Float64) - Association Volume

Input models

idealmodel: Ideal Model

Description

Co-Oriented Fluid Functional Equation for Electrostatic interactions (COFFEE)

References

Langenbach, K. (2017). Co-Oriented Fluid Functional Equation for Electrostatic interactions (COFFEE). Chemical Engineering Science, 174, 40-55 (2017). doi:10.1016/j.ces.2017.08.025