Contents

Contents

Index

Multi component properties

Consistency and Stability

TP Flash

Index

Clapeyron.ActivityBubblePressure

Clapeyron.ActivityBubbleTemperature

Clapeyron.ActivityDewPressure

Clapeyron.ActivityDewTemperature

Clapeyron.ChemPotBubblePressure

Clapeyron.ChemPotBubbleTemperature

Clapeyron.ChemPotDewPressure

Clapeyron.ChemPotDewTemperature

Clapeyron.DETPFlash

Clapeyron.FugBubblePressure

Clapeyron.FugBubbleTemperature

Clapeyron.FugDewPressure

Clapeyron.FugDewTemperature

Clapeyron.MichelsenTPFlash

Clapeyron.RRTPFlash

Clapeyron.LLE_pressure

Clapeyron.LLE_temperature

Clapeyron.UCEP_mix

Clapeyron.UCST_mix

Clapeyron.VLLE_pressure

Clapeyron.VLLE_temperature

Clapeyron.azeotrope_pressure

Clapeyron.azeotrope_temperature

Clapeyron.bubble_pressure

Clapeyron.bubble_temperature

Clapeyron.crit_mix

Clapeyron.dew_pressure

Clapeyron.dew_temperature

Clapeyron.diffusive_stability

Clapeyron.gibbs_duhem

Clapeyron.gibbs_solvation

Clapeyron.isstable

Clapeyron.mechanical_stability

Clapeyron.numphases

Clapeyron.tp_flash

Clapeyron.tpd

Multi component properties

Clapeyron.bubble_pressure — Function

bubble_pressure(model::EoSModel, T, x, method = ChemPotBubblePressure())

Calculates the bubble pressure and properties at a given temperature. Returns a tuple, containing:

Bubble Pressure [Pa]

liquid volume at Bubble Point [m³]

vapour volume at Bubble Point [m³]

Gas composition at Bubble Point

By default, uses equality of chemical potentials, via ChemPotBubblePressure

Clapeyron.bubble_temperature — Function

bubble_temperature(model::EoSModel, p, x,method::BubblePointMethod = ChemPotBubbleTemperature())

Calculates the bubble temperature and properties at a given pressure. Returns a tuple, containing:

Bubble Temperature [K]

liquid volume at Bubble Point [m³]

vapour volume at Bubble Point [m³]

Gas composition at Bubble Point

By default, uses equality of chemical potentials, via ChemPotBubbleTemperature

Clapeyron.dew_pressure — Function

dew_pressure(model::EoSModel, T, y,method = ChemPotDewPressure())

Calculates the dew pressure and properties at a given temperature. Returns a tuple, containing:

Dew Pressure [Pa]

liquid volume at Dew Point [m³]

vapour volume at Dew Point [m³]

Liquid composition at Dew Point

By default, uses equality of chemical potentials, via ChemPotDewPressure

Clapeyron.dew_temperature — Function

dew_temperature(model::EoSModel, p, y, method = ChemPotDewTemperature())

calculates the dew temperature and properties at a given pressure. Returns a tuple, containing:

Dew Temperature [K]

liquid volume at Dew Point [m³]

vapour volume at Dew Point [m³]

Liquid composition at Dew Point

By default, uses equality of chemical potentials, via ChemPotDewTemperature

Clapeyron.azeotrope_pressure — Function

azeotrope_pressure(model::EoSModel, T; v0 = x0_azeotrope_pressure(model,T))

calculates the azeotrope pressure and properties at a given temperature. Returns a tuple, containing:

Azeotrope Pressure [Pa]

liquid volume at Azeotrope Point [m³]

vapour volume at Azeotrope Point [m³]

Azeotrope composition

Clapeyron.azeotrope_temperature — Function

azeotrope_temperature(model::EoSModel, T; v0 = x0_bubble_pressure(model,T,[0.5,0.5]))

Calculates the azeotrope temperature and properties at a given pressure. Returns a tuple, containing:

Azeotrope Temperature [K]

liquid volume at Azeotrope Point [m³]

vapour volume at Azeotrope Point [m³]

Azeotrope composition

Clapeyron.LLE_pressure — Function

LLE_pressure(model::EoSModel, T, x; v0 = x0_LLE_pressure(model,T,x))

calculates the Liquid-Liquid equilibrium pressure and properties at a given temperature.

Returns a tuple, containing:

LLE Pressure [Pa]

liquid volume of composition x₁ = x at LLE Point [m³]

liquid volume of composition x₂ at LLE Point [m³]

Liquid composition x₂

Clapeyron.LLE_temperature — Function

LLE_temperature(model::EoSModel, p, x; T0 = x0_LLE_temperature(model,p,x))

calculates the Liquid-Liquid equilibrium temperature and properties at a given pressure.

Returns a tuple, containing:

LLE Pressure [Pa]

liquid volume of composition x₁ = x at LLE Point [m³]

liquid volume of composition x₂ at LLE Point [m³]

Liquid composition x₂

Clapeyron.VLLE_pressure — Function

VLLE_pressure(model::EoSModel, T; v0 = x0_LLE_pressure(model,T))

calculates the Vapor-Liquid-Liquid equilibrium pressure and properties of a binary mixture at a given temperature.

Returns a tuple, containing:

VLLE Pressure [Pa]

Liquid volume of composition x₁ at VLLE Point [m³]

Liquid volume of composition x₂ at VLLE Point [m³]

Vapour volume of composition y at VLLE Point [m³]

Liquid composition x₁

Liquid composition x₂

Liquid composition y

Clapeyron.VLLE_temperature — Function

VLLE_temperature(model::EoSModel, p; T0 = x0_LLE_temperature(model,p))

calculates the Vapor-Liquid-Liquid equilibrium temperature and properties of a binary mixture at a given temperature.

Returns a tuple, containing:

VLLE temperature [K]

Liquid volume of composition x₁ at VLLE Point [m³]

Liquid volume of composition x₂ at VLLE Point [m³]

Vapour volume of composition y at VLLE Point [m³]

Liquid composition x₁

Liquid composition x₂

Liquid composition y

Clapeyron.crit_mix — Function

crit_mix(model::EoSModel,z;v0=x=x0_crit_mix(model,z))

Returns the critical mixture point at a ginven composition.

Returns a tuple, containing:

Critical Mixture Temperature [K]

Critical Mixture Pressure [Pa]

Critical Mixture Volume [m³]

Clapeyron.UCEP_mix — Function

UCEP_mix(model::EoSModel;v0=x0_UCEP_mix(model))

Calculates the Upper Critical End Point of a binary mixture.

returns:

UCEP Temperature [K]

UCEP Pressure [Pa]

liquid volume at UCEP Point [m³]

vapour volume at UCEP Point [m³]

liquid molar composition at UCEP Point

vapour molar composition at UCEP Point

Clapeyron.UCST_mix — Function

UCST_mix(model::EoSModel,T;v0=x0_UCST_mix(model,T))

Calculates the Upper critical solution point of a mixture at a given Temperature.

returns:

UCST Pressure [Pa]

volume at UCST Point [m³]

molar composition at UCST Point

Clapeyron.gibbs_solvation — Function

gibbs_solvation(model::EoSModel, T)

Calculates the solvation free energy as:

g_solv = -R̄*T*log(K)

where the first component is the solvent and second is the solute.

Bubble/Dew methods

Clapeyron.ChemPotBubblePressure — Type

ChemPotBubblePressure(kwargs...)

Function to compute bubble_pressure via chemical potentials. It directly solves the equality of chemical potentials system of equations.

Inputs:

y0 = nothing: optional, initial guess for the vapor phase composition

p0 = nothing: optional, initial guess for the bubble pressure [Pa]

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

atol = 1e-8: optional, absolute tolerance of the non linear system of equations

rtol = 1e-12: optional, relative tolerance of the non linear system of equations

max_iters = 10000: optional, maximum number of iterations

nonvolatiles = nothing: optional, Vector of strings containing non volatile compounds. those will be set to zero on the vapour phase.

Clapeyron.FugBubblePressure — Type

FugBubblePressure(kwargs...)

Function to compute bubble_pressure via fugacity coefficients. First it uses successive substitution to update the phase composition and a outer newtown loop to update the pressure. If no convergence is reached after itmax_newton iterations, the system is solved using a multidimensional non-linear system of equations.

Inputs:

y0 = nothing: optional, initial guess for the vapor phase composition

p0 = nothing: optional, initial guess for the bubble pressure [Pa]

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

itmax_newton = 10: optional, number of iterations to update the pressure using newton's method

itmax_ss = 5: optional, number of iterations to update the liquid phase composition using successive substitution

tol_x = 1e-8: optional, tolerance to stop successive substitution cycle

tol_p = 1e-8: optional, tolerance to stop newton cycle

tol_of = 1e-8: optional, tolerance to check if the objective function is zero.

nonvolatiles = nothing: optional, Vector of strings containing non volatile compounds. those will be set to zero on the vapour phase.

Clapeyron.ActivityBubblePressure — Type

ActivityBubblePressure(kwargs...)

Function to compute bubble_pressure using Activity Coefficients. On activity coefficient models it solves the problem via succesive substitucion. On helmholtz-based models, it uses the Chapman approximation for activity coefficients.

Inputs:

gas_fug = true: if the solver uses gas fugacity coefficients. on ActivityModel is set by default to false

poynting = true: if the solver use the poynting correction on the liquid fugacity coefficients. on ActivityModel is set by default to false

y0 = nothing: optional, initial guess for the vapor phase composition

p0 = nothing: optional, initial guess for the bubble pressure [Pa]

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

atol = 1e-8: optional, absolute tolerance of the non linear system of equations

rtol = 1e-12: optional, relative tolerance of the non linear system of equations

itmax_ss = 40: optional, maximum number of sucesive substitution iterations

Clapeyron.ChemPotBubbleTemperature — Type

ChemPotBubbleTemperature(kwargs...)

Function to compute bubble_temperature via chemical potentials. It directly solves the equality of chemical potentials system of equations.

Inputs:

y = nothing: optional, initial guess for the vapor phase composition.

T0 = nothing: optional, initial guess for the bubble temperature [K].

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

atol = 1e-8: optional, absolute tolerance of the non linear system of equations

rtol = 1e-12: optional, relative tolerance of the non linear system of equations

max_iters = 10000: optional, maximum number of iterations

nonvolatiles = nothing: optional, Vector of strings containing non volatile compounds. those will be set to zero on the vapour phase.

Clapeyron.FugBubbleTemperature — Type

FugBubbleTemperature(kwargs...)

Method to compute bubble_temperature via fugacity coefficients. First it uses successive substitution to update the phase composition and a outer newtown loop to update the temperature. If no convergence is reached after itmax_newton iterations, the system is solved using a multidimensional non-linear system of equations.

Inputs:

y = nothing: optional, initial guess for the vapor phase composition.

T0 = nothing: optional, initial guess for the bubble temperature [K].

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

itmax_newton = 10: optional, number of iterations to update the temperature using newton's method

itmax_ss = 5: optional, number of iterations to update the liquid phase composition using successive substitution

tol_x = 1e-8: optional, tolerance to stop successive substitution cycle

tol_T = 1e-8: optional, tolerance to stop newton cycle

tol_of = 1e-8: optional, tolerance to check if the objective function is zero.

nonvolatiles: optional, Vector of strings containing non volatile compounds. those will be set to zero on the vapour phase.

Clapeyron.ActivityBubbleTemperature — Type

ActivityBubbleTemperature(kwargs...)

Function to compute bubble_temperature using Activity Coefficients. On activity coefficient models it solves the problem via succesive substitucion. On helmholtz-based models, it uses the Chapman approximation for activity coefficients.

Inputs:

gas_fug = true: if the solver uses gas fugacity coefficients. on ActivityModel is set by default to false

poynting = true: if the solver use the poynting correction on the liquid fugacity coefficients. on ActivityModel is set by default to false

y0 = nothing: optional, initial guess for the vapor phase composition

T0 = nothing: optional, initial guess for the bubble temperature [K]

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

atol = 1e-8: optional, absolute tolerance of the non linear system of equations

rtol = 1e-12: optional, relative tolerance of the non linear system of equations

itmax_ss = 40: optional, maximum number of sucesive substitution iterations

Clapeyron.ChemPotDewPressure — Type

ChemPotDewPressure(kwargs...)

Function to compute dew_pressure via chemical potentials. It directly solves the equality of chemical potentials system of equations.

Inputs:

x0 = nothing: optional, initial guess for the liquid phase composition

p0 = nothing: optional, initial guess for the dew pressure [Pa]

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

atol = 1e-8: optional, absolute tolerance of the non linear system of equations

rtol = 1e-12: optional, relative tolerance of the non linear system of equations

max_iters = 10000: optional, maximum number of iterations

noncondensables = nothing: optional, Vector of strings containing non condensable compounds. those will be set to zero on the liquid phase.

Clapeyron.FugDewPressure — Type

FugDewPressure(kwargs...)

Method to compute dew_pressure via fugacity coefficients. First it uses successive substitution to update the phase composition and a outer newtown loop to update the pressure. If no convergence is reached after itmax_newton iterations, the system is solved using a multidimensional non-linear system of equations.

Inputs:

x0 = nothing: optional, initial guess for the liquid phase composition

p0 = nothing: optional, initial guess for the dew pressure [Pa]

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

itmax_newton = 10: optional, number of iterations to update the pressure using newton's method

itmax_ss = 5: optional, number of iterations to update the liquid phase composition using successive substitution

tol_x = 1e-8: optional, tolerance to stop successive substitution cycle

tol_p = 1e-8: optional, tolerance to stop newton cycle

tol_of = 1e-8: optional, tolerance to check if the objective function is zero.

noncondensables = nothing: optional, Vector of strings containing non condensable compounds. those will be set to zero on the liquid phase.

Clapeyron.ActivityDewPressure — Type

ActivityDewPressure(kwargs...)

Function to compute dew_pressure using Activity Coefficients. On activity coefficient models it solves the problem via succesive substitucion. On helmholtz-based models, it uses the Chapman approximation for activity coefficients.

Inputs:

gas_fug = true: if the solver uses gas fugacity coefficients. on ActivityModel is set by default to false

poynting = true: if the solver use the poynting correction on the liquid fugacity coefficients. on ActivityModel is set by default to false

x0 = nothing: optional, initial guess for the liquid phase composition

p0 = nothing: optional, initial guess for the dew pressure [Pa]

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

atol = 1e-8: optional, absolute tolerance of the non linear system of equations

rtol = 1e-12: optional, relative tolerance of the non linear system of equations

itmax_ss = 40: optional, maximum number of sucesive substitution iterations

Clapeyron.ChemPotDewTemperature — Type

ChemPotDewTemperature(kwargs...)

Function to compute temperature via chemical potentials. It directly solves the equality of chemical potentials system of equations.

Inputs:

x0 = nothing: optional, initial guess for the liquid phase composition

T0 =nothing: optional, initial guess for the dew temperature [K]

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

atol = 1e-8: optional, absolute tolerance of the non linear system of equations

rtol = 1e-12: optional, relative tolerance of the non linear system of equations

max_iters = 10000: optional, maximum number of iterations

noncondensables = nothing: optional, Vector of strings containing non condensable compounds. those will be set to zero on the liquid phase.

Clapeyron.FugDewTemperature — Type

FugDewTemperature(kwargs...)

Method to compute dew_temperature via fugacity coefficients. First it uses successive substitution to update the phase composition and a outer newtown loop to update the temperature. If no convergence is reached after itmax_newton iterations, the system is solved using a multidimensional non-linear system of equations.

Inputs:

x0 = nothing: optional, initial guess for the liquid phase composition

T0 = nothing: optional, initial guess for the dew temperature [K]

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

itmax_newton = 10: optional, number of iterations to update the temperature using newton's method

itmax_ss = 5: optional, number of iterations to update the liquid phase composition using successive substitution

tol_x = 1e-8: optional, tolerance to stop successive substitution cycle

tol_T = 1e-8: optional, tolerance to stop newton cycle

tol_of = 1e-8: optional, tolerance to check if the objective function is zero.

noncondensables = nothing: optional, Vector of strings containing non condensable compounds. those will be set to zero on the liquid phase.

Clapeyron.ActivityDewTemperature — Type

ActivityDewTemperature(kwargs...)

Function to compute dew_temperature using Activity Coefficients. On activity coefficient models it solves the problem via succesive substitucion. On helmholtz-based models, it uses the Chapman approximation for activity coefficients.

Inputs:

gas_fug = true: if the solver uses gas fugacity coefficients. on ActivityModel is set by default to false

poynting = true: if the solver use the poynting correction on the liquid fugacity coefficients. on ActivityModel is set by default to false

x0 = nothing: optional, initial guess for the liquid phase composition

T0 = nothing: optional, initial guess for the dew temperature [K]

vol0 = nothing: optional, initial guesses for the liquid and vapor phase volumes

atol = 1e-8: optional, absolute tolerance of the non linear system of equations

rtol = 1e-12: optional, relative tolerance of the non linear system of equations

itmax_ss = 40: optional, maximum number of sucesive substitution iterations

Consistency and Stability

Clapeyron.gibbs_duhem — Function

gibbs_duhem(model,V,T,z=[1.0])

performs a Gibbs-Duhem check on the input conditions:

∑zᵢμᵢ - G ≈ 0

Where G is the total gibbs energy. it can help diagnose if a user-defined eos is consistent.

return |∑zᵢμᵢ - G|, ∑zᵢμᵢ and G at the specified conditions.

Clapeyron.isstable — Function

isstable(model,V,T,z)::Bool

Performs stability tests for a (V,T,z) pair, and warn if any tests fail. returns true/false.

Checks:

mechanical stability: isothermal compressibility is not negative.

diffusive stability: all eigenvalues of ∂²A/∂n² are positive.

Clapeyron.mechanical_stability — Function

mechanical_stability(model,V,T,z)::Bool

Performs a mechanical stability for a (V,T,z) pair, returns true/false. Checks if isothermal compressibility is not negative.

Clapeyron.diffusive_stability — Function

diffusive_stability(model,V,T,z)::Bool

Performs a diffusive stability for a (V,T,z) pair, returns true/false. Checks if all eigenvalues of ∂²A/∂n² are positive.

Clapeyron.tpd — Function

tpd(model,p,T,z;verbose=false)

Calculates the Tangent plane distance function (tpd). It returns:

a vector with trial phase compositions where tpd < 0

a vector with the tpd values

a vector with symbols indicating the phase of the input composition

a vector with symbols indicating the phase of the trial composition

It iterates over each two-phase combination, starting from pure trial compositions, it does succesive substitution, then Gibbs optimization.

If the vectors are empty, then the procedure couldn't find a negative tpd. That is an indication that the phase is (almost) surely stable.

TP Flash

Clapeyron.tp_flash — Function

tp_flash(model, p, T, n, method::TPFlashMethod = DETPFlash())

Routine to solve non-reactive multicomponent flash problem. The default method uses Global Optimization. see DETPFlash

Inputs:

T, Temperature

p, Pressure

n, vector of number of moles of each species

Outputs - Tuple containing:

xᵢⱼ, Array of mole fractions of species j in phase i

nᵢⱼ, Array of mole numbers of species j in phase i, [mol]

G, Gibbs Free Energy of Equilibrium Mixture [J]

Clapeyron.DETPFlash — Type

DETPFlash(;numphases = 2;max_steps = 1e4*(numphases-1),population_size =20,time_limit = Inf,verbose = false, logspace = false)

Method to solve non-reactive multicomponent flash problem by finding global minimum of Gibbs Free Energy via Differential Evolution.

User must assume a number of phases, numphases. If true number of phases is smaller than numphases, model should predict either (a) identical composition in two or more phases, or (b) one phase with negligible total number of moles. If true number of phases is larger than numphases, a thermodynamically unstable solution will be predicted.

The optimizer will stop at max_steps evaluations or at time_limit seconds

Clapeyron.RRTPFlash — Type

RRTPFlash{T}(;K0 = nothing,rtol= 1e-10,atol = 1e-10,max_iters = 100)

Method to solve non-reactive multicomponent flash problem by using successive substitution in the Rachford-Rice equation.

Only two phases are supported. if K0 is nothing, it will be calculated via the Wilson correlation.

The optimizer will stop at max_iters evaluations, when the absolute tolerance is less than atol or when the relative tolerance is less than rtol

Clapeyron.MichelsenTPFlash — Type

MichelsenTPFlash{T}(;kwargs...)

Method to solve non-reactive multicomponent flash problem by Michelsen's method.

Only two phases are supported. if K0 is nothing, it will be calculated via the Wilson correlation.

Keyword Arguments:

equilibrium = equilibrium type ":vle" for liquid vapor equilibria, ":lle" for liquid liquid equilibria

K0 (optional), initial guess for the constants K

x0 (optional), initial guess for the composition of phase x

y0 = optional, initial guess for the composition of phase y

vol0 = optional, initial guesses for phase x and phase y volumes

K_tol = tolerance to stop the calculation

ss_iters = number of Successive Substitution iterations to perform

nacc = accelerate successive substitution method every nacc steps. Should be a integer bigger than 3. Set to 0 for no acceleration.

second_order = wheter to solve the gibbs energy minimization using the analytical hessian or not

noncondensables = arrays with names (strings) of components non allowed on the liquid phase. Not allowed with lle equilibria

nonvolatiles = arrays with names (strings) of components non allowed on the vapour phase. Not allowed with lle equilibria

Clapeyron.numphases — Function

numphases(method::TPFlashMethod)

return the number of phases supported by the TP flash method. by default its set to 2. it the method allows it, you can set the number of phases by doing method(;numphases = n).